==> scripts/01_global_ingredient_demand.Rmd <==
---
title: "Calculate demand for ingredients from global salmon production"
author: "Rich Cottrell"
date: "14/03/2022"
output: html_document
---


```{r}
library(tidyverse)
library(here)
library(janitor)
library(countrycode)
library(data.table)
library(dtplyr)

source(here("src/directories.R"))

select <- dplyr::select



```


#AQUACULTURE TRENDS DRIVING FEED DEMAND

Import aquaculture data and isolate Atlantic salmon. The top 10 producers produce 99.9% of global production so we can use this weight feed conversion ratio to get the approximate global average
```{r}

aquaculture <- readRDS(here("data/tidy_data/production-data/aquaculture_production_tidy.rds"))

aquaculture_species_groups <- unique(aquaculture$species)

salmon_species <- aquaculture_species_groups[grep("salmon", aquaculture_species_groups)]

salmon_aquaculture <- aquaculture %>% filter(species == "Atlantic salmon" & year %in% c(2019)) 


aquaculture %>% filter(species == "Atlantic salmon" & year %in% c(1990, 2019)) |> group_by(year) |> summarise(sum = sum(value))

(top_salmon_producers <- salmon_aquaculture %>% 
    group_by(country) %>% 
    summarise(value = sum(value, na.rm = TRUE)) %>% 
    arrange(-value) %>% 
    mutate(prop = value/sum(value),
           cum_prop = cumsum(prop)) %>% 
    slice(1:10)
    
    
    )




```

Import the feed conversion ratios and join to production. 

Feed conversion ratios were sourced from Tacon and Metian 2008 which has country specific information for salmon feed conversion ratios. Many of these are old and so we use the minimum values to represent country level economic feed conversion ratios and weight these values for each country by the proportion of global production they represent to create a weighted mean. This is approximately 1.2 and it is this value that we use for calculated total feed demand (production * eFCR).
```{r}

fcrs <- read_csv(here("data/raw_data/FCR.csv"))


(fcr_trends <- 
  top_salmon_producers %>% 
  left_join(fcrs, by = "country") %>% 
  mutate(eFCR_min = if_else(is.na(eFCR_min), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_min),
         eFCR_max = if_else(is.na(eFCR_max), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_max),
         eFCR_mean = if_else(is.na(eFCR_mean), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_mean)) |> 
  mutate(weighted_mean = sum(prop*eFCR_min)/sum(prop))
)

efcr <- fcr_trends |> pull(weighted_mean) |> unique() |> round(digits = 1)

```


Bring in diet data from downloaded from Teams (and stored in raw_data in this repo) and create separate scenario datasets for each diet. Also create a summary for reference for group totals across marine proteins, marine oils, plant proteins, plant oils, and carbohyrdate sources.

```{r}

#all_diets <- read_csv(here("data/raw_data/diet_formats.csv")) %>% dplyr::select(-Sources)
both_diets <- read_csv(here("data/raw_data/feed_composition_pressures.csv")) |> select(-X5) |> 
  mutate(ingredients = if_else(condition = ingredients== "canola/camelina oil", true = "canola oil", false = ingredients)) |> 
  mutate(ingredients = gsub(pattern = "-wild", replacement = "", ingredients))

diet_summaries <- 
  both_diets %>% 
  filter(grepl("totals", groups)) %>% 
  dplyr::select(c(groups, !matches('[2-9]+')), -ingredients) %>% 
  set_names(~str_replace_all(., pattern = "_1", replacement = ""))

saveRDS(object=diet_summaries, file = here("data/tidy_data/diets/diet_summaries.rds"))
  

both_diets_list <- 
  both_diets %>% 
  filter(!grepl("totals", groups)) |> 
  filter(groups != "Totals") |>  
  drop_na(ingredients) |> 
  set_names(~str_replace_all(., pattern = "_1", replacement = "")) |> 
  pivot_longer(names_to = "diet", values_to = "prop", -c(groups, ingredients)) |> 
  arrange(diet) |>  
  mutate(prop = prop/100) |> 
  group_split(diet)
   

map(.x = both_diets_list,  .f = \(this_diet){ 
  this_file_name <- unique(this_diet$diet); 
  saveRDS(object = this_diet, here(paste0(sprintf("data/tidy_data/diet-scenarios/%s", this_file_name), ".rds")))
  })



```

Now calculate ingredient demand per diet to supply 1000 tonnes of salmon based on the efcr of 1.2.

```{r}
diet_files <- list.files(here("data/tidy_data/diet-scenarios"), full.names = TRUE)


#calculate ingredient demand for 100,000 tonnes using the weighted efcr

map_df(diet_files, readRDS) |> 
  mutate(production_tonnes = 100000, 
         efcr = efcr) |> 
  mutate(ingredient_demand_tonnes = prop*production_tonnes*efcr) |> 
  filter(ingredient_demand_tonnes>0) |> 
  saveRDS(here("data/tidy_data/demand/total_ingredient_demand_by_diet.rds"))


```



```{r}

#create the ingredient list to match crop raw materials with ingredients

ingredient_list <- tibble(ingredient = readRDS(here("data/tidy_data/demand/total_ingredient_demand_by_diet.rds")) |> pull(ingredients) |> unique()) |> 
  write_csv(here("data/tidy_data/diets/ingredient_list.csv"))


#clear the environment
rm(list = ls(all.names = TRUE))
```
