==> scripts/02_crop_raw_materials_demand.Rmd <==
---
title: "Calculate embodied crop materials from feed demand"
output: html_document
---

Next step is convert crop ingredient demand into raw material demand which will vary by processing technique (e.g. SPC vs SBM) and source. So first, we need to know where the ingredients are largely sourced from. THis may change with information that we gain from Biomar but the idea is eventually to make this flexible enough to select any from the top producing countries. This requires knowing the raw material that supports the ingredient and the major places this raw material is grown.

```{r}
library(tidyverse)
library(here)
library(janitor)
library(countrycode)
library(data.table)
library(dtplyr)
library(terra)
library(sf)
library(jsonlite)
#library(ggnewscale)

source(here("src/directories.R"))

select <- dplyr::select
values <- terra::values

#allocation method (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"
```

#CROP INGREDIENTS

Have added codes and FAO_names for plant ingredients in teams so join these to production data.

```{r}

ingredient_by_raw_material <- 
  read_csv(here("data/tidy_data/diets/plant_ingredient_codes.csv")) |> 
  select(-reference) |> 
  drop_na(ingredient) |> 
  add_row(ingredient = "pea starch", raw_material = "peas", FAOSTAT_name = "Peas, green", FAO_code = 187, map_spam_code = "opul") |> 
  mutate(FAOSTAT_name = if_else(FAOSTAT_name == "Peas, dry; Peas, green", true = "Peas, dry", false = FAOSTAT_name)) 
 

crop_production_raw <- readRDS(file = here("data/tidy_data/production-data/crops_production_tidy.rds"))


(crop_production <- 
  crop_production_raw |>
  filter(element == "Production"  & area!= "China" & item %in% ingredient_by_raw_material$FAOSTAT_name & year %in% seq(from = 2015, to = 2019)) |> 
  group_by(area, item) |> 
  summarise(mean_value = mean(value, na.rm = TRUE)) |> 
    ungroup() |> 
  mutate(area = case_when(grepl("Ivoire", area) ~ "Cote d'Ivoire",
                          TRUE ~ area),
         iso3c = countrycode(sourcevar = area, origin = "country.name", destination = "iso3c", warn = TRUE)) |> 
  filter(!is.na(iso3c)) |> 
    arrange(item,-mean_value)
)


#looking across the top 10 producers for each crop
top_10_crop_producers <- 
  crop_production |> 
  group_split(item) |> 
  map_df(\(.){ 
    #this takes the top ten producers - unless it is for Guar (Pulses nes) where I used the suggestion that most guar production comes from India, Pakistan and Africa (so subsequently took the top 8 African producers of pulses nes to make up the top 10)
    
    if(unique(.$item) == "Pulses nes"){
      . |> filter(area %in% c("India", "Pakistan", "Kenya", 
                              "Ethiopia", "United Republic of Tanzania", 
                              "Sudan", "Sierra Leone", "Nigeria", 
                              "Guinea",  "Central African Republic"))
      }else{
        . |>  slice(1:10)}
  }
)


write.csv(x = top_10_crop_producers, file = here("data/tidy_data/production-data/top_crop_producers.csv"))
saveRDS(object = top_10_crop_producers, file = here("data/tidy_data/production-data/top_crop_producers.rds"))


```

Now we have the top 10 producers for each crop that is used as a raw material in feed, we need to get conversion factors from ingredients to each of these raw materials in different places and also the gross energy values of each ingredient and co-product so that energetic allocation and mass allocation conversion to embodied raw material can be achieved. This was done separately in the Sustainable Aquafeeds Project Teams interface and imported here.

```{r}

#pull in crop allocation codes

crop_allocation_factors <- read_csv(here("data/tidy_data/allocation/crop_ingredient_allocation_factors.csv"))

#import codes that will be compatible with MAPSPAM layers

plant_ingredient_codes <- 
  read_csv(here("data/tidy_data/diets/plant_ingredient_codes.csv")) |> 
  select(-reference) |> 
  drop_na(ingredient) |> 
  #add_row(ingredient = "pea starch", raw_material = "peas", FAOSTAT_name = "Peas, green", FAO_code = 187, map_spam_code = "opul") |> this row is for if you want pea starch to come from green peas and dried peas - just using dried peas for simplicity
  mutate(FAOSTAT_name = if_else(FAOSTAT_name == "Peas, dry; Peas, green", true = "Peas, dry", false = FAOSTAT_name),
         ingredient = if_else(ingredient == "canola/camelina oil", true = "canola oil", false = ingredient))


crop_conversion_lookup <- 
  plant_ingredient_codes |> 
  left_join(crop_allocation_factors, by = c("ingredient", "FAOSTAT_name" = "item")) |> 
  drop_na()
  
write_csv(x = crop_conversion_lookup, here("data/tidy_data/allocation/crop_conversion_lookup.csv"))


# 
# 
# 
# #join codes to top 20 producers and conversion factors data
# (top_10_crop_w_cf <- read_csv(here("data/tidy_data/production-data/top_crop_producers_conversions.csv")) |> 
#     select(-X7, -mean_value) |> 
#   left_join(plant_ingredient_codes |> select(-raw_material), by = c("ingredient", "item" = "FAOSTAT_name"))
# )
# 
# 
# 
# 
# 
# 
# 
# 
# #get yield data from main ingredients
# cf_data <- read_csv(here("data/tidy_data/production-data/top_crop_producers_conversions.csv")) |> 
#   mutate(ingredient = if_else(ingredient == "canola/camelina oil", true = "canola oil", false = ingredient)) |> filter(ingredient!="soy oil")
# 
# # get coproduct yield data
# cf_coproduct_data <- read_csv(here("data/raw_data/allocation/coproduct_conversions.csv"))
# 
# #join product and coproduct yield (conversion) data
# all_cf_data <- 
#   cf_data |> 
#   select(-X7) |> 
#   filter(!ingredient %in% c("tapioca starch", "lupins")) |> #not included in the diet information
#   left_join(cf_coproduct_data, by = c("ingredient", "iso3c"="country"))
# 
# 
# #bring in the gross energy values
# ge_values <- read_csv(here("data/raw_data/allocation/feed_coproduct_ge_allocation.csv")) |> 
#   select(-c(unit,reference, notes)) |> 
#   filter(!ingredient %in% c("fishmeal", "fish oil") & !coproduct %in% c("soybean meal", "soy protein isolate")) |> 
#   mutate(coproduct = if_else(coproduct == "soybean meal" & ingredient == "soybean meal", true = "soy protein concentrate", false = coproduct ))# |> 
#   #mutate(gross_energy_coproduct = if_else(coproduct == "guar gum", true = 1.7, false = gross_energy_coproduct))
# 
# #calculate the allocation factors for gross energy and mass allocation
# 
# (allocation_raw <- 
#   ge_values |> 
#   left_join(all_cf_data, by = c("ingredient", "coproduct")) |> 
#   mutate(energy_cf_product = gross_energy_product*cf,
#          energy_cf_coproduct = gross_energy_coproduct*coproduct_cf) |> 
#   ungroup() |> 
#   group_by(ingredient, area) |> 
#   nest() |> 
#   mutate(ge_allocation_factor = map(data, ~(.$gross_energy_product/ (sum(unique(.$energy_cf_product), sum(.$energy_cf_coproduct))))),
#          mass_allocation_factor = map(data, ~((.$cf/ (sum(unique(.$cf), sum(.$coproduct_cf))))/.$cf)) #need to divide by yield again here otherwise it is just the partitioning factor (for mass allocation only)
#          ) |> 
#   unnest(c(data, ge_allocation_factor, mass_allocation_factor)))
# 
# write_csv(x = allocation_raw, file = here("data/tidy_data/allocation/crop_ingredient_allocation_factors.csv"))

```

Join the ingredient demand and allocation data to generate crop raw material demand

```{r}

#select allocation method column
this_allocation_column <- paste0(this_allocation_method, "_factor")
this_allocation_abbreviation <- str_extract(this_allocation_method, "[^_]+")



#get saved allocation data
(allocation_factors <- read_csv(here("data/tidy_data/allocation/crop_ingredient_allocation_factors.csv")) |> 
  select(item, ingredient, this_allocation_column) |> distinct() |> 
  left_join(plant_ingredient_codes |> select(-raw_material), by = c("ingredient", "item" = "FAOSTAT_name")) |> 
  distinct()
)


selected_allocation_column <- allocation_factors[which(names(allocation_factors) == this_allocation_column)]

#join the ingredient demand by diet to the top 10 producer data so for each ingredient there will be an raw material demand per source, and diet scenario. Create new raw_material demand column by dividing total ingredient demand by the conversion factor (i.e. the ingredient extraction rate from the raw material in the first place)

(crop_raw_material_demand <- 
  readRDS(here("data/tidy_data/demand/total_ingredient_demand_by_diet.rds")) |> 
  ungroup() |> 
  left_join(allocation_factors, by = c("ingredients" = "ingredient")) |> 
  drop_na(item) |> 
  rename(FAOSTAT_name = item) |> 
  mutate(total_crop_demand =  unlist(across(contains(this_allocation_column)) * ingredient_demand_tonnes)))

#export total crop demand 
saveRDS(object = crop_raw_material_demand, file = sprintf(here("data/tidy_data/demand/total_crop_demand_%s.rds"), this_allocation_method))

```

```{r}
#clear the environment
rm(list = ls(all.names = TRUE))
```
