==> scripts/03_spatial_fish_demand_allocation.Rmd <==
  ---
title: "Calculate embodied fish demand in feeds"
output: html_document
---

```{r}
library(tidyverse)
library(here)
library(janitor)
library(countrycode)
library(data.table)
library(dtplyr)
library(raster)
library(terra)
library(sf)
library(jsonlite)
#library(ggnewscale)

source(here("src/directories.R"))


select <- dplyr::select
values <- terra::values


#allocation method (un-comment preference)

source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"


```


Now we need to work out how the ingredient demand converts to embodied fish biomass. This will depend on species which depends on geographic region and also the degree of trimmings used for FMFO in this region. So we need different conversions from different species and forage and trimmings sources.
  

```{r}
#select column for the correct allocation method
this_allocation_column <- paste0(this_allocation_method, "_factor")


#bring in total ingredient demand data and isolate fishmeal and oil demand

total_ingredient_demand <- readRDS(here("data/tidy_data/demand/total_ingredient_demand_by_diet.rds"))

fmfo_demand <- total_ingredient_demand |> filter(grepl("Marine", groups))

rm(total_ingredient_demand)

(foragefish_allocation_factors <- read_csv(here("data/tidy_data/allocation/forage_fish_allocation_factors.csv")) |> 
  select(common_name, sci_name, ingredient, this_allocation_column) |> 
  mutate(sci_name = case_when(sci_name == "Brevoorti patronus" ~ "Brevoortia patronus",
                              TRUE ~ sci_name)) |> distinct() |> drop_na(this_allocation_column))


#bring in trimmings weightings and add to conversion factors

trimmings_wghts <- read_csv(here("data/tidy_data/demand/trimmings_weightings.csv")) |> 
  group_by(group) |> 
  summarise(trimmings_weight = sum(share_of_trimmings))


(trimmings_allocation_factors <- 
  read_csv(here("data/tidy_data/allocation/trimmings_allocation_factors.csv")) |> 
  select(common_name, sci_name, ingredient, this_allocation_column) |> 
  mutate(group = case_when(grepl("sprat|Sprat", common_name) ~ "Sprats",
                           grepl("anchovy|anchoita|Anchovies|anchoveta|Anchovy", common_name) ~ "Anchovies",
                           grepl("Tuna|tuna|Albacore", common_name) ~ "Tunas",
                           grepl("cod|pollack|Pollock", common_name) ~ "Cods",
                            grepl("sardinella|Sardinella|sardine", common_name) ~ "Sardines",
                           grepl("mackerel", common_name) ~ "Mackerels",
                           grepl("herring|Herring", common_name) ~ "Herrings",
                           grepl("menhaden|Mehaden", common_name) ~ "Menhaden",
                           grepl("Capelin", common_name) ~ "Capelin",
                           grepl("Blue whiting|Blue Whiting", common_name) ~ "Blue whiting",
                           grepl("Sandeel", common_name) ~ "Sandeel"))
  )


# # #global values for later
# globalav_foragefish_fm_mass <- foragefish_cfs |> filter(common_name == "Global average" & ingredient == "Fishmeal") |> pull(mass_value)
# globalav_foragefish_fm_ge <- foragefish_cfs |> filter(common_name == "Global average" & ingredient == "Fishmeal") |> pull(ge_value)
# 
# globalav_foragefish_fo_mass <- foragefish_cfs |> filter(common_name == "Global average" & ingredient == "Fish oil") |> pull(mass_value)
# globalav_foragefish_fo_ge <- foragefish_cfs |> filter(common_name == "Global average" & ingredient == "Fish oil") |> pull(ge_value)
```

Isolate catch for trimmings first
```{r}

trimmings_spp_catch <- readRDS(here("data/large_data/trimmings_spp_catch.rds")) |> filter(IYear == 2017)

trimmings_xyz <- 
  trimmings_spp_catch |> 
  group_by(LonCentre, LatCentre) |> 
  summarise(total_catch = sum(ReportedIND)) |> 
  #select(LonCentre, LatCentre, total_catch) |> 
  rename(x=LonCentre, y= LatCentre, z = total_catch) 


total_trimmings_raster <- rast(trimmings_xyz, type="xyz")

plot(total_trimmings_raster)


#FAO areas of most prominence from Biomar sustainability report 2022 for both wild caught and trimmings https://www.biomar.com/globalassets/.global/sustainability-report/biomar-global-sustainability-report-2022.pdf

#bring in the FAO regions shape file

fao_areas_shp <- 
  read_sf("https://www.fao.org/fishery/geoserver/fifao/ows?service=WFS&request=GetFeature&version=1.1.0&typeName=FAO_AREAS_NOCOASTLINE&outputFormat=json") |> filter(F_LEVEL == "MAJOR") |> 
  mutate(F_AREA = as.numeric(F_AREA)) |> 
  group_by(F_AREA) |> 
  group_split()

this_area <- fao_areas_shp[[13]]



#get xyz file for the FAO areas in 0.5 resolution (catch data resolution)

fao_xyz <- 
  
  map_df(.x = fao_areas_shp, .f = \(this_area){
  
  this_area_shp <- this_area
  
  this_area_raster <- terra::rasterize(x = vect(this_area_shp), y = rast(res = 0.5), field = this_area_shp$F_AREA)
  
  this_area_xyz <- terra::as.data.frame(x = this_area_raster, xy = TRUE)  |> mutate(fao_area_name = this_area_shp$NAME_EN)
  
  return(this_area_xyz)
  
}) |> rename(fao_area_code = layer,
             LonCentre = x,
             LatCentre = y)



#check FAO shapefile appears correctly
ggplot()+
  geom_sf(data = fao_areas_shp |> bind_rows(), 
          aes(fill = F_CODE))


#join catch and fao xyz to identify the catch for species used in trimmings across FAO regions

trimmings_spp_fao_area <- 
  trimmings_spp_catch |> 
  left_join(fao_xyz) |> 
  group_by(fao_area_name, fao_area_code) |> 
  summarise(trim_spp_catch = sum(ReportedIND, na.rm=TRUE)) |> 
  ungroup() |> 
  arrange(-trim_spp_catch) |> 
  mutate(prop_trimmings = trim_spp_catch/sum(trim_spp_catch, na.rm=TRUE))


```



```{r}
forage_catch <- readRDS(here("data/large_data/spatial_forage_catch_2017.rds"))


#plot forage catch to identify main regions for production. The areas I'm interested in is the Humboldt system, the caribeban coast off latin america, the gulf of mexico, the north atlantic, west africa, the western INdian Ocean (India, Red Sea) the east asia regions
forage_xyz <- forage_catch |> 
  group_by(LonCentre, LatCentre) |> 
  summarise(total_catch = sum(ReportedIND)) |> 
  #select(LonCentre, LatCentre, total_catch) |> 
  rename(x=LonCentre, y= LatCentre, z = total_catch) 

total_forage_raster <- rast(forage_xyz, type="xyz")

plot((total_forage_raster))

sum(values(total_forage_raster), na.rm = TRUE)


fao_areas_shp <- 
  read_sf("https://www.fao.org/fishery/geoserver/fifao/ows?service=WFS&request=GetFeature&version=1.1.0&typeName=FAO_AREAS_NOCOASTLINE&outputFormat=json") |> filter(F_LEVEL == "MAJOR") |> 
  mutate(F_AREA = as.numeric(F_AREA)) |> 
  group_by(F_AREA) |> 
  group_split()

#test the function
this_area <- fao_areas_shp[[1]]



#get xyz file for the FAO areas in 0.5 resolution (catch data resolution)

fao_xyz <- 
  
  map_df(.x = fao_areas_shp, .f = \(this_area){
  
  this_area_shp <- this_area
  
  this_area_raster <- terra::rasterize(x = vect(this_area_shp), y = rast(res = 0.5), field = this_area_shp$F_AREA)
  
  this_area_xyz <- terra::as.data.frame(x = this_area_raster, xy = TRUE)  |> mutate(fao_area_name = this_area_shp$NAME_EN)
  
  return(this_area_xyz)
  
}) |> rename(fao_area_code = layer,
             LonCentre = x,
             LatCentre = y)

#join FAO area xyz  to catch data
forage_spp_fao_area <- 
  forage_catch |> 
  left_join(fao_xyz, 
            by = c("LonCentre","LatCentre")) |> 
  drop_na(fao_area_name) |>
  group_by(fao_area_name, fao_area_code) |> 
  summarise(forage_spp_catch = sum(ReportedIND, na.rm=TRUE)) |> 
  ungroup() |> 
  arrange(-forage_spp_catch) |> 
  mutate(prop_forage_catch = forage_spp_catch/sum(forage_spp_catch, na.rm=TRUE))



forage_spp_fao_area |> pull(fao_area_code) |> unique()
```



Now join the trimmings and the forage fish catch potential for each fao fishing area, assign a proportion that is coming from each area, and then choose which areas will be used for sourcing for the fishmeal and oil demand.

```{r}
fish_sourcing_areas <- 
  forage_spp_fao_area |> 
  left_join(trimmings_spp_fao_area) |> 
  mutate(max_fm_trim_prop = case_when(fao_area_code == 27 ~ 0.3,
                                      TRUE~0),# proportion of fishmeal from trimmings based loosely on on biomar sustainability report 2022 - https://www.biomar.com/globalassets/.global/sustainability-report/biomar-global-sustainability-report-2022.pdf
         min_fm_forage_prop = 1-max_fm_trim_prop,
         max_fo_trim_prop = 0.4, # proportion of fish oil from trimmings based loosely on on biomar sustainability report 2022 - https://www.biomar.com/globalassets/.global/sustainability-report/biomar-global-sustainability-report-2022.pdf
         min_fo_forage_prop = 1-max_fo_trim_prop) |> 
  filter(fao_area_code %in% c(27, 87, 31, 61, 67)) # this mix of FAO fishing areas are based off values in Aas et al 2022 and Biomar sustainability report

saveRDS(object = fish_sourcing_areas, file = sprintf(here("data/tidy_data/production-data/fao_regions_fish_production_potential_%s.rds"), this_allocation_method))



```




Now we have to calculate the yield of fishmeal and oil per cell of the fisheries data for both trimmings and forage.

First start with sources for trimmings. 

- Join data for fao fishing area xyz and filter by the sourcing areas of interest above
- create and join duplicate data frames for fishmeal and fish oil and assign an umbrella group to each species (e.g. Mackerels, anchovies)
- Join the embodied fish conversion factors for species used in trimmings for both fm and fo
- Divide fish biomass in each cell by the FMFO conversion factors to get FMFO yields for each record (row) in data
- By each fishing area, calculate the proportion of total available FMFO each record accounts for

```{r}

#create duplicate data for fishmeal and oil, filtered by the correct FAO fishing areas to join the conversion factors more efficiently

(trim_fmfo_yield <- 
   #bind FMFO dataframes together
  bind_rows(
    #fishmeal data frame filtered for FAO fishing areas of interest
    trimmings_spp_catch |> 
            left_join(fao_xyz) |> 
            filter(fao_area_code %in% fish_sourcing_areas$fao_area_code) |> 
            mutate(ingredient = "Fishmeal"),
    #fish oil dataframe filtered for FAO fishing areas of interest
trimmings_spp_catch |> 
  left_join(fao_xyz) |> 
            filter(fao_area_code %in% fish_sourcing_areas$fao_area_code) |> 
  mutate(ingredient = "Fish oil")) |>
  #now assign groups for species 
   mutate(group = case_when(grepl("sprat", CommonName) ~ "Sprats",
                           grepl("anchovy|anchoita|Anchovies|nchoveta", CommonName) ~ "Anchovies",
                           grepl("Tuna|tuna|Albacore", CommonName) ~ "Tunas",
                           grepl("cod|pollack", CommonName) ~ "Cods",
                            grepl("sardinella|Sardinella|sardine|pilchard", CommonName) ~ "Sardines",
                           grepl("mackerel", CommonName) ~ "Mackerels",
                           grepl("herring", CommonName) ~ "Herrings",
                           grepl("menhaden", CommonName) ~ "Menhaden",
                           grepl("Capelin", CommonName) ~ "Capelin",
                           grepl("Blue whiting", CommonName) ~ "Blue whiting")) |> 
  #add the conversion factors for trimmings
  left_join(trimmings_allocation_factors, by = c("CommonName" = "common_name", 
                                                 "TaxonName" = "sci_name", 
                                                 "ingredient", 
                                                 "group")) |> 
  left_join(trimmings_wghts, by = "group") |> 
  rename(byproduct_allocation_factor = this_allocation_column) |> 
  #fill in NA values through species group values for mass and energetic allocation
  mutate(byproduct_allocation_factor = case_when(#when the na group is anchovies make it Peruvian anchoveta conversion factor
                                      is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Anchovies" ~ 
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Peruvian anchovy" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Anchovies"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Peruvian anchovy" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                      
                                      #when the na species is chub mackerels make it Pacific mackerel conversion factor
                                      is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Mackerels"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Pacific mackerel" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Mackerels"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Pacific mackerel" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                      #when the na group is a menhaden sp make it Gulf menhaden conversion factor
                                      
                                       is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Menhaden"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Gulf menhaden" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Menhaden"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Gulf menhaden" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                       
                                       #Where sardines make it the Atlantic sardine (pilchard) conversion factor
                                       is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Sardines"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Atlantic sardine" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Sardines"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Atlantic sardine" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                      #Where cods make it the Alaska pollack conversion factor (the binomial in Reg's data is different)
                                       is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Cods"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Alaska pollack" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Cods"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Alaska pollack" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                      #where sprats is na - make conversion factors same as European sprat
                                      is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Sprats"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "European sprat" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Sprats"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "European sprat" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                      #where tunas is na - make conversion factors same as Skipjack tuna
                                      is.na(byproduct_allocation_factor) & ingredient == "Fishmeal" & group == "Tunas"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Skipjack tuna" & ingredient == "Fishmeal") |> 
                                         pull(this_allocation_column),
                                       
                                       is.na(byproduct_allocation_factor) & ingredient == "Fish oil" & group == "Tunas"~
                                        trimmings_allocation_factors |> 
                                         filter(common_name == "Skipjack tuna" & ingredient == "Fish oil") |> 
                                         pull(this_allocation_column),
                                      
                                       TRUE ~ byproduct_allocation_factor)) |> 
  
  # add the proportion of demand coming from trimmings from different FAO fishing areas
   left_join(fish_sourcing_areas |> select(fao_area_name, fao_area_code, max_fm_trim_prop, max_fo_trim_prop)) |>
  #calculate the yield of FMFO for each cell
     mutate(yield = ReportedIND/byproduct_allocation_factor) |> 
    mutate(ingredient = gsub(" ", "_", tolower(ingredient))) |> 
    select(-c(IUUIND, DiscardsIND, ReportedNIND, IUUNIND, DiscardsNIND, ISSCAAPName, foragefish, OceanAreasqkm)) |> 
  #turn to wide format to reverse duplicate rows created from duplicate dataframe for FM and FO
    pivot_wider(names_from = "ingredient", values_from = c(byproduct_allocation_factor, yield)) |> 
  mutate(type="trimmings"))
 

fwrite(trim_fmfo_yield, file = here("data/large_data/embodied_fmfo_potential_from_trimmings.csv"))

```

Now repeat this for forage fish

- Join data for fao fishing area xyz and filter by the sourcing areas of interest above
- create and join duplicate data frames for fishmeal and fish oil and assign an umbrella group to each species (e.g. Mackerels, anchovies)
- Join the embodied fish conversion factors for forage fish species for both fm and fo
- Divide fish biomass  in each cell by the FMFO conversion factors to get FMFO yields for each record (row) in data
- By each fishing area, calculate the proportion of total available FMFO each record accounts for


```{r}
(forage_fmfo_yield <- 
   #bind FMFO dataframes together
  bind_rows(
    #fishmeal data frame filtered for FAO fishing areas of interest
    forage_catch |> 
            left_join(fao_xyz) |> 
            filter(fao_area_code %in% fish_sourcing_areas$fao_area_code) |> 
            mutate(ingredient = "Fishmeal"),
    #fish oil dataframe filtered for FAO fishing areas of interest
forage_catch |> 
  left_join(fao_xyz) |> 
            filter(fao_area_code %in% fish_sourcing_areas$fao_area_code) |> 
  mutate(ingredient = "Fish oil")) |> 
  #add the conversion factors for forage
  left_join(foragefish_allocation_factors, by = c("ingredient" , "TaxonName" = "sci_name", "CommonName" = "common_name")) |> 
  #add group information
  rename(forage_allocation_factor = this_allocation_column) |> 
  mutate(group = case_when(grepl("sprat", CommonName) ~ "Sprats",
                           grepl("anchovy|anchoita|Anchovies|nchoveta|nchovies", CommonName) ~ "Anchovies",
                           grepl("Tuna|tuna|Albacore", CommonName) ~ "Tunas",
                           grepl("cod|pollack", CommonName) ~ "Cods",
                            grepl("sardinella|Sardinella|sardine|pilchard", CommonName) ~ "Sardines",
                           grepl("ackerel", CommonName) ~ "Mackerels",
                           grepl("erring", CommonName) ~ "Herrings",
                           grepl("menhaden", CommonName) ~ "Menhaden",
                           grepl("Capelin", CommonName) ~ "Capelin",
                           grepl("Blue whiting", CommonName) ~ "Blue whiting")) |> 
  #fill in NA values through species group values for mass and energetic allocation
  mutate(forage_allocation_factor = case_when(#when the na group is ones we know from trimmings - assign to groups or prominent species
                                      
                                      #For the Peruvian Anchoveta specify its own one 
                                      is.na(forage_allocation_factor) & TaxonName == "Engraulis ringens" & ingredient == "Fishmeal" ~ 
                                        foragefish_allocation_factors |>
                                        filter(sci_name == "Engraulis ringens" & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column),
                                      is.na(forage_allocation_factor) & TaxonName == "Engraulis ringens" & ingredient == "Fish oil" ~  
                                        foragefish_allocation_factors |>
                                        filter(sci_name == "Engraulis ringens" & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column),
                                
                                      
                                      #for other species in the herring group assign the mean from herring species
                                       is.na(forage_allocation_factor) & group == "Herrings" & ingredient == "Fishmeal"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("erring", common_name) & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column) |> mean(),
                                      is.na(forage_allocation_factor) & group == "Herrings" & ingredient == "Fish oil"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("erring", common_name) & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                      
                                      #for other species in the mackerel group assign the mean from herring species
                                       is.na(forage_allocation_factor) & group == "Mackerels" & ingredient == "Fishmeal"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("ackerel", common_name) & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column) |> mean(),
                                      is.na(forage_allocation_factor) & group == "Mackerels" & ingredient == "Fish oil"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("ackerel", common_name) & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                      #for other species in anchovy group assign mean from anchovy species
                                      is.na(forage_allocation_factor) & group == "Anchovies" & ingredient == "Fishmeal"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("anchovy|anchoita|Anchovies|nchoveta|nchovies", common_name) & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column) |> mean(),
                                      is.na(forage_allocation_factor) & group == "Anchovies" & ingredient == "Fish oil"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("anchovy|anchoita|Anchovies|nchoveta|nchovies", common_name) & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                        #for other species in sardines group assign mean from sardine/sardinella species
                                      is.na(forage_allocation_factor) & group == "Sardines" & ingredient == "Fishmeal"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("sardinella|Sardinella|sardine|pilchard", common_name) & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                      is.na(forage_allocation_factor) & group == "Sardines" & ingredient == "Fish oil"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("sardinella|Sardinella|sardine|pilchard", common_name) & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                      #for menhaden do the same
                                       is.na(forage_allocation_factor) & group == "Menhaden" & ingredient == "Fishmeal"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("enhaden", common_name) & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                      is.na(forage_allocation_factor) & group == "Menhaden" & ingredient == "Fish oil"~
                                        foragefish_allocation_factors |> 
                                        filter(grepl("enhaden", common_name) & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column) |> mean(),
                                      
                                      #for all other species assign a global mean
                            
                                       is.na(forage_allocation_factor) & ingredient == "Fishmeal"~
                                        foragefish_allocation_factors |> 
                                        filter(is.na(common_name) & ingredient == "Fishmeal") |> 
                                        pull(this_allocation_column),
                                      
                                      is.na(forage_allocation_factor) & ingredient == "Fish oil"~
                                        foragefish_allocation_factors |> 
                                        filter(is.na(common_name) & ingredient == "Fish oil") |> 
                                        pull(this_allocation_column),
                        
                                      TRUE ~ forage_allocation_factor)) |> 
  
  # add the proportion of demand coming from forage fish from different FAO fishing areas
   left_join(fish_sourcing_areas |> select(fao_area_name, fao_area_code, min_fm_forage_prop, min_fo_forage_prop)) |>
  
  #calculate the yield of FMFO for each cell
     mutate(yield = ReportedIND/forage_allocation_factor) |> 
    mutate(ingredient = gsub(" ", "_", tolower(ingredient))) |> 
    select(-c(IUUIND, DiscardsIND, ReportedNIND, IUUNIND, DiscardsNIND, ISSCAAPName, foragefish, OceanAreasqkm)) |> 
  #turn to wide format to reverse duplicate rows created from duplicate dataframe for FM and FO
    pivot_wider(names_from = "ingredient", values_from = c("forage_allocation_factor", "yield")) |>  
  #calculate the proportion each record represents for each FAO fishing area - this can be multiplied by demand under each diet
filter(!across(everything(), is.infinite) %>% purrr::reduce(`|`)) |> #this drops one record of Euphausia superba where reported industrial catch was 0.2 tonnes
 group_by(fao_area_code) |> 
  nest() |>
  mutate(prop_fm = map(data, ~(.$yield_fishmeal/sum(.$yield_fishmeal, na.rm = TRUE))),
         prop_fo = map(data, ~(.$yield_fish_oil/sum(.$yield_fish_oil, na.rm = TRUE)))) |> 
  unnest(cols = c(data, prop_fm, prop_fo)) |> 
  ungroup() |> 
  mutate(type="foragefish")) 


fwrite(x = forage_fmfo_yield, file =  here("data/large_data/embodied_fmfo_potential_from_foragefish.csv"))


```




Find out how much fishmeal and oil is available globally from the species used in trimmings and using forage fish and distribute the demand for fmfo across cellsm, and convert back to embodied fish for calculations.


```{r}
 
trim_fmfo_yield <- fread(here("data/large_data/embodied_fmfo_potential_from_trimmings.csv"))
forage_fmfo_yield <- fread(here("data/large_data/embodied_fmfo_potential_from_foragefish.csv"))


#forage
(total_fm_available<- forage_fmfo_yield  |>  pull(yield_fishmeal) |> sum())
(total_fo_available <- forage_fmfo_yield$yield_fish_oil |> sum())



#trimmings - not meaningful as biomass divided by econ allocation become silly numbers. Because we attribute less to it doesn't mean there is physically more there
(total_fm_available_trim <- trim_fmfo_yield$yield_fishmeal |> sum())
(total_fo_available_trim <- trim_fmfo_yield$yield_fish_oil |> sum())


#Forage fish 

#use the proportions for each region to divide the demand up and work out how the spp_based demand converts to embodied fish


#test this function
this_diet <- "marine_diet"


forage_supply_demand_by_diet <- 
  map_df(.x =  c("marine_diet", "plant_diet"), .f = \(this_diet){
    
    this_df <- forage_fmfo_yield |> 
      mutate(diet = this_diet) |> 
      mutate(#extract total fmfo demand for each diet
             total_fm_needed_for_diet = fmfo_demand |> filter(diet == this_diet & ingredients == "fishmeal") |> pull(ingredient_demand_tonnes),
             total_fo_needed_for_diet = fmfo_demand |> filter(diet == this_diet & ingredients == "fish oil") |> pull(ingredient_demand_tonnes),
             #distribute fishmeal and oil demand according to its distribution
             distributed_fm_demand = total_fm_needed_for_diet * prop_fm,
             distributed_fo_demand = total_fo_needed_for_diet * prop_fo,
            
             #calculate embodied demand
             #calculated the embodied fish from fm and fo
             fm_embodied_fish = distributed_fm_demand * min_fm_forage_prop * forage_allocation_factor_fishmeal ,
             fo_embodied_fish = distributed_fo_demand * min_fo_forage_prop * forage_allocation_factor_fish_oil,
             
             #calculate the total fmfo demand
             total_fmfo_demand = distributed_fm_demand + distributed_fo_demand,
             total_fmfo_demand = distributed_fm_demand + distributed_fo_demand,
             
             #calculate the total embodied fish
             total_embodied_fish = fm_embodied_fish +fo_embodied_fish
             ) |> 
      relocate(diet, .before = fao_area_code) |> 
      as.data.frame()
    
  })

fwrite(forage_supply_demand_by_diet, file = here("data/large_data/embodied_foragefish_per_cell.csv"))


#Create a total embodied fish demand by diet and source
embodied_foragefish_demand_by_diet_source <- 
  forage_supply_demand_by_diet |> 
  group_split(diet, fao_area_code) |> 
  map_df(\(this_df){
    
    return(tibble(diet = unique(this_df$diet), 
                  fao_area_code = unique(this_df$fao_area_code), 
                  embodied_fish_demand_from_fm = sum(this_df$fm_embodied_fish),
                  embodied_fish_demand_from_fo = sum(this_df$fo_embodied_fish),
                  embodied_fish_demand = sum(this_df$total_embodied_fish)))
    
  }) |> 
  pivot_longer(cols = -c(diet, fao_area_code), names_to = "demand_from", values_to = "embodied_foragefish" ) |> 
  mutate(allocation = this_allocation_method) |> 
  mutate(demand_from = case_when(grepl("from_fm", demand_from) ~ "Fishmeal",
                                 grepl("from_fo", demand_from) ~ "Fish oil",
                                 TRUE ~ "Total")) |> arrange(diet, fao_area_code, allocation)



saveRDS(object = embodied_foragefish_demand_by_diet_source, file = here("data/tidy_data/demand/embodied_foragefish_demand_by_source_diet.rds"))


#Trimmings


#use the proportions for each region and weight by group to divide the demand up and work out how the spp_based demand converts to embodied fish

trim_supply_demand_by_diet <- 
  map_df(.x =  c("marine_diet", "plant_diet"), .f = \(this_diet){
    
    this_trimmings_df <- trim_fmfo_yield |> as_tibble()
    
    these_weight_sums <- 
      map_df(this_trimmings_df |> 
      group_by(fao_area_code) |> 
      group_split(), \(this_df){
        
       return(tibble(fao_area_code = this_df |> pull(fao_area_code) |> unique(), trimmings_weight_sum = this_df |> pull(trimmings_weight) |> unique() |> sum()))
      })
    
this_demand_df <- 
      this_trimmings_df |>
      left_join(these_weight_sums) |> 
      mutate(diet = this_diet) |> 
      mutate(#extract total fmfo demand for each diet
             total_fm_needed_for_diet = fmfo_demand |> filter(diet == this_diet & ingredients == "fishmeal") |> pull(ingredient_demand_tonnes),
             total_fo_needed_for_diet = fmfo_demand |> filter(diet == this_diet & ingredients == "fish oil") |> pull(ingredient_demand_tonnes)) |> 
      mutate(weighted_fm_needed = total_fm_needed_for_diet*trimmings_weight/trimmings_weight_sum,
             weighted_fo_needed = total_fo_needed_for_diet*trimmings_weight/trimmings_weight_sum) |> 
      #now see the yield for each area from each group that cell represents
      as_tibble() |> 
      group_by(fao_area_code, group) |> 
      nest() |> 
  mutate(prop_fm = map(data, ~(.$yield_fishmeal/sum(.$yield_fishmeal))),
         prop_fo = map(data, ~(.$yield_fish_oil/sum(.$yield_fish_oil)))) |> 
  unnest(cols = c(data, prop_fm, prop_fo)) |> 
      mutate(
        #distribute fishmeal and oil demand according to its distribution and the weight of that group in trimmings
             distributed_fm_demand =  weighted_fm_needed * prop_fm,
             distributed_fo_demand = weighted_fo_needed * prop_fo,
            
             #calculate embodied demand
             #calculated the embodied fish from fm and fo
             fm_embodied_fish = distributed_fm_demand * max_fm_trim_prop * byproduct_allocation_factor_fishmeal,
             fo_embodied_fish = distributed_fo_demand * max_fo_trim_prop * byproduct_allocation_factor_fish_oil,
        
             #calculate the total fmfo demand
             total_fmfo_demand = distributed_fm_demand + distributed_fo_demand,
   
             #calculate the total embodied fish
             total_embodied_fish = fm_embodied_fish + fo_embodied_fish,

             ) |> 
      relocate(diet, .before = fao_area_code)
      
    return(this_demand_df)
    
  }) |> ungroup()

fwrite(trim_supply_demand_by_diet, file = here("data/large_data/embodied_fishfromtrimmings_per_cell.csv"))



#Create a total embodied fish demand by diet and source
embodied_trimingsfish_demand_by_diet_source <- 
  trim_supply_demand_by_diet |> 
  group_split(diet, fao_area_code) |> 
  map_df(\(this_df){
    
    return(tibble(diet = unique(this_df$diet), 
                  fao_area_code = unique(this_df$fao_area_code), 
                  embodied_fish_demand_from_fm = sum(this_df$fm_embodied_fish),
                  embodied_fish_demand_from_fo = sum(this_df$fo_embodied_fish),
                  embodied_fish_demand = sum(this_df$total_embodied_fish)))
    
  }) |> 
  pivot_longer(cols = -c(diet, fao_area_code), names_to = "demand_from", values_to = "embodied_fish" ) |> 
  mutate(allocation = this_allocation_method) |> 
  mutate(demand_from = case_when(grepl("from_fm", demand_from) ~ "Fishmeal",
                                 grepl("from_fo", demand_from) ~ "Fish oil",
                                 TRUE ~ "Total")) |> arrange(diet, fao_area_code, allocation)



saveRDS(object = embodied_trimingsfish_demand_by_diet_source, file = here("data/tidy_data/demand/embodied_trimmingsfish_demand_by_source_diet.rds"))




```



```{r}
rm(list = ls(all.names = TRUE))
```
