==> scripts/3_query_IUCN_api.Rmd <==
---
title: "Query IUCN API"
author: "Rich Cottrell"
date: "23/02/2022"
output: html_document
---

```{r}
library(tidyverse)
library(furrr)
library(tictoc)
library(janitor)
library(here)
library(rredlist)
library(jsonlite)
library(parallel)

source(here("src/directories.R"))
source(here("src/iucn_fxs.R"))




```

Get a species list
```{r}


sppcount_url <- sprintf("https://apiv3.iucnredlist.org/api/v3/speciescount?token=%s", api_key)
nspp <- jsonlite::fromJSON(sppcount_url) %>% .$speciescount %>% as.integer()

pagelist <- c(0:(ceiling(nspp/10000)-1)) #10000 is number of species presented per page


spp_list_df <- bind_rows(mclapply(X = pagelist, FUN = get_spp_api, mc.cores = detectCores())) %>% 
  select(-infra_name, -infra_rank) %>% 
  setNames(names(.) %>% 
             str_replace("_name", ""))


saveRDS(object = spp_list_df, file = here("data/raw_data/iucn/spp_list.rds"), compress = TRUE)


```

Get species habitat data. This takes a couple of dats to run. Don't run if not needed. If needed, adjust the 'iucn_habitat_dir' to your storage.

```{r}

spp_list <- 
  readRDS(file = here("data/raw_data/iucn/spp_list.rds"))  %>% pull(taxonid) 
  
iucn_habitat_dir <- file.path(iucn_dir, "habitat_info")

#pre loop prep
chunk_size <- 500 #chunk size for processing
spp_list_split <- split(spp_list, ceiling(seq_along(spp_list)/chunk_size)) #list of species
n_species <- 1:length(spp_list) 
n_species_index <- split(n_species, ceiling(seq_along(n_species)/chunk_size)) #position of species in list


for (j in 1:length(spp_list_split)){

  #index by species identity
  spp_ids <- spp_list_split[[j]]
  first_sp <- first(spp_ids)
  last_sp <- last(spp_ids)
  
  #index by number of species
  spp_index <- n_species_index[[j]]
  first_spp_index <- first(spp_index)
  last_spp_index <- last(spp_index)
  
  chunk_file <- file.path(sprintf(paste0(iucn_threats_dir,"/spp_habitat_%s-%s.rds"), first_spp_index, last_spp_index))
  
  if(!file.exists(chunk_file)){
    
    message("Getting habitat info for species positions ", first_spp_index, "-", last_spp_index)
    
    future::plan(strategy = multisession, workers =8)
    habitat_list <- future_map(spp_ids,  get_habitat_api)
    
    habitat_list_df <- bind_rows(habitat_list)
    
    saveRDS(object = habitat_list_df, file = file.path(sprintf(paste0(iucn_habitat_dir, "/spp_habitats_%s-%s.rds"), first_spp_index, last_spp_index)))
    
  }
  
  
}

habitat_files <- list.files(path = iucn_habitat_dir, full.names = TRUE)


all_habitats<-  map_df(habitat_files, readRDS)

#save to rdsi
saveRDS(object = all_habitats, file = file.path(iucn_habitat_dir, "1_all_habitats.rds"))

#save to repo
saveRDS(object = all_habitats, file = here("data/raw_data/iucn/habitats_by_spp.rds"))



#accidentally downloaded habitats instead of threats so renaming habitat files
# rename_habitat_files <- function(this_filename){
#   
#   file.rename(from = this_filename, to = str_replace(string = this_filename, pattern = "threatss", replacement = "habitat"))
#   
# }
# 
# 
# map(habitat_files, rename_habitat_files)


```


Get species threats data. This takes a couple of days to run. Don't run if not needed. If needed, adjust the 'iucn_threat_dir' to your storage.

```{r}

spp_list <- 
  readRDS(file = here("data/raw_data/iucn/spp_list.rds"))  %>% pull(taxonid) 
  
iucn_threat_dir <- file.path(iucn_dir, "threat_info")

#pre loop prep
chunk_size <- 500 #chunk size for processing
spp_list_split <- split(spp_list, ceiling(seq_along(spp_list)/chunk_size)) #list of species
n_species <- 1:length(spp_list) 
n_species_index <- split(n_species, ceiling(seq_along(n_species)/chunk_size)) #position of species in list


for (j in 1:length(spp_list_split)){

  #index by species identity
  spp_ids <- spp_list_split[[j]]
  first_sp <- first(spp_ids)
  last_sp <- last(spp_ids)
  
  #index by number of species
  spp_index <- n_species_index[[j]]
  first_spp_index <- first(spp_index)
  last_spp_index <- last(spp_index)
  
  chunk_file <- file.path(sprintf(paste0(iucn_threat_dir,"/spp_threats_%s-%s.rds"), first_spp_index, last_spp_index))
  
  if(!file.exists(chunk_file)){
    
    message("Getting threat info for species positions ", first_spp_index, "-", last_spp_index)
    
    future::plan(strategy = multisession, workers =8)
    threat_list <- future_map(spp_ids,  get_threat_api_casey)
    
    threat_list_df <- bind_rows(threat_list)
    
    saveRDS(object = threat_list_df, file = file.path(sprintf(paste0(iucn_threat_dir, "/spp_threats_%s-%s.rds"), first_spp_index, last_spp_index)))
    
  }
  
  
}

threat_files <- list.files(path = file.path(iucn_threat_dir), full.names = TRUE)

all_threats <- map_df(threat_files, readRDS)

saveRDS(object = all_threats, file = file.path(iucn_threat_dir, "1_all_threats.rds"))

saveRDS(object = all_threats, file = here("data/raw_data/iucn/threats_by_spp.rds"))


```
