==> scripts/04_crop_source_country_rasters.Rmd <==
---
title: "Create country-based rasters"
author: "Rich Cottrell"
date: "20/04/2022"
output: html_document
---
Libraries
```{r}
library(tidyverse)
library(sf)
library(here)
library(terra)
library(rnaturalearth)
library(fasterize)
library(parallel)
library(purrr)




```


Write country shapefiles to raster for terrestrial systems

```{r}

# country shapefiles
countries_shp <- ne_countries(scale = "medium", type = "countries", returnclass = "sf")

#crop feed countries
crop_countries <- countries_shp |> filter(!is.na(iso_a3)) |> pull(iso_a3) |> unique()

#check the crop from the demand data and teh shapefile are now all the same
length(crop_countries)
  
#if it returns null it's because the file already exists

map(.x = crop_countries, .f =  \(this_country){
  
  message("processing ", this_country)
  
  this_country_filepath <- here(sprintf("data/spatial/country-rasters/%s.tif", this_country))
  
  if(!file.exists(this_country_filepath)){
    
    this_country_map <- countries_shp |> filter(iso_a3 == this_country) |> mutate(iso_n3 = as.double(iso_n3))
    
    this_country_raster <- rasterize(x = vect(this_country_map), y = rast(res=0.5, crs = "EPSG:4326"))
    
    writeRaster(x = this_country_raster, filename = here(sprintf("data/spatial/country-rasters/%s.tif", this_country)), overwrite=TRUE)
    
  }
})





#clear the environment
rm(list = ls(all.names = TRUE))
```
