==> scripts/4_calculate_global_feed_demand.Rmd <==
---
title: "Calculate_global_feed_demand"
author: "Rich Cottrell"
date: "14/03/2022"
output: html_document
---


```{r}
library(tidyverse)
library(here)
library(janitor)
library(countrycode)

source(here("src/directories.R"))

select <- dplyr::select

```


Import aquaculture data and isolate salmon. The top 10 producers produce 99% of global production so we can just go with that.
```{r}

aquaculture <- readRDS(here("data/tidy_data/production-data/aquaculture_production_tidy.rds"))

aquaculture_species_groups <- unique(aquaculture$species)

salmon_species <- aquaculture_species_groups[grep("salmon", aquaculture_species_groups)]

salmon_aquaculture <- aquaculture %>% filter(species %in% salmon_species & year==2019)

(top_producers <- salmon_aquaculture %>% 
    group_by(country) %>% 
    summarise(value = sum(value, na.rm = TRUE)) %>% 
    arrange(-value) %>% 
    mutate(prop = value/sum(value),
           cum_prop = cumsum(prop)) %>% 
    slice(1:10)
    
    
    )


aquaculture %>% filter(country %in% top_producers$country) %>% filter(grepl("salmon", species)) %>% pull(species) %>% unique()

```


Import the feed conversion ratios and join to production. Where country information from tacon and metian is not available use global average for that country. And get absolute feed demand. To do this we use the min eFCR given the Tacon and Metian data is old and there is a general upward trend in feed efficiency through time.
```{r}

fcrs <- read_csv(here("data/raw_data/FCR.csv"))

feed_demand <- 
  top_producers %>% 
  left_join(fcrs, by = "country") %>% 
  mutate(eFCR_min = if_else(is.na(eFCR_min), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_min),
         eFCR_max = if_else(is.na(eFCR_max), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_max),
         eFCR_mean = if_else(is.na(eFCR_mean), true = fcrs %>% filter(country=="Globe") %>% pull(eFCR_mean), false = eFCR_mean)) %>% 
  mutate(feed_demand = value*eFCR_min)
  
saveRDS(object = feed_demand, file = here("data/tidy_data/demand/global_feed_demand.rds"))

```


Bring in diet data

```{r}

all_diets <- read_csv(here("data/raw_data/diet_formats.csv")) %>% dplyr::select(-Sources)

diet_summaries <- 
  all_diets %>% 
  filter(grepl("totals", groups)) %>% 
  dplyr::select(c(groups, !matches('[2-9]+')), -ingredients) %>% 
  set_names(~str_replace_all(., pattern = "_1", replacement = "s"))

saveRDS(object=diet_summaries, file = here("data/tidy_data/diets/diet_summaries.rds"))
  

all_diets_list <- 
  all_diets %>% 
  filter(!grepl("totals", groups)) %>% 
  filter(groups != "Totals") %>% 
  pivot_longer(names_to = "diet", values_to = "prop", -c(groups, ingredients)) %>% 
  arrange(diet) %>% 
  mutate(prop = prop/100) %>% 
  group_split(diet)
  

map(.x = all_diets_list,  .f = function(this_diet){ 
  this_file_name <- unique(this_diet$diet); 
  saveRDS(object = this_diet, here(paste0(sprintf("data/tidy_data/diet-scenarios/%s", this_file_name), ".rds")))
  })


```


Now calculate ingredient demand per diet and for each country

```{r}
diet_files <- list.files(here("data/tidy_data/diet-scenarios"), full.names = TRUE)

feed_demand_files <- readRDS(here("data/tidy_data/demand/global_feed_demand.rds")) %>% 
  dplyr::select(country, feed_demand) %>% 
  group_split(country)


#calculate demand for each country individually because of different FCRs

for(i in 1:length(feed_demand_files)){
  
  this_country_data <- feed_demand_files[[i]]
  
  this_country <- unique(this_country_data$country)
  
  this_countrys_feed_demand <- unique(this_country_data$feed_demand)
  
  map_df(diet_files, readRDS) %>% 
  mutate(country = this_country,
         ingredient_demand = this_countrys_feed_demand * prop) %>% 
  saveRDS(here(sprintf("data/tidy_data/ingredient-demand-by-country/ingredient_demand_%s.rds", this_country)))
  
}

#Summarise and store total feed demand per ingredient

list.files(here("data/tidy_data/ingredient-demand-by-country"), full.names = TRUE) %>% 
  map_df(readRDS) %>% 
  group_by(groups, ingredients, diet) %>% 
  summarise(total_ingredient_demand = sum(ingredient_demand, na.rm = TRUE)) %>% 
  saveRDS(here("data/tidy_data/demand/total_ingredient_demand.rds"))

```


Next step is convert ingredient demand into raw material demand which will vary by processing technique (e.g. SPC vs SBM) and source. So first, we need to know where the ingredients are largely sourced from. THis may change with information that we gain from Biomar but the idea is eventually to make this flexible enough to select any from the top producing countries. This requires knowing the raw material that supports the ingredient and the major places this raw material is grown.

```{r}

#create the ingredient list to match raw materials with ingredients

ingredient_list <- tibble(ingredient = readRDS(here("data/tidy_data/demand/total_ingredient_demand.rds")) |> pull(ingredients) |> unique()) |> 
  write_csv(here("data/tidy_data/diets/ingredient_list.csv"))


```


CROP INGREDIENTS

Have added codes and FAO_names for plant ingredients in teams so join these to production data. 

```{r}

ingredient_by_raw_material <- 
  read_csv(here("data/tidy_data/diets/plant_ingredient_codes.csv")) |> 
  select(-reference) |> 
  drop_na(ingredient) |> 
  add_row(ingredient = "pea starch", raw_material = "peas", FAOSTAT_name = "Peas, green", FAO_code = 187, map_spam_code = "opul") |> 
  mutate(FAOSTAT_name = if_else(FAOSTAT_name == "Peas, dry; Peas, green", true = "Peas, dry", false = FAOSTAT_name)) 
 

crop_production_raw <- readRDS(file = here("data/tidy_data/production-data/crops_production_tidy.rds"))

(crop_production <- 
  crop_production_raw |>
  filter(element == "Production"  & area!= "China" & item %in% ingredient_by_raw_material$FAOSTAT_name & year %in% seq(from = 2015, to = 2020)) |> 
  group_by(area, item) |> 
  summarise(mean_value = mean(value, na.rm = TRUE)) |> 
    ungroup() |> 
  mutate(area = case_when(grepl("Ivoire", area) ~ "Cote d'Ivoire",
                          TRUE ~ area),
         iso3c = countrycode(sourcevar = area, origin = "country.name", destination = "iso3c", warn = TRUE)) |> 
  filter(!is.na(iso3c)) |> 
    arrange(item,-mean_value)
)


#need to decide whether we should pick the top 10 producers for things like guar so countries like pakistan are included for pulses nes
top_10_crop_producers <- 
  crop_production |> 
  group_split(item) |> 
  map_df(\(.){ 
    #this takes the top ten producers - unless it is for Guar (Pulses nes) where I used the suggestion that most guar production comes from India, Pakistan and Africa (so subsequently took the top 8 African producers of pulses nes to make up the top 10)
    
    if(unique(.$item) == "Pulses nes"){
      . |> filter(area %in% c("India", "Pakistan", "Kenya", 
                              "Ethiopia", "United Republic of Tanzania", 
                              "Sudan", "Sierra Leone", "Nigeria", 
                              "Guinea",  "Central African Republic"))
      }else{
        . |>  slice(1:10)}
  }
)


write.csv(x = top_10_crop_producers, file = here("data/tidy_data/production-data/top_crop_producers.csv"))
saveRDS(object = top_10_crop_producers, file = here("data/tidy_data/production-data/top_crop_producers.rds"))



```


Now we have the top 10 producers for each crop that is used as a raw material in feed, we need to get conversion factors from ingredients to  each of these raw materials in different places. This was done separately in the Sustainable Aquafeeds Project Teams interface and subsequently transferred to 'tidy_data' in this project. This can be combined with the global demand 

```{r}

#import codes that will be compatible with MAPSPAM layers

plant_ingredient_codes <- 
  read_csv(here("data/tidy_data/diets/plant_ingredient_codes.csv")) |> 
  select(-reference) |> 
  drop_na(ingredient) |> 
  #add_row(ingredient = "pea starch", raw_material = "peas", FAOSTAT_name = "Peas, green", FAO_code = 187, map_spam_code = "opul") |> this row is for if you want pea starch to come from green peas and dried peas - just using dried peas for simplicity
  mutate(FAOSTAT_name = if_else(FAOSTAT_name == "Peas, dry; Peas, green", true = "Peas, dry", false = FAOSTAT_name))

#join codes to top 10 producers and conversion factors data
(top_10_crop_w_cf <- read_csv(here("data/tidy_data/production-data/top_crop_producers_conversions.csv")) |> 
    select(-X7, -mean_value) |> 
  left_join(plant_ingredient_codes |> select(-raw_material), by = c("ingredient", "item" = "FAOSTAT_name"))
)


#join the ingredient demand by diet to the top 10 producer data so for each ingredient there will be an raw material demand per source, and diet scenario. Create new raw_material demand column by dividing total ingredient demand by the conversion factor (i.e. the ingredient extraction rate from the raw material in the first place)
crop_raw_material_demand <- readRDS(here("data/tidy_data/demand/total_ingredient_demand.rds")) |> 
  ungroup() |> 
  left_join(top_10_crop_w_cf, by = c("ingredients" = "ingredient")) |> 
  drop_na(item) |> 
  rename(FAOSTAT_name = item,
         source_country = area,
         source_iso3c = iso3c) |> 
  mutate(total_crop_demand = total_ingredient_demand/cf)


saveRDS(object = crop_raw_material_demand, file = here("data/tidy_data/demand/total_crop_demand.rds"))
  

```
