==> scripts/05_spatial_crop_demand_allocation.Rmd <==
---
title: "Spatial Allocation of Crop Demand"
output: html_document
---

```{r}
library(tidyverse)
library(here)
library(terra)
library(parallel)
library(doParallel)

source(here("src/directories.R"))


base_raster_ea <- rast(here("data/spatial/base_raster_gall_peters.tif"))

select <- dplyr::select


#allocation method (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"


```


We're  interested from where the different crop ingredients could be sourced across the top producing nations. Need to check whether the demand from the global salmon industry can be met by crop producing countries

```{r}
crop_demand <- readRDS(here("data/tidy_data/demand/total_crop_demand.rds"))

crop_production <- readRDS(here("data/tidy_data/production-data/top_crop_producers.rds")) |> rename(national_production = mean_value) 

#Check that demand for each raw material on each country in the various diets does not outstrip national production - all possible

(crop_demand_production_compare <- 
  crop_demand |> 
    left_join(crop_production, by = c("FAOSTAT_name" = "item")) |> 
    mutate(demand_production_ratio = total_crop_demand/national_production,
           check = demand_production_ratio>1)
    )




saveRDS(object = crop_demand_production_compare, file = here("data/tidy_data/demand/crop_demand_production_compare.rds"))


```

Make some decisions about which countries to include. The idea is to deliberately select countries that have variable environmental midpoints to illustrate the relative contribution of source and diet. Import the agrifootprint data and filter by the countries that can support

```{r}
#import the crop lca data
all_crop_pressures <-  readRDS(file = here("data/tidy_data/LCA/crop_lca.rds"))

sort(unique(all_crop_pressures$impact))
sort(unique(all_crop_pressures$raw_material))

#filter for the pressures of interest
crop_main_pressures <- all_crop_pressures |> filter(impact %in% c("Global warming", "Water use",  "Water consumption", "Land use", "Freshwater eutrophication", "Marine eutrophication"))

#check that five pressures are available across all materials (i.e. GWP, FW eutrophication, marine eutrophication, land use, water use). 
crop_main_pressures |> 
  select(raw_material, impact) |> 
  distinct() |> 
  group_by(raw_material) |> 
  count()



#match up country-specific pressure data with country production/demand data. Want to constrain the countries used for LCA variability by those that can plausibly source demand

(plausible_crop_sourcing_list <- crop_demand_production_compare |>  group_split(ingredients))

which_crop_countries <- \(this_df){
  
  this_raw_material <- this_df|> pull(FAOSTAT_name) |> unique()
  these_countries <- this_df  |> pull(iso3c) |> unique()
  this_return_df <- bind_cols(FAOSTAT_name = this_raw_material, iso3c = these_countries, plausible = TRUE)
  
}

(plausible_crop_sourcing <- map_df(.x = plausible_crop_sourcing_list, .f = which_crop_countries) |> distinct())

#filter impact information for those it is plausible to source from. 
plausible_crops_w_impact <- crop_main_pressures |> filter(raw_material != "Lupins") |> 
  left_join(plausible_crop_sourcing,  by = c("iso3c", "FAOSTAT_name")) |> 
  drop_na(plausible) 


#save crop production with LCA
saveRDS(plausible_crops_w_impact, file = here("data/tidy_data/demand/crop_sourcing_countries_plausible_impacts_all_allocation_methods.rds"))


#are at least 5 countries available where they can plausibly source demand and with which we have impacts? Yes
plausible_crops_w_impact |> 
  group_by(raw_material, impact, allocation) |> 
  count()  # accounts for two allocation methods



#these countries for each crop create the greatest range of midpoint impacts across sources
crop_sourcing_countries <- 
    plausible_crops_w_impact |>
   filter(allocation == this_allocation_method) |> 
    ungroup() |> 
    #filter(raw_material != "Guar bean") |>
    group_by(raw_material, FAOSTAT_name, iso3c, allocation) |> 
    summarise(cum_value = sum(value)) |> 
    # summarise(min = min(cum_value),
    #           mid = if_else(condition = length(cum_value)%%2==0 & length(cum_value)>1, 
    #                         true = sort(cum_value)[length(cum_value)/2], 
    #                         false = median(cum_value)),
    #           max = max(cum_value)) |> 
  ungroup() |> 
  arrange(allocation, raw_material, -cum_value) |> 
  group_by(allocation) |> 
  group_split()

map(crop_sourcing_countries, \(this_df){
  
  saveRDS(object = this_df, file = sprintf(here("data/tidy_data/demand/crop_sourcing_countries_plausible_w_cum_impacts_%s.rds"), this_allocation_method))
  
  
})



  #   # add_row(raw_material = "Guar bean", FAOSTAT_name = "Pulses nes", min = guar_seed_cum_value, mid = guar_seed_cum_value, max = guar_seed_cum_value) |> 
  # # pivot_longer(names_to = "value_range", values_to = "cum_value", -c(raw_material, FAOSTAT_name)) |> 
  # left_join(plausible_crops_w_impact |> 
  #             group_by(raw_material, FAOSTAT_name, iso3c, allocation) |> 
  #             summarise(cum_value = sum(value)), by = c("raw_material", "FAOSTAT_name", "cum_value"))

#saveRDS(object = crop_sourcing_countries, file = here("data/tidy_data/demand/crop_sourcing_countries_plausible_impacts.rds"))
  


```


Now we can use the countries which can plausibly support salmon feed demands and for which we have midpoint impacts for to spatially allocate crop demand to these producing countries - rather than just using the top 3 producers for example.

```{r}
#import sourcing countries data and crop demand data if only running this chunk
crop_sourcing_countries<- readRDS(file = sprintf(here("data/tidy_data/demand/crop_sourcing_countries_plausible_w_cum_impacts_%s.rds"), this_allocation_method))


crop_demand_production_compare <- readRDS(here("data/tidy_data/demand/crop_demand_production_compare.rds"))

ingredient_per_country_per_diet_all <- 
  crop_demand_production_compare |>  
  filter(ingredient_demand_tonnes != 0) |> 
  group_split(diet, ingredients, FAOSTAT_name)


this_raw_material_source <- ingredient_per_country_per_diet_all[[1]]

#filter each combo of diet, ingredient and raw-material by the source countries decided above
ingredient_per_country_per_diet <- 
  
  map_df(ingredient_per_country_per_diet_all, \(this_raw_material_source){
  
  this_raw_material <- this_raw_material_source$FAOSTAT_name |> unique()
  
  
  #gonna hard code the countries we want to use based on 5 countries, plausibility, and wide range of midpoint impacts
  these_sources <- 
    case_when(this_raw_material == "Broad beans, horse beans, dry" ~ c("AUS", "ITA", "FRA", "GBR", "DEU"),
              this_raw_material == "Pulses nes" ~ "IND",
              this_raw_material == "Linseed" ~ c("RUS", "UKR", "CHN", "FRA", "CAN"),
              this_raw_material == "Maize" ~ c("USA", "CHN", "ARG", "UKR", "BRA"),
              this_raw_material == "Peas, dry" ~ c("AUS", "RUS", "CAN", "LTU", "FRA"),
              this_raw_material == "Rapeseed" ~ c("CAN", "POL", "DEU", "GBR", "AUS"),
              this_raw_material == "Soybeans" ~ c("USA", "BRA", "PRY", "CHN", "UKR"),
              this_raw_material == "Sunflower seed" ~ c("HUN", "USA", "ARG", "RUS", "ROU"),
              this_raw_material == "Wheat" ~ c("RUS", "UKR", "AUS", "DEU", "FRA"))

  
  filter_df <-  this_raw_material_source|> filter(iso3c %in% these_sources) |> 
    mutate(allocation = this_allocation_method)
  
}) |> 
  group_split(diet, ingredients, FAOSTAT_name, iso3c)


saveRDS(bind_rows(ingredient_per_country_per_diet), file = here("data/tidy_data/demand/sourcing_countries_crops.rds"))


#for testing the function below
this_ingredient_source <- ingredient_per_country_per_diet[[5]]

#allocating demand on spatial production layers by diet, ingredient, and source country
AllocateSpatialCropDemand <- \(this_ingredient_source){
  
  #parameters within diet-ingredient-source demand data
  this_diet <- unique(this_ingredient_source$diet)
  
  this_country <- unique(this_ingredient_source$iso3c)
  
  this_ingredient <- unique(this_ingredient_source$ingredients)
  
  this_mapspam_code <- unique(this_ingredient_source$map_spam_code)
  
  #the file_name of interest for this diet, ingredient, and country
  this_demand_raster_filename <- here(sprintf("data/spatial/%s/production/%s_AllTech_%s_%s.tif", this_diet, this_ingredient, this_allocation_method, this_country))
  
  message("processing ", this_diet, ", ", this_ingredient, ", ", this_allocation_method, ", ", this_country )
  
  #if the file doesn't exist, create it
  if(!file.exists(this_demand_raster_filename)){
    
    #spatial adjustments of production rasters by current production and applying demand for each diet
    
  
      
      this_demand_quantity <- unique(this_ingredient_source$total_crop_demand)
      
      this_country_production <- unique(this_ingredient_source$national_production)
      
      this_demand_production_ratio <- unique(this_ingredient_source$demand_production_ratio)
      
      this_country_raster <- rast(here(sprintf("data/spatial/country-rasters/%s.tif", this_country)))
      
      this_global_production_raster <- rast(list.files(here("data/spatial/crop-layers-reprojected"), pattern = sprintf("_P_%s_A", toupper(this_mapspam_code)), full.names = TRUE))
      
      this_country_production_raster <- this_global_production_raster*this_country_raster
      
      this_updated_country_production_raster <-  this_country_production_raster* this_country_production/sum(c(terra::values(this_country_production_raster, na.rm=TRUE)))
      
      this_demand_raster <- this_updated_country_production_raster*this_demand_production_ratio
      
      terra::writeRaster(x=this_demand_raster, filename =  this_demand_raster_filename, overwrite=TRUE)
      
    }
    
    
    
  }



map(.x = ingredient_per_country_per_diet, .f = AllocateSpatialCropDemand)

rm(list = ls())

```
