==> scripts/06_exploring_biodiversity_shapefiles.Rmd <==
---
title: "Exploring biodiversity files"
output: html_document
---

This script explores the raw biodiversity shapefiles from IUCN and birds of the world stored on the rdsi storage.

Libraries
```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(sf)
library(data.table)
library(dtplyr)
library(terra)
library(parallel)

source(here("src/directories.R"))

source(here("src/spatial.R"))

base_raster_ea <- rast(here("data/spatial/base_raster_gall_peters.tif"))


```



Convert birds of the world from gdb to shapefile. This works on my local device but not on my NECTAR Cloud RStudio, weirdly. So copied the exported version to the RDSI storage in 'iucn_dir'. 
```{r}



birds_layers <- st_layers(file.path(iucn_dir, "spp_shapefiles/BOTW/BOTW.gdb"))

#this is the directory used on NECTAR with RDSI mount
birds_gdb_file <- file.path(iucn_dir, "spp_shapefiles/BOTW/BOTW.gdb")

#this is the directory used for Rich's local machine
birds_local_gdb_file <- file.path(iucn_rich_local_dir, "BOTW/BOTW.gdb")

botw <- st_read(dsn = birds_local_gdb_file, layer = "All_Species")

n_groups <- 9
(len_groups <- (nrow(botw)/n_groups))


botw %>% 
  mutate(index = rep(1:n_groups, each=len_groups)) %>% 
  group_by(index) %>% 
  nest() %>% 
  mutate(txt = walk2(.x = data, .y = index, ~st_write(obj = .x, dsn = paste0(iucn_rich_local_dir, "/BOTW/BOTW_PART_", .y, ".shp"))))



```



Create all shapefiles for each taxon
```{r}

taxon_list <- list.files(file.path(iucn_dir, "spp_shapefiles"))[-grep(pattern = "BOTW|hybas_table|FW_", list.files(file.path(iucn_dir, "spp_shapefiles")))] #excluding birds of the world and hybas tables and fw species. Botw has a different format which means iteration fails; hybas tables are needed to intersect the range maps; the range of some fw species is smaller than one pixel so rast() fails 

#create taxon directories
map(.x = taxon_list, \(this_taxon){
  dir.create(paste0("/mnt/rdsi/raw_data/iucn/spp_rasters/", this_taxon))
})


#This can take a long time to run and was killing memory trying tun all at once - so splitting it into groups of three to keep manageable. Sometimes have to rerun when results from some cores fail


#########################################################################################################################


#RUN 1
map(.x = taxon_list[1:5], .f = \(this_taxon){
  
  message("processing ", this_taxon) 
  
  #read in shp files - can be multiple per taxon
  this_taxons_shpfiles <- list.files(file.path(iucn_dir, sprintf("spp_shapefiles/%s/", this_taxon)), pattern = "\\.shp$")
  
  #read in each in the list of shp files
    
  this_taxon_shp_list <- mclapply(X=this_taxons_shpfiles, 
                       FUN = \(this_shp){
                         st_read(dsn = paste0("/mnt/rdsi/raw_data/iucn/spp_shapefiles/", this_taxon) , layer =  tools::file_path_sans_ext(this_shp))  
                       },
                       mc.cores=detectCores()-2)
  
  #filter for extant species and those we know to have ongoing (or past likely to reoccur threats that cause at least slow but substantial declines)
  this_taxon_shp <- bind_rows(this_taxon_shp_list) |> 
    filter(grepl("Extant", legend) & presence == 1) #|> 
    #filter(id_no %in% current_severe_threats_of_interest$id_no)
  
  #isolate the species
  these_spp <- unique(this_taxon_shp$id_no)
  
  
  #cycle through the species
  mclapply(X = these_spp, FUN = \(this_spp){
    
    spp_raster_filepath <- file.path(iucn_dir, sprintf("spp_rasters/%s/spp_%s.tif", this_taxon, this_spp))
    
    #does this file path exist? If not create the file
    if(!file.exists(spp_raster_filepath)){
      
      this_spp_shp <- this_taxon_shp |> filter(id_no == this_spp) |> st_transform(crs = equal_area_gp_proj)
      
      this_spp_raster <- rasterize(vect(this_spp_shp), base_raster_ea, touches = TRUE)
      
      writeRaster(x = this_spp_raster, filename = spp_raster_filepath, overwrite=TRUE)
      }
  },
  
  mc.cores = detectCores()-2)}
  )


#########################################################################################################################
#RUN 2

map(.x = taxon_list[6:10], .f = \(this_taxon){
  
  message("processing ", this_taxon) 
  
  #read in shp files - can be multiple per taxon
  this_taxons_shpfiles <- list.files(file.path(iucn_dir, sprintf("spp_shapefiles/%s/", this_taxon)), pattern = "\\.shp$")
  
  #read in each in the list of shp files
    
  this_taxon_shp_list <- mclapply(X=this_taxons_shpfiles, 
                       FUN = \(this_shp){
                         st_read(dsn = paste0("/mnt/rdsi/raw_data/iucn/spp_shapefiles/", this_taxon) , layer =  tools::file_path_sans_ext(this_shp))  
                       },
                       mc.cores=detectCores()-2)
  
  
  #filter for extant species and those we know to have ongoing (or past likely to reoccur threats that cause at least slow but substantial declines)
  this_taxon_shp <- bind_rows(this_taxon_shp_list) |> 
    filter(grepl("Extant", legend) & presence == 1) #|> 
    #filter(id_no %in% current_severe_threats_of_interest$id_no)
  
  #isolate the species
  these_spp <- unique(this_taxon_shp$id_no)
  
  
  #cycle through the species
  mclapply(X = these_spp, FUN = \(this_spp){
    
    spp_raster_filepath <- file.path(iucn_dir, sprintf("spp_rasters/%s/spp_%s.tif", this_taxon, this_spp))
    
    #does this file path exist? If not create the file
    if(!file.exists(spp_raster_filepath)){
      
      this_spp_shp <- this_taxon_shp |> filter(id_no == this_spp) |> st_transform(crs = equal_area_gp_proj)
      
      this_spp_raster <- rasterize(vect(this_spp_shp), base_raster_ea , touches = TRUE)
      
      writeRaster(x = this_spp_raster, filename = spp_raster_filepath, overwrite=TRUE)
      }
  },
  
  mc.cores = detectCores()-2)}
  )


#########################################################################################################################
#RUN 3

map(.x = taxon_list[11:15], .f = \(this_taxon){
  
  message("processing ", this_taxon) 
  
  #read in shp files - can be multiple per taxon
  this_taxons_shpfiles <- list.files(file.path(iucn_dir, sprintf("spp_shapefiles/%s/", this_taxon)), pattern = "\\.shp$")
  
  #read in each in the list of shp files
    
  this_taxon_shp_list <- mclapply(X=this_taxons_shpfiles, 
                       FUN = \(this_shp){
                         st_read(dsn = paste0("/mnt/rdsi/raw_data/iucn/spp_shapefiles/", this_taxon) , layer =  tools::file_path_sans_ext(this_shp))  
                       },
                       mc.cores=detectCores()-2)
  
  #filter for extant species and those we know to have ongoing (or past likely to reoccur threats that cause at least slow but substantial declines)
  this_taxon_shp <- bind_rows(this_taxon_shp_list) |> 
    filter(grepl("Extant", legend) & presence == 1) #|> 
    #filter(id_no %in% current_severe_threats_of_interest$id_no)
  
  #isolate the species
  these_spp <- unique(this_taxon_shp$id_no)
  
  
  #cycle through the species
  mclapply(X = these_spp, FUN = \(this_spp){
    
    spp_raster_filepath <- file.path(iucn_dir, sprintf("spp_rasters/%s/spp_%s.tif", this_taxon, this_spp))
    
    #does this file path exist? If not create the file
    if(!file.exists(spp_raster_filepath)){
      
      this_spp_shp <- this_taxon_shp |> filter(id_no == this_spp) |> st_transform(crs = equal_area_gp_proj)
      
      this_spp_raster <- rasterize(vect(this_spp_shp), base_raster_ea, touches = TRUE)
      
      writeRaster(x = this_spp_raster, filename = spp_raster_filepath, overwrite=TRUE)
      }
  },
  
  mc.cores = detectCores()-2)}
  )




#########################################################################################################################
#RUN 4

map(.x = taxon_list[16:20], .f = \(this_taxon){
  
  message("processing ", this_taxon) 
  
  #read in shp files - can be multiple per taxon
  this_taxons_shpfiles <- list.files(file.path(iucn_dir, sprintf("spp_shapefiles/%s/", this_taxon)), pattern = "\\.shp$")
  
  #read in each in the list of shp files
    
  this_taxon_shp_list <- mclapply(X=this_taxons_shpfiles, 
                       FUN = \(this_shp){
                         st_read(dsn = paste0("/mnt/rdsi/raw_data/iucn/spp_shapefiles/", this_taxon) , layer =  tools::file_path_sans_ext(this_shp))  
                       },
                       mc.cores=detectCores()-2)
  
  #filter for extant species and those we know to have ongoing (or past likely to reoccur threats that cause at least slow but substantial declines)
  this_taxon_shp <- bind_rows(this_taxon_shp_list) |> 
    filter(grepl("Extant", legend) & presence == 1) #|> 
    #filter(id_no %in% current_severe_threats_of_interest$id_no)
  
  #isolate the species
  these_spp <- unique(this_taxon_shp$id_no)
  
  
  #cycle through the species
  mclapply(X = these_spp, FUN = \(this_spp){
    
    spp_raster_filepath <- file.path(iucn_dir, sprintf("spp_rasters/%s/spp_%s.tif", this_taxon, this_spp))
    
    #does this file path exist? If not create the file
    if(!file.exists(spp_raster_filepath)){
      
      this_spp_shp <- this_taxon_shp |> filter(id_no == this_spp) |> st_transform(crs = equal_area_gp_proj)
      
      this_spp_raster <- rasterize(vect(this_spp_shp), base_raster_ea, touches = TRUE)
      
      writeRaster(x = this_spp_raster, filename = spp_raster_filepath, overwrite=TRUE)
      }
  },
  
  mc.cores = detectCores()-2)}
  )




#########################################################################################################################
#RUN 5

map(.x = taxon_list[21:25], .f = \(this_taxon){
  
  message("processing ", this_taxon) 
  
  #read in shp files - can be multiple per taxon
  this_taxons_shpfiles <- list.files(file.path(iucn_dir, sprintf("spp_shapefiles/%s/", this_taxon)), pattern = "\\.shp$")
  
  #read in each in the list of shp files
    
  this_taxon_shp_list <- mclapply(X=this_taxons_shpfiles, 
                       FUN = \(this_shp){
                         st_read(dsn = paste0("/mnt/rdsi/raw_data/iucn/spp_shapefiles/", this_taxon) , layer =  tools::file_path_sans_ext(this_shp))  
                       },
                       mc.cores=detectCores()-2)
  
  #filter for extant species and those we know to have ongoing (or past likely to reoccur threats that cause at least slow but substantial declines)
  this_taxon_shp <- bind_rows(this_taxon_shp_list) |> 
    filter(grepl("Extant", legend) & presence == 1) #|> 
    #filter(id_no %in% current_severe_threats_of_interest$id_no)
  
  #isolate the species
  these_spp <- unique(this_taxon_shp$id_no)
  
  
  #cycle through the species
  mclapply(X = these_spp, FUN = \(this_spp){
    
    spp_raster_filepath <- file.path(iucn_dir, sprintf("spp_rasters/%s/spp_%s.tif", this_taxon, this_spp))
    
    #does this file path exist? If not create the file
    if(!file.exists(spp_raster_filepath)){
      
      this_spp_shp <- this_taxon_shp |> filter(id_no == this_spp) |> st_transform(crs = equal_area_gp_proj)
      
      this_spp_raster <- rasterize(vect(this_spp_shp), base_raster_ea, touches = TRUE)
      
      writeRaster(x = this_spp_raster, filename = spp_raster_filepath, overwrite=TRUE)
      }
  },
  
  mc.cores = detectCores()-2)}
  )



#########################################################################################################################
#RUN 6

map(.x = taxon_list[26:length(taxon_list)], .f = \(this_taxon){
  
  message("processing ", this_taxon) 
  
  #read in shp files - can be multiple per taxon
  this_taxons_shpfiles <- list.files(file.path(iucn_dir, sprintf("spp_shapefiles/%s/", this_taxon)), pattern = "\\.shp$")
  
  #read in each in the list of shp files
    
  this_taxon_shp_list <- mclapply(X=this_taxons_shpfiles, 
                       FUN = \(this_shp){
                         st_read(dsn = paste0("/mnt/rdsi/raw_data/iucn/spp_shapefiles/", this_taxon) , layer =  tools::file_path_sans_ext(this_shp))  
                       },
                       mc.cores=detectCores()-2)
  
  #filter for extant species and those we know to have ongoing (or past likely to reoccur threats that cause at least slow but substantial declines)
  this_taxon_shp <- bind_rows(this_taxon_shp_list) |> 
    filter(grepl("Extant", legend) & presence == 1) #|> 
    #filter(id_no %in% current_severe_threats_of_interest$id_no)
  
  #isolate the species
  these_spp <- unique(this_taxon_shp$id_no)
  
  
  #cycle through the species
  mclapply(X = these_spp, FUN = \(this_spp){
    
    spp_raster_filepath <- file.path(iucn_dir, sprintf("spp_rasters/%s/spp_%s.tif", this_taxon, this_spp))
    
    #does this file path exist? If not create the file
    if(!file.exists(spp_raster_filepath)){
      
      this_spp_shp <- this_taxon_shp |> filter(id_no == this_spp) |> st_transform(crs = equal_area_gp_proj)
      
      this_spp_raster <- rasterize(vect(this_spp_shp), base_raster_ea, touches = TRUE)
      
      writeRaster(x = this_spp_raster, filename = spp_raster_filepath, overwrite=TRUE)
      }
  },
  
  mc.cores = detectCores()-2)}
  )










```
