==> scripts/08_fish_disturbance_pressures.Rmd <==
---
title: "Spatial Allocation of Forage fish demand from each diet and source"
output: html_document
---

# Overall approach

Using the marine footprint approach outlined in Cashion et al https://onlinelibrary.wiley.com/doi/full/10.1111/faf.12222 based off of the primary production required for catch in a given cell



We need to calculate this separately for fishmeal and oil as we are interested in the allocation of pressures to ingredients
  
```{r, setup}

library(tidyverse)
library(here)
library(terra)
library(sf)
library(data.table)
library(dtplyr)
library(future)
library(furrr)
#library(doParallel)
library(rfishbase)
library(RColorBrewer)

source(here("src/directories.R"))

select <- dplyr::select
resample <- terra::resample


#allocation method (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"


```


Equal area projection resources
```{r}

gall_peters <- "+proj=cea +lon_0=0 +x_0=0 +y_0=0 +lat_ts=45 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
#raster template 
base_rast <- rast(res=0.5)
ext(base_rast) <- c(-180, 180, -90, 90)
values(base_rast) <- 1:ncell(base_rast)


```


MARINE FOOTPRINT


Create a cell area xyz file for 0.5 res wgs84 raster

```{r}

cell_area <- terra::as.data.frame(cellSize(rast(res=0.5), unit="km"), xy=TRUE) |> rename(LonCentre = x, LatCentre = y)

```



Import net primary productivity data
```{r}

#make npp into an xy file

mean_npp <- rast(here("data/spatial/00-net-primary-productivity/npp_2013_2022_mean.tif")) |> focal(w=9, fun = mean, na.policy = "only", na.rm=TRUE)
max_npp <- rast(here("data/spatial/00-net-primary-productivity/npp_2013_2022_max.tif")) |> focal(w=9, fun = mean, na.policy = "only", na.rm=TRUE)

#mean 
mean_npp_xyz <- terra::as.data.frame(mean_npp, xy = TRUE) |> 
  rename(mean_npp_t_km2_yr=focal_mean) |> 
  rename(LonCentre = x, 
         LatCentre = y)
#max
max_npp_xyz <- terra::as.data.frame(max_npp, xy = TRUE) |> 
  rename(max_npp_t_km2_yr=focal_mean) |> 
  rename(LonCentre = x, 
         LatCentre = y)

```

Import ecosystem transfer-efficiency

```{r}

ecosystem_te <- read_csv(here("data/tidy_data/ecosystem-te/ecosystem_te.csv")) |> select(TE, fao_area_code) |> group_by(fao_area_code) |> summarise(te=mean(TE, na.rm = TRUE)) |> mutate(te= te/100)


```


Get forage fish trophic levels - needed for calculating the primary productivity required

```{r}

#Separate the catch records that have Genus and species.
forage_spp <- readRDS(here("data/large_data/spatial_forage_catch_2017.rds")) |> 
  mutate(binomial = stringi::stri_count(TaxonName, regex="\\S+")) |> 
  filter(binomial==2) |>  
  pull(TaxonName) |> 
  unique() 

this_sp <- forage_spp[[1]]


(binomial_tls <- rfishbase::ecology(species_list =  forage_spp, server = getOption("fishbase")) |>
  select(Species, DietTroph, FoodTroph) |> 
  mutate(DietTroph = if_else(is.na(DietTroph), true = FoodTroph, false = DietTroph)) |> 
  drop_na(DietTroph) |> 
  select(-FoodTroph) |> 
  distinct()) #both extracted manually from fishbase as not coming up from app search

  
#Add back the Genus only or family records and match to any species with the same Genus

(forage_higherorder <- readRDS(here("data/large_data/spatial_forage_catch_2017.rds")) |> mutate(binomial = stringi::stri_count(TaxonName, regex="\\S+")) |> filter(binomial==1) |>  pull(TaxonName) |> unique()
)

forage_nonbinomial_tls <- tibble(TaxonName = forage_higherorder, DietTroph = NA) |> 
  mutate(DietTroph = case_when(TaxonName %in% 
                                 c("Engraulidae", "Clupeiformes", "Stolephorus") ~ binomial_tls |> 
                                 filter(grepl("Engraulis",Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Scomber") ~ binomial_tls |> 
                                 filter(grepl("Scomber", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Sardinella") ~ binomial_tls |> 
                                 filter(grepl("Sardinella", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Rastrelliger") ~ binomial_tls |> 
                                 filter(grepl("Rastrelliger", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Trachurus") ~ binomial_tls |> 
                                 filter(grepl("Trachurus", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Harengula", "Clupeidae", "Clupeoids") ~ binomial_tls |> 
                                 filter(grepl("Clupea", Species)) |> 
                                 pull(DietTroph),
                               TaxonName %in% c("Atherinidae") ~ binomial_tls |> 
                                 pull(DietTroph) |> 
                                 mean())) |> 
  add_row(TaxonName = "Euphausia superba", DietTroph = 2.25) |> 
    add_row(TaxonName = "Clupea pallasii pallasii", DietTroph = 3.16)
                               
                               
                               
forage_tls <- 
  bind_rows(binomial_tls |> rename(TaxonName = Species),
          forage_nonbinomial_tls)

```


Get trimmings species trophic levels  - needed for calculating the primary productivity required


```{r}

#Separate the catch records that have Genus and species
(trim_spp <- readRDS(here("data/large_data/trimmings_spp_catch.rds")) |> 
   mutate(binomial = stringi::stri_count(TaxonName, regex="\\S+")) |> 
   filter(binomial==2) |> 
   pull(TaxonName) |> 
   unique())


  #test the function
this_sp <- forage_spp[[1]]


(trimmings_binomial_tls <- rfishbase::ecology(species_list =  trim_spp, server = getOption("fishbase")) |> 
  select(Species, DietTroph, FoodTroph) |> 
  mutate(DietTroph = if_else(is.na(DietTroph), true = FoodTroph, false = DietTroph)) |>
  select(-FoodTroph) |> 
  mutate(DietTroph = case_when(Species == "Engraulis capensis" ~ mean(c(3.14,3.12,2.51)), #mean of otehr Engraulis values
                               TRUE ~ DietTroph)) |> 
  distinct() |> 
    drop_na() |> 
    add_row(Species = "Theragra chalcogramma", DietTroph = 3.57)) #add Alaska pollack that fell out of the species filter

  
#Add back the Genus only or family records and match to any species with the same Genus

(trim_higherorder <- readRDS(here("data/large_data/trimmings_spp_catch.rds")) |> mutate(binomial = stringi::stri_count(TaxonName, regex="\\S+")) |> filter(binomial==1) |>  pull(TaxonName) |> unique()
)

trim_nonbinomial_tls <- tibble(TaxonName = trim_higherorder, DietTroph = NA) |> 
  mutate(DietTroph = case_when(TaxonName %in% 
                                 c("Engraulidae") ~ trimmings_binomial_tls |>
                                 filter(grepl("Engraulis",Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Sardinella") ~ trimmings_binomial_tls |> 
                                 filter(grepl("Sardinella", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Clupeidae") ~ trimmings_binomial_tls |> 
                                 filter(grepl("Clup", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean(),
                               TaxonName %in% c("Thunnus") ~ trimmings_binomial_tls |> 
                                 filter(grepl("Thunnus", Species)) |> 
                                 pull(DietTroph) |> 
                                 mean()))
                               
                               
                               
trim_tls <- 
  bind_rows(trimmings_binomial_tls |> rename(TaxonName = Species),
          trim_nonbinomial_tls)


```



Attach NPP, TL, data to forage and trimmings catch data and calculate the primary productivity required for the catch, the proportion of the annual PPR that is and tmultiply this proportion by the size of the cell.    

```{r}
carbon_conversion <- 9

(forage_npp <- 
    fread(here("data/large_data/embodied_foragefish_per_cell.csv")) |> 
    lazy_dt(immutable = FALSE) |> 
    left_join(ecosystem_te, by="fao_area_code") |> 
    left_join(forage_tls, by = "TaxonName") |>
    mutate(te = case_when(is.na(te) ~ ecosystem_te |> pull(te) |> mean(),
                          TRUE ~ te)) |> 
    left_join(mean_npp_xyz, by = c("LonCentre", "LatCentre")) |> 
    left_join(max_npp_xyz, by = c("LonCentre", "LatCentre")) |> 
    left_join(cell_area) |> 
    mutate(mean_npp_t_C_yr = mean_npp_t_km2_yr*area) |> 
    mutate(max_npp_t_C_yr = max_npp_t_km2_yr*area) |>
    as_tibble() |> 
    mutate(ppr_fm_mass = (fm_embodied_fish/carbon_conversion)*(1/te)^(DietTroph-1), #T C yr-1
           ppr_fo_mass = (fo_embodied_fish/carbon_conversion)*(1/te)^(DietTroph-1)) |> 
    group_by(diet, fao_area_code, LonCentre, LatCentre) |> 
    summarise(mean_npp_t_C_yr = mean(mean_npp_t_C_yr) ,
              max_npp_t_C_yr = mean(max_npp_t_C_yr),
              ppr_fm = sum(ppr_fm_mass),
              ppr_fo = sum(ppr_fo_mass)) |> 
    ungroup() |> 
    mutate(ppr_prop_fm = case_when(ppr_fm/max_npp_t_C_yr<1 ~ ppr_fm/mean_npp_t_C_yr,
                                        TRUE ~ ppr_fm/max_npp_t_C_yr),
           
           ppr_prop_fo = case_when(ppr_fo/max_npp_t_C_yr<1 ~ ppr_fo/mean_npp_t_C_yr,
                                        TRUE ~ ppr_fo/max_npp_t_C_yr)) |> 
    arrange(LonCentre, LatCentre) |> 
    left_join(cell_area) |> 
    mutate(area_fm = ppr_prop_fm*area,
           area_fo = ppr_prop_fo*area) |> 
    mutate(source = "forage"))



(trim_npp <- 
  fread(here("data/large_data/embodied_fishfromtrimmings_per_cell.csv")) |> 
  left_join(ecosystem_te, by="fao_area_code") |> 
  left_join(trim_tls, by = "TaxonName") |>
  mutate(te = case_when(is.na(te) ~ ecosystem_te |> pull(te) |> mean(),
                        TRUE ~ te)) |> 
  left_join(mean_npp_xyz) |> 
    left_join(max_npp_xyz) |> 
    left_join(cell_area) |> 
    #work around for missing npp data for cod
    group_by(fao_area_code) |> 
    mutate_at(vars(mean_npp_t_km2_yr, max_npp_t_km2_yr), 
            ~replace_na(., 
                        mean(., na.rm = TRUE))) |> 
    ungroup() |> 
    mutate(mean_npp_t_C_yr = mean_npp_t_km2_yr*area) |> 
    mutate(max_npp_t_C_yr = max_npp_t_km2_yr*area) |>
    mutate(ppr_fm = (fm_embodied_fish/carbon_conversion)*(1/te)^(DietTroph-1), #T C yr-1
           ppr_fo = (fo_embodied_fish/carbon_conversion)*(1/te)^(DietTroph-1)) |> 
    as_tibble() |> 
    group_by(diet, fao_area_code, LonCentre, LatCentre) |> 
    summarise(mean_npp_t_C_yr = mean(mean_npp_t_C_yr) ,
              max_npp_t_C_yr = mean(max_npp_t_C_yr),
              ppr_fm = sum(ppr_fm),
              ppr_fo = sum(ppr_fo)) |> 
    ungroup() |> 
     mutate(ppr_prop_fm = case_when(ppr_fm/max_npp_t_C_yr<1 ~ ppr_fm/mean_npp_t_C_yr,
                                        TRUE ~ ppr_fm/mean_npp_t_C_yr),
           
           ppr_prop_fo = case_when(ppr_fo/max_npp_t_C_yr<1 ~ ppr_fo/mean_npp_t_C_yr,
                                        TRUE ~ ppr_fo/mean_npp_t_C_yr)) |> 
         arrange(LonCentre, LatCentre) |> 
    left_join(cell_area) |> 
    mutate(area_fm = ppr_prop_fm*area,
           area_fo = ppr_prop_fo*area) |> 
    mutate(source = "trimmings")
    ) 


#Save the data for TLs only

(forage_tl_by_area <- 
  fread(here("data/large_data/embodied_foragefish_per_cell.csv")) |> 
    lazy_dt(immutable = FALSE) |> 
    left_join(ecosystem_te, by="fao_area_code") |> 
    left_join(forage_tls, by = "TaxonName") |> 
    select(diet, fao_area_code, CommonName, total_embodied_fish, DietTroph) |> 
  as_tibble())

(trim_tl_by_area <- 
  fread(here("data/large_data/embodied_fishfromtrimmings_per_cell.csv")) |> 
  lazy_dt(immutable = FALSE) |> 
  left_join(ecosystem_te, by="fao_area_code") |> 
  left_join(trim_tls, by = "TaxonName") |> 
    select(diet, fao_area_code, CommonName, total_embodied_fish, DietTroph) |> 
  as_tibble()
)

all_tls_by_area <- 
  bind_rows(forage_tl_by_area, trim_tl_by_area) |> 
  mutate(weighted_tl = total_embodied_fish*DietTroph)

saveRDS(object = all_tls_by_area, file = sprintf(here("data/large_data/fish_tls_by_area_%s.rds"), this_allocation_method))

#save the prop of npp files for later

prop_npp <- bind_rows(forage_npp |> select(diet, fao_area_code,LonCentre, LatCentre,mean_npp_t_C_yr, ppr_fm, ppr_fo, ppr_prop_fm, ppr_prop_fo, source),
            trim_npp |> select(diet, fao_area_code,LonCentre, LatCentre, mean_npp_t_C_yr, ppr_fm, ppr_fo, ppr_prop_fm, ppr_prop_fo, source))


saveRDS(object = prop_npp, file = sprintf(here("data/large_data/prop_npp_%s.rds"), this_allocation_method))





#Bind the forage and trimmings catch npp per cell

all_fish_npp <- 
  bind_rows(forage_npp |> select(diet, fao_area_code,LonCentre, LatCentre, area_fm, area_fo, source),
            trim_npp |> select(diet, fao_area_code,LonCentre, LatCentre, area_fm, area_fo, source)) |> 
  group_by(diet, fao_area_code, LonCentre, LatCentre) |> 
  summarise(area_fm = sum(area_fm),
            area_fo = sum(area_fo)) |> 
  left_join(cell_area)


saveRDS(object = all_fish_npp, file = sprintf(here("data/large_data/fish_disturbance_join_%s.rds"), this_allocation_method))


#km per tonne of fishmeal/oil estimates
all_fish_npp |> filter(fao_area_code == "27" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE)/ (38400)
all_fish_npp |> filter(fao_area_code == "31" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE)/ (38400)
all_fish_npp |> filter(fao_area_code == "61" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE)/ (38400)
all_fish_npp |> filter(fao_area_code == "67" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE)/ (38400)
all_fish_npp |> filter(fao_area_code == "87" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE)/ (38400)



#fishmeal trim
trim_npp |> filter(fao_area_code == "27" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE)/ (38400*0.3)

#fo trim 
trim_npp |> filter(fao_area_code == "27" & diet == "marine_diet") |> 
  pull(area_fo)  |> sum(na.rm=TRUE) /(27600*0.4)


#fishmeal forage
forage_npp |> filter(fao_area_code == "87" & diet == "marine_diet") |> 
  pull(area_fm)  |> sum(na.rm=TRUE) / (38400)

#fo forage
forage_npp |> filter(fao_area_code == "87" & diet == "marine_diet") |> 
  pull(area_fo)  |> sum(na.rm=TRUE) / (27600*0.6)

```


Now export the km2 equivalent as rasters

```{r}

# Split the dataframe by diet and fao area
all_fish_npp_list <- 
  all_fish_npp |> 
  group_by(diet, fao_area_code) |> 
  group_split()


#test function
this_diet_fao_area <- all_fish_npp_list[[3]]


#spit out the disturbance rasters for fo and fm under both mass and energetic allocation

map(.x = all_fish_npp_list, .f = \(this_diet_fao_area){
  
  this_diet <- this_diet_fao_area |> pull(diet) |> unique()
  
  this_fao_area_code <- this_diet_fao_area |> pull(fao_area_code) |> unique()
  
  
  # FISHMEAL DISTRUBANCE 
  
  message("Processing fishmeal disturbance raster - ", this_diet, " - fao area ", this_fao_area_code, " - ",  this_allocation_method)
  
  fm_saveName <- sprintf(here("data/spatial/%s/int/fishmeal_disturbance_km2_%s_%s.tif"), this_diet, this_allocation_method, this_fao_area_code)
  
  fm_area_rast <- this_diet_fao_area |> 
    select(LonCentre, LatCentre, area_fm) |> 
    rast(type = "xyz", crs(rast(res=0.5))) |> 
    resample(rast(res=0.5))
  
  
  fm_area_km2_rast <- fm_area_rast/cellSize(fm_area_rast, unit = "km")
  
  #output area raster
  
  writeRaster(x= fm_area_km2_rast, filename = fm_saveName, overwrite=TRUE)
  
  
  # FISH OIL DISTURBANCE
  
  message("Processing fish oil disturbance raster - ", this_diet, " - fao area ", this_fao_area_code, " - ",  this_allocation_method)
  
  fo_saveName <- sprintf(here("data/spatial/%s/int/fish oil_disturbance_km2_%s_%s.tif"), this_diet, this_allocation_method, this_fao_area_code)
  
  fo_area_rast <- this_diet_fao_area |> 
    select(LonCentre, LatCentre, area_fo) |> 
   rast(type = "xyz", crs(rast(res=0.5))) |> 
   resample(rast(res=0.5))

  fo_area_km2_rast <- fo_area_rast/cellSize(fo_area_rast, unit = "km")
  
  writeRaster(x= fo_area_km2_rast, filename = fo_saveName, overwrite=TRUE)

  
})





```




Some supplementary plots of the fish embodied by zone

```{r}

#Fishmeal - Mass allocation

fm_stack <- list.files(here("data/spatial/marine_diet/int"), pattern = paste0("fishmeal_disturbance_km2_", this_allocation_method), full=TRUE) |> rast()

fm_all <- app(fm_stack, sum, na.rm=TRUE)



ggplot()+
  geom_tile(data = fm_all |> 
              terra::as.data.frame(xy=TRUE) |> filter(sum>0), aes(x=x, y=y, fill=(sum+1)))+
  geom_sf(data = rnaturalearth::ne_countries(scale = "medium", returnclass = "sf"), fill="grey90", colour="grey90")+
  scale_fill_gradientn(colours = brewer.pal(n=8, "BuPu"),
                       trans = "log10")+
  theme(panel.background = element_rect(fill="transparent"))+
  labs(fill=bquote(Marine~footprint~km^2), title = "Disturbance, fishmeal, mass allocation")

ggsave(file = here("figures/presos/fishmeal_forage_mass_disturbance.jpg"), height = 4, width=7)



```


#Clear environment

```{r}
rm(list = ls())

```
