==> scripts/09_fish_GHG_pressures.Rmd <==
---
title: "08_forage_fish_GHG_pressures"
author: "Rich Cottrell"
date: "29/06/2022"
output: html_document
---

```{r}
library(tidyverse)
library(terra)
library(sf)
library(data.table)
library(dtplyr)
library(here)

source(here("src/directories.R"))

select <- dplyr::select


#allocation method (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"

```

Bringing in the emission intensity data and tidy to distinct categories with functional groups associated with the Watson data.
```{r}

ei_data <- read_csv(here("data/large_data/all_catch_emissions_2017.csv")) 
 

```

Now join the embodied fish demand xy data with the emission intensity data and calculate the emissions from each gear in each cell. Then we can save the rasters for each diet and source, for both fishmeal and oil ingredients

```{r}
embodied_fish_trim <- fread(here("data/large_data/embodied_fishfromtrimmings_per_cell.csv"))
embodied_fish_forage <- fread(here("data/large_data/embodied_foragefish_per_cell.csv"))

#take a look at the differences between gears for watson data and the ei data 
unique(ei_data$ParkerGearName)
unique(embodied_fish_trim$VBDesc)

#Gear classifications are different. So we need the join file with Parker Data.
parker_gear_join <- read_csv(here("data/raw_data/fisheries/parker_gear_join.csv"))

#now join parker gear data to the embodied fish by the parker gear categories, species and description from watson. Adjust the emissions intensities to reflect per tonne pressures

#trimmings

embodied_fish_trim_new_gear <- 
  embodied_fish_trim |> 
  lazy_dt(immutable = FALSE) |> 
  left_join(parker_gear_join, by="Gear") |> 
  left_join(ei_data, by = c("Taxonkey", "Descript", "ParkerGearName")) |> 
  as_tibble() |>
  mutate(ei_kgco2_tonnecatch = ei_kgco2_kgcatch*1000)|> 
  mutate(ei_kg_CO2_eq_fm = fm_embodied_fish*ei_kgco2_tonnecatch,
         ei_kg_CO2_eq_fo = fo_embodied_fish*ei_kgco2_tonnecatch)

embodied_fish_trim_new_gear |> filter(is.na(ei_kgco2_kgcatch))
#no missing emissions intensity data - what we want.



#forage fish

embodied_fish_forage_new_gear <- 
  embodied_fish_forage |> 
  lazy_dt(immutable = FALSE) |> 
  left_join(parker_gear_join, by="Gear") |> 
  left_join(ei_data|> 
  filter(Descript %in% c("pelagic <30 cm",  "pelagic 30 - 90 cm",  "benthopelagic 30 - 90 cm", "bathypelagic <30 cm" , "bathypelagic 30 - 90 cm", "krill" )), by = c("Taxonkey", "Descript", "ParkerGearName")) |> #limit the description to these for render fisheries
  as_tibble() |>
  mutate(ei_kgco2_tonnecatch = ei_kgco2_kgcatch*1000)|> 
  mutate(ei_kg_CO2_eq_fm = fm_embodied_fish*ei_kgco2_tonnecatch,
         ei_kg_CO2_eq_fo = fo_embodied_fish*ei_kgco2_tonnecatch)

#no missing emissions intensity data - what we want.
embodied_fish_forage_new_gear |> filter(is.na(ei_kgco2_kgcatch))



#check the embedded fish across the fm and fo per diet
embodied_fish_trim_new_gear |> 
  select(diet, fao_area_code, fm_embodied_fish, fo_embodied_fish) |> 
  group_by(diet, fao_area_code) |> 
  summarise(fo_embodied = sum(fo_embodied_fish, na.rm=TRUE),
            fm_embodied = sum(fm_embodied_fish, na.rm = TRUE))

embodied_fish_forage_new_gear |> 
  select(diet, fao_area_code, fm_embodied_fish, fo_embodied_fish) |> 
  group_by(diet, fao_area_code) |> 
  summarise(fo_embodied = sum(fo_embodied_fish, na.rm=TRUE),
            fm_embodied = sum(fm_embodied_fish, na.rm = TRUE))



#combine the trimmings and forage data to break into diet and fao_area later

embodied_fish_new_gear <- 
  
  bind_rows(embodied_fish_forage_new_gear |> 
            select(diet, fao_area_code, Gear,TaxonName, LonCentre, LatCentre, VBDesc, 
                   ei_kgco2_tonnecatch, ei_kg_CO2_eq_fm, ei_kg_CO2_eq_fo),
          
          embodied_fish_trim_new_gear |> select(diet, fao_area_code, Gear, TaxonName, LonCentre, LatCentre, VBDesc, 
                   ei_kgco2_tonnecatch, ei_kg_CO2_eq_fm,  ei_kg_CO2_eq_fo))


saveRDS(object = embodied_fish_new_gear, file = sprintf(here("data/large_data/embodied_fish_ghgs_%s.rds"), this_allocation_method))

#with tonnage embodied involved
embodied_fish_new_gear_with_catch <- 
  bind_rows(embodied_fish_forage_new_gear |> 
            select(diet, fao_area_code, Gear,TaxonName, LonCentre, LatCentre, VBDesc, fm_embodied_fish, fo_embodied_fish,
                   ei_kgco2_tonnecatch, ei_kg_CO2_eq_fm, ei_kg_CO2_eq_fo),
          
          embodied_fish_trim_new_gear |> select(diet, fao_area_code, Gear, TaxonName, LonCentre, LatCentre, VBDesc, fm_embodied_fish, fo_embodied_fish,
                   ei_kgco2_tonnecatch, ei_kg_CO2_eq_fm,  ei_kg_CO2_eq_fo))
 
 saveRDS(object = embodied_fish_new_gear_with_catch, file = sprintf(here("data/large_data/embodied_fish_ghgs_with_embodied_fish_%s.rds"), this_allocation_method))

 


```




# Fishmeal
```{r}

ghg_list_fm <- 
  embodied_fish_new_gear |> 
  select(diet, fao_area_code, LonCentre, LatCentre, ei_kg_CO2_eq_fm) |> 
  group_by(diet, fao_area_code, LonCentre, LatCentre) |> 
  summarise(total_kgCO2_fm = sum(ei_kg_CO2_eq_fm, na.rm=TRUE)) |> 
  ungroup() |>
  group_by(diet, fao_area_code) |> 
  group_split()

#testing function
this_df <- ghg_list_fm[[1]] |> pull(total_kgCO2_fm) |> sum()

ghg_raw_r_list_fm <- 
  
  map(.x = ghg_list_fm, .f = \(this_df) {
  
  this_source <- unique(this_df$fao_area_code)
  
  this_diet <- unique(this_df$diet)
  
  saveName <- sprintf(here("data/spatial/%s/int/fishmeal_ghg_km2_%s_%s.tif"), this_diet, this_allocation_method, this_source)
  
  if(!file.exists(saveName)){
    
    this_xy <- this_df |> select(LonCentre, LatCentre, total_kgCO2_fm)
    
    this_r <- rast(this_xy, type = "xyz", crs = "EPSG:4326") |> extend(ext(rast(res=0.5)))
    
    this_r_km2 <- this_r/cellSize(this_r, unit="km")
    
    writeRaster(x = this_r_km2, filename = saveName, overwrite=TRUE)
    
    return(this_r_km2)
  }
})

#check the rasters
#plot(rast(ghg_raw_r_list_fm))

```

# Fish oil

```{r}

ghg_list_fo <- 
  embodied_fish_new_gear |> 
  select(diet, fao_area_code, LonCentre, LatCentre, ei_kg_CO2_eq_fo) |> 
  group_by(diet, fao_area_code, LonCentre, LatCentre) |> 
  summarise(total_kgCO2_fo = sum(ei_kg_CO2_eq_fo, na.rm=TRUE)) |> 
  ungroup() |> 
  group_by(diet, fao_area_code) |> 
  group_split()



ghg_raw_r_list_fo <- 
  
  map(.x = ghg_list_fo, .f = \(this_df) {
  
  this_source <- unique(this_df$fao_area_code)
  
  this_diet <- unique(this_df$diet)
  
  
  saveName <- sprintf(here("data/spatial/%s/int/fish oil_ghg_km2_%s_%s.tif"), this_diet, this_allocation_method, this_source)
  
  if(!file.exists(saveName)){
    
    this_xy <- this_df |> select(LonCentre, LatCentre, total_kgCO2_fo)
    
    this_r <- rast(this_xy, type = "xyz", crs = "EPSG:4326") |> extend(ext(rast(res=0.5)))
    
    this_r_km2 <- this_r/cellSize(this_r, unit="km")
    
    writeRaster(x = this_r_km2, filename = saveName, overwrite=TRUE)
    
    return(this_r_km2)
  }
})

#check the rasters
#plot(rast(ghg_raw_r_list_fo))



```



Table for supplementary summary of gears and emissions for Northwest Pacific (61) and Southeast Pacific (87)
```{r}

# trimmings_species <- readRDS(file = here("data/large_data/trimmings_spp_catch.rds")) |> pull(TaxonName) |> unique()
# forage_species <- readRDS(file = here("data/large_data/spatial_forage_catch_2017.rds")) |> pull(TaxonName) |> unique()
# 
# 
# emissions_summary <- 
#   readRDS(sprintf(here("data/large_data/embodied_fish_ghgs_with_embodied_fish_%s.rds"), this_allocation_method)) |> 
#   filter(fao_area_code %in% c(61, 87)) |> 
#   group_by(diet, fao_area_code, TaxonName, VBDesc) |> 
#   summarise(fm_emissions_total = sum(ei_kg_CO2_eq_fm, na.rm=TRUE),
#             fo_emissions_total = sum(ei_kg_CO2_eq_fo, na.rm = TRUE))
# 
# savesummary <- \(x){
#   this_diet <- x$diet |> unique()
#   this_area <- x$fao_area_code |> unique()
#   write_csv(x, file = sprintf(here("explore/ghg_summaries_%s_%s_%s.csv"), this_allocation_method, this_diet, this_area))
# }
# 
# emissions_summary |> 
#   ungroup() |> 
#   filter(fm_emissions_total>0 | fo_emissions_total>0) |> 
#   arrange(-fm_emissions_total) |> 
#   mutate(source = case_when(fm_emissions_total == 0 ~ "Trimmings",
#                             TRUE~ "Whole fish")) |>
#   group_by(diet, fao_area_code) |> 
#   group_split() |> 
#   map(savesummary)

```
```{r}
#clear the environment
rm(list = ls(all.names = TRUE))
```
