==> scripts/10_fish_nutrient_pressures.Rmd <==
---
title: "09_forage_fish_nutrient_pressures"
author: "Rich Cottrell"
date: "15/07/2022"
output: html_document
---


```{r setup}

library(tidyverse)
library(terra)
library(here)

source(here("src/directories.R"))

select <- dplyr::select



#allocation method (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"



```

For each diet and source import the rasters from ghg and make all values zero - then re-export into "int" folder with nutrient file names - for both nitrogen and phosphorus

# Nitrogen pollution

Fishmeal

```{r}

fm_template_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fishmeal_ghg_km2", full.names = TRUE)
fm_template_files <- fm_template_files[grepl(this_allocation_method, fm_template_files)]

this_file <- fm_template_files[[3]]

map(.x = fm_template_files, .f = \(this_file){
  
  filename <- tools::file_path_sans_ext(this_file)
  
  this_source <- str_sub(filename, start = nchar(filename)-1, end = nchar(filename))
  
  saveName_N_md <- sprintf(here("data/spatial/marine_diet/int/fishmeal_N_km2_%s_%s.tif"), this_allocation_method, this_source)
  saveName_N_pd <- sprintf(here("data/spatial/plant_diet/int/fishmeal_N_km2_%s_%s.tif"), this_allocation_method, this_source)
  
  if(!file.exists(saveName_N_md) | !file.exists(saveName_N_pd)){
    
    this_r <- rast(this_file)*0
    names(this_r) <- "total_N_km2"
    
    #save N rasters
    terra::writeRaster(x = this_r, filename = saveName_N_md, overwrite = TRUE)
    
    terra::writeRaster(x = this_r, filename = saveName_N_pd, overwrite = TRUE)
    
  }
  
})

```

Fish oil

```{r}

fo_template_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fish oil_ghg", full.names = TRUE)
fo_template_files <- fo_template_files[grepl(this_allocation_method, fo_template_files)]


this_file <- fo_template_files[[3]]


map(.x = fo_template_files, .f = \(this_file){
  
  filename <- tools::file_path_sans_ext(this_file)
  
   this_source <- str_sub(filename, start = nchar(filename)-1, end = nchar(filename))
  
  saveName_N_md <- sprintf(here("data/spatial/marine_diet/int/fish oil_N_km2_%s_%s.tif"), this_allocation_method, this_source)
  saveName_N_pd <- sprintf(here("data/spatial/plant_diet/int/fish oil_N_km2_%s_%s.tif"), this_allocation_method, this_source)
  
  if(!file.exists(saveName_N_md) | !file.exists(saveName_N_pd)){
    
    this_r <- rast(this_file)*0
    names(this_r) <- "total_N_km2"
    
    #save N rasters
    terra::writeRaster(x = this_r, filename = saveName_N_md, overwrite = TRUE)
    
    terra::writeRaster(x = this_r, filename = saveName_N_pd, overwrite = TRUE)
    
  }
  
})

```



# Phosphorus pollution

Fishmeal

```{r}
fm_template_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fishmeal_ghg_km2", full.names = TRUE)
fm_template_files <- fm_template_files[grepl(this_allocation_method, fm_template_files)]

this_file <- fm_template_files[[3]]


map(.x = fm_template_files, .f = \(this_file){
  
  filename <- tools::file_path_sans_ext(this_file)
  
  this_source <- str_sub(filename, start = nchar(filename)-1, end = nchar(filename))
  
  saveName_P_md <- sprintf(here("data/spatial/marine_diet/int/fishmeal_P_km2_%s_%s.tif"), this_allocation_method, this_source)
  saveName_P_pd <- sprintf(here("data/spatial/plant_diet/int/fishmeal_P_km2_%s_%s.tif"), this_allocation_method, this_source)
  
  if(!file.exists(saveName_P_md) | !file.exists(saveName_P_pd)){
  
  this_r <- rast(this_file)*0
  names(this_r) <- "total_P_km2"
  
  #save P rasters
  terra::writeRaster(x = this_r, filename = saveName_P_md, overwrite = TRUE)
  
  terra::writeRaster(x = this_r, filename = saveName_P_pd, overwrite = TRUE)
  }
  
  })


```


Fish oil 

```{r}
fo_template_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fish oil_ghg", full.names = TRUE)
fo_template_files <- fo_template_files[grepl(this_allocation_method, fo_template_files)]


this_file <- fo_template_files[[3]]


map(.x = fo_template_files, .f = \(this_file){
  
  filename <- tools::file_path_sans_ext(this_file)
  
  this_source <- str_sub(filename, start = nchar(filename)-1, end = nchar(filename))
  
  saveName_P_md <- sprintf(here("data/spatial/marine_diet/int/fish oil_P_km2_%s_%s.tif"), this_allocation_method, this_source)
  saveName_P_pd <- sprintf(here("data/spatial/plant_diet/int/fish oil_P_km2_%s_%s.tif"), this_allocation_method, this_source)
  
  if(!file.exists(saveName_P_md) | !file.exists(saveName_P_pd)){
  
  this_r <- rast(this_file)*0
  names(this_r) <- "total_P_km2"
  
  #save P rasters
  terra::writeRaster(x = this_r, filename = saveName_P_md, overwrite = TRUE)
  
  terra::writeRaster(x = this_r, filename = saveName_P_pd, overwrite = TRUE)
  }
  
  })


```

Now sum N and P for fishmeal and oil nutrient rasters

Fishmeal first

```{r}
fm_N_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fishmeal_N", full.names = TRUE)
fm_N_files <- fm_N_files[grepl(this_allocation_method, fm_N_files)]
fm_P_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fishmeal_P", full.names = TRUE)
fm_P_files <- fm_P_files[grepl(this_allocation_method, fm_P_files)]

this_N_file <- fm_N_files[[1]]
this_P_file <- fm_P_files[[1]]

map2(.x = fm_N_files, .y = fm_P_files, .f=\(this_N_file, this_P_file){
  
  filename_N <- tools::file_path_sans_ext(this_N_file)
  filename_P <- tools::file_path_sans_ext(this_P_file)
  
  this_source <- c(str_sub(filename_N, start = nchar(filename_N)-1, end= nchar(filename_N)), str_sub(filename_P, start = nchar(filename_P)-1, end= nchar(filename_P))) |> unique()
  
  #because these are zero - save to both marine and plant
  saveName_nutrient_md <- sprintf(here("data/spatial/marine_diet/int/fishmeal_nutrient_km2_%s_%s.tif"), this_allocation_method, this_source)
  saveName_nutrient_pd <- sprintf(here("data/spatial/plant_diet/int/fishmeal_nutrient_km2_%s_%s.tif"), this_allocation_method, this_source)
  
  if(!file.exists(saveName_nutrient_md) | !file.exists(saveName_nutrient_pd)){
  
  this_r <- rast(this_N_file)+rast(this_P_file)
  names(this_r) <- "total_nutrient_km2"
  
  #save nutrient fishmeal rasters
  terra::writeRaster(x = this_r, filename = saveName_nutrient_md, overwrite = TRUE)
  
  terra::writeRaster(x = this_r, filename = saveName_nutrient_pd, overwrite = TRUE)
  }
  
} )

```


Now fish oil

```{r}
fo_N_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fish oil_N", full.names = TRUE)
fo_N_files <- fo_N_files[grepl(this_allocation_method, fo_N_files)]
fo_P_files <- list.files(path = here("data/spatial/marine_diet/int"), pattern = "fish oil_P", full.names = TRUE)
fo_P_files <- fo_P_files[grepl(this_allocation_method, fo_P_files)]

#test function
this_N_file <- fo_N_files[[1]]
this_P_file <- fo_P_files[[1]]

map2(.x = fo_N_files, .y = fo_P_files, .f=\(this_N_file, this_P_file){
  
  filename_N <- tools::file_path_sans_ext(this_N_file)
  filename_P <- tools::file_path_sans_ext(this_P_file)
  
  this_source <- c(str_sub(filename_N, start = nchar(filename_N)-1, end= nchar(filename_N)), str_sub(filename_P, start = nchar(filename_P)-1, end= nchar(filename_P))) |> unique()
  
  #because these are zero - save to both marine and plant
  saveName_nutrient_md <- sprintf(here("data/spatial/marine_diet/int/fish oil_nutrient_km2_%s_%s.tif"), this_allocation_method, this_source)
  saveName_nutrient_pd <- sprintf(here("data/spatial/plant_diet/int/fish oil_nutrient_km2_%s_%s.tif"), this_allocation_method, this_source)
  
  if(!file.exists(saveName_nutrient_md) | !file.exists(saveName_nutrient_pd)){
  
  this_r <- rast(this_N_file)+rast(this_P_file)
  names(this_r) <- "total_nutrient_km2"
  
  #save nutrient fishmeal rasters
  terra::writeRaster(x = this_r, filename = saveName_nutrient_md, overwrite = TRUE)
  
  terra::writeRaster(x = this_r, filename = saveName_nutrient_pd, overwrite = TRUE)
  }
  
} )

```

```{r}
#clear the environment
rm(list = ls(all.names = TRUE))
```
