==> scripts/16_analysis_visualisation_ge.Rmd <==
---
title: "16_analysis_visualisation"
author: "Rich Cottrell"
date: "01/08/2022"
output: html_document
---
Set up
```{r}

library(tidyverse)
library(ggpubr)
library(patchwork)
library(here)
library(sf)
library(terra)
library(rnaturalearth)
#library(rnaturalearthdata)
library(RColorBrewer)
library(gameofthrones)
library(parallel)
library(ggdist)
library(data.table)
library(dtplyr)


moll_crs <- "ESRI:53009"
gp_crs <- "+proj=cea +lon_0=0 +x_0=0 +y_0=0 +lat_ts=45 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"

# taking cumulative impoact palette from Halpern et al 
red <- "#B90000"
d_brown <- "#515256"
m_brown <- "#B27B54"
l_brown <- "#BC995F"
green <- "#A8B072"
yellow<- "#EFCE71"
l_yellow <- "#F7F6C1"
light_gray <- "#F8F9FA"


discrete_pal <- c(l_yellow, yellow, green, l_brown, m_brown, d_brown, "firebrick")
continuous_pal <-  colorRampPalette(discrete_pal, space="Lab", bias = 3.5)(10000)

final_palette <- c(light_gray, continuous_pal, red)

#source files
source(here("src/fxns.R"))

#spatial products
countries <-    ne_countries(scale = "medium", returnclass = "sf") |> st_transform(crs = moll_crs)
bbox <- ne_download(scale = 50, type = "wgs84_bounding_box", category = "physical", returnclass = "sf") |> st_transform(crs = moll_crs)
coastline <- ne_coastline(scale = 50, returnclass = "sf") |> st_transform(crs = moll_crs)

select <- dplyr::select
values <- terra::values
```

Explanatory plot of the two diets

```{r}

marine_diet <- readRDS(here("data/tidy_data/diet-scenarios/marine_diet.rds")) |> 
  filter(prop >0) |> 
  add_row(groups = "Microingredients", ingredients = "microingredients", diet = "marine_diet", prop = 0.02) |> 
  mutate(ingredients = case_when(ingredients == "canola/camelina oil" ~ "Canola oil",
                                  TRUE ~ str_to_sentence(ingredients)),
         ingredients = gsub(pattern = "-wild", replacement = "", ingredients)) |> 
  mutate(ingredients = factor(ingredients, levels = c("Fishmeal", "Fish oil", "Guar meal", "Soybean meal", "Corn gluten meal", "Wheat gluten", "Wheat", "Microingredients"))) |> 
  mutate(cum_prop = cumsum(prop)) |> 
  mutate(label_y_pos = if_else(condition = prop != prop[1], true = prop/2+lag(cum_prop), false = cum_prop/2)) |> 
  mutate(label_y_pos = 1-label_y_pos) 

plant_diet <- readRDS(here("data/tidy_data/diet-scenarios/plant_diet.rds")) |> 
  filter(prop >0) |> 
  add_row(groups = "Microingredients", ingredients = "microingredients", diet = "plant_diet", prop = 0.04) |> 
  mutate(ingredients = case_when(ingredients == "canola/camelina oil" ~ "Canola oil",
                                  TRUE ~ str_to_sentence(ingredients)),
         ingredients = gsub(pattern = "-wild", replacement = "", ingredients)) |> 
  mutate(ingredients = factor(ingredients, levels = c("Fishmeal", "Fish oil", "Soy protein concentrate", "Corn gluten meal","Faba beans", "Pea protein concentrate",  "Sunflower meal", "Wheat gluten", "Canola oil", "Linseed oil", "Wheat", "Pea starch", "Corn starch", "Microingredients"))) |> 
  mutate(cum_prop = cumsum(prop)) |> 
  mutate(label_y_pos = if_else(condition = prop != prop[1], true = prop/2+lag(cum_prop), false = cum_prop/2))|> 
   mutate(label_y_pos = 1-label_y_pos) 


#check totals 
marine_diet |> pull(prop) |> sum() #adds to 1
plant_diet |> pull(prop) |> sum() #adds to 1

ingredient_levels <- c( "Fishmeal",  "Fish oil" ,"Guar meal" , "Soybean meal","Corn gluten meal" , "Wheat gluten" ,"Wheat" , "Soy protein concentrate", "Faba beans" ,   "Pea protein concentrate","Sunflower meal","Canola oil","Linseed oil","Pea starch","Corn starch", "Microingredients")


#get a plot and legend for each diet


marine_diet_plot <- ggplot(data = marine_diet,
       aes(x=diet, y = prop, fill = ingredients))+
  geom_col()+
  scale_fill_manual(values =  rev(colorRampPalette(colors = rcartocolor::carto_pal(n = 9, name = "Earth"))(16))[c(1:7,16)])

marine_legend <- get_legend(marine_diet_plot)


plant_diet_plot <- ggplot(data = plant_diet,
       aes(x=diet, y = prop, fill = ingredients))+
  geom_col()+
  scale_fill_manual(values =  rev(colorRampPalette(colors = rcartocolor::carto_pal(n = 9, name = "Earth"))(16))[c(1,2,5:16)])

plant_legend <- get_legend(plant_diet_plot)



#combine both diets

both_diets <- bind_rows(marine_diet, plant_diet) |> 
  
  mutate(label_x_pos = if_else(diet =="marine_diet", true =0.5, false = 2.5)) |> 
  mutate(diet = if_else(condition = diet == "marine_diet", true = "Marine-\ndominant", false = "Plant-\ndominant")) |> 
  mutate(hjust = if_else(condition = diet == "marine_diet", true = 1, false = 0)) |> 
  mutate(ingredients = factor(ingredients, levels = rev(ingredient_levels))) |> 
  mutate(diet = factor(diet, levels = rev(c("Plant-\ndominant", "Marine-\ndominant"))))




both_diets_p <- ggplot(data = both_diets,
       aes(x = diet , y = prop, fill = ingredients))+
  geom_col()+
  scale_fill_manual(values =  (colorRampPalette(colors = rcartocolor::carto_pal(n = 7, name = "Earth"))(16)))+
  theme_pubr()+
  theme(legend.title = element_blank(),
        legend.text = element_text(size=7),
        text = element_text(size=8),
        legend.position = "right",
        #legend.box.spacing = unit(-0.1, "cm"),
        panel.grid = element_blank(),
        legend.key.size = unit(0.25, "cm")
        )+
  labs(x = "Feed scenario", y = "Proportion")+
  guides(fill = guide_legend(byrow=TRUE, reverse = FALSE, ncol = 1))


ggsave(filename = here("figures/feed_composition.jpg"), dpi = 600, width = 10, height = 9, units = "cm")



```

Some statistics on the number of species/crop raw materials/countries/marine regions needed to support demand. 

```{r}

forage_fish <- data.table::fread(here("data/tidy_data/demand/embodied_fish_per_cell.csv"))
forage_fish_md <- forage_fish |> filter(diet == "marine_diet") |> as_tibble()
unique(forage_fish$CommonName)
sum(forage_fish_md$total_catch)

#Number of countries for crops
countries_regions <-
  c(list.files("/mnt/rdsi/github/feed_pressure_mapping/data/spatial/marine_diet/pressures", full.names = TRUE), list.files("/mnt/rdsi/github/feed_pressure_mapping/data/spatial/plant_diet/pressures", full.names = TRUE)) |>
  map(.f = \(this_element){
    sans_ext <- tools::file_path_sans_ext(this_element)
    
    last_3 <-  str_sub(string = sans_ext, start = nchar(sans_ext)-2, end = nchar(sans_ext))
    return(last_3)
  }) |> 
  unlist() |> 
  unique() 


length(countries_regions[!countries_regions %in% c("EPC", "NAT", "OTH")]) # 18 countries used and 3 marine regions


#number of crop types
total_crop_demand <- readRDS("/mnt/rdsi/github/feed_pressure_mapping/data/tidy_data/demand/total_crop_demand.rds") 

total_crop_demand |> pull(FAOSTAT_name) |> unique()
total_crop_demand |> filter(FAOSTAT_name == "Pulses nes")




#forage fish embodied demand by regions
embodied_fish_bysource <- readRDS(here("data/tidy_data/demand/embodied_fish_demand_by_source_diet.rds")) |> 
  mutate(source_code = case_when(source_code == "EPC" ~ "Eastern Pacific",
                                 source_code == "NAT" ~ "North Atlantic",
                                 source_code == "OTH" ~ "All other regions")) |> 
  mutate(source_code = factor(source_code, levels = c("Eastern Pacific", "North Atlantic", "All other regions"))) |> 
  mutate(diet = if_else(diet=="marine_diet", true = "Marine-dominant\nfeed", false = "Plant-dominant\nfeed"))


#how much more embodied fish does a marine diet create relative to plant?
source_codes <- unique(embodied_fish_bysource$source_code) |> as.character()

diff_list_ge <- list()
diff_list_mass <- list()
amount_list <- list()

for(s in source_codes){
 
  marine_plant_ratio_ge <-  (embodied_fish_bysource |> filter(source_code == s & allocation == "Energetic allocation" & diet == "Marine-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))/(embodied_fish_bysource |> filter(source_code == s & allocation == "Energetic allocation" & diet == "Plant-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))
  
  marine_plant_ratio_mass <-  (embodied_fish_bysource |> filter(source_code == s & allocation == "Mass allocation" & diet == "Marine-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))/(embodied_fish_bysource |> filter(source_code == s & allocation == "Mass allocation" & diet == "Plant-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))
  
  
 diff_list_ge[which(source_codes==s)] <- marine_plant_ratio_ge
 
 diff_list_mass[which(source_codes==s)] <- marine_plant_ratio_mass
 
}
unlist(diff_list_ge) #~2.2 across all regions and allocation methods=
unlist(diff_list_mass)

#Differences among regions within marine-dominant feed
embodied_fish_bysource |> 
  filter(diet == "Marine-dominant\nfeed" & demand_from == "Total") |> 
  arrange(-embodied_fish) |> 
  mutate(region_diff = max(embodied_fish)-embodied_fish)


embodied_fish_bysource |>
  filter(demand_from =="Total") |> 
  group_by(diet, allocation) |> 
  summarise(mean = mean(embodied_fish, na.rm=TRUE),
            sd = sd(embodied_fish, na.rm=TRUE))


ggplot(data = embodied_fish_bysource) +
  aes(x = demand_from, y = embodied_fish/1000, fill = source_code)+
  geom_bar(stat= "identity", position = "dodge")+
  facet_grid(rows = vars(diet), cols = vars(allocation), scales = "fixed")+
  
  labs(fill = "Origin", y = "Embodied fish demand (1000s T)")+
  scale_fill_manual(values = c("#edf8b1", "#7fcdbb", "#2c7fb8"))+
  theme_bw()+
  theme(legend.position = c(0.75, 0.43),
        legend.background = element_rect(fill="transparent"), 
        legend.direction = "horizontal",
        axis.title.x = element_blank(),
        panel.grid = element_blank(), 
        text = element_text(size=12),
        legend.text = element_text(size = 10))+
  guides(fill = guide_legend(title.position = "top"))


ggsave(filename = here("figures/presos/embodied_forage_fish_demand.jpg"), device="jpg", dpi = 600, width = 18, height = 15, units="cm")
ggsave(filename = here("figures/supplementary/embodied_forage_fish_demand.jpg"), device="jpg", dpi = 600, width = 18, height = 15, units="cm")


#find reasons for differences in demand

#blue whiting looks like the main reason with European sprat also. Comparing the embodied fish versus the catch both Sprat and Whiting rank as higher in embodied fish than for catch

(embodied_fish_rank <- 
  fread("/mnt/rdsi/github/feed_pressure_mapping/data/tidy_data/demand/embodied_fish_per_cell.csv") |> 
  lazy_dt() |> 
  group_by(diet, source_code, TaxonName, CommonName) |> 
  summarise(total_demand_ge = sum(total_embodied_fish_ge, na.rm = TRUE),
            total_demand_mass = sum(total_embodied_fish_mass, na.rm = TRUE)) |> 
  filter(diet == "marine_diet") |> 
  arrange(-total_demand_mass) |> 
  as_tibble())


(fish_catch_rank <- 
  fread("/mnt/rdsi/github/feed_pressure_mapping/data/tidy_data/demand/embodied_fish_per_cell.csv") |> 
  lazy_dt() |> 
  group_by(diet, source_code, TaxonName, CommonName) |> 
  summarise(total_catch = sum(total_catch, na.rm = TRUE)) |> 
  filter(diet == "marine_diet") |> 
  arrange(-total_catch) |> 
  as_tibble())

```

Bring in the source-ingredient combinations within each diet.  

```{r}

marine_diets <- readRDS(here("data/tidy_data/diets/ingredient_source_combinations_marine_diet.rds")) 

plant_diets <- readRDS(here("data/tidy_data/diets/ingredient_source_combinations_plant_diet.rds")) 

```

# Compare marine-dominant and plant-dominant feeds from a selectiom of different sources to illustrate the concept.


Combine all values for pressures for a marine diet

```{r}

marine_source_ingredient <- 
  marine_diets |> 
  select(ingredient) |> 
  mutate(source_1 = c(1, 2, 2, 1, 2, 5, 1),
         source_2 = c(2, 1, 1, 1, 4, 2, 3),
         source_3 = c(3, 3, 3, 1, 3, 2, 4)) |> 
   mutate(row = row_number()) |> 
  group_split(row) 
  
  # select(ingredient, source_1 = `5`, source_2 = `9`, source_3 = `21`) |> 
  # 
  # mutate(source_1 = case_when(ingredient == "guar meal"~ as.integer(1),
  #                           TRUE ~ source_1),
  #        source_2 = case_when(ingredient == "guar meal"~ as.integer(1),
  #                           TRUE ~ source_2),
  #        source_3 = case_when(ingredient == "guar meal"~ as.integer(1),
  #                           TRUE ~ source_3)) |> 
  # 




#test dataframe
this_df <- marine_source_ingredient[[4]]


#Now prepare the files corresponding to the first source for each ingredient for each pressure


sources <- c("source_1", "source_2", "source_3")

source = "source_1"

#gross energy allocation 

marine_source_file_list <- list()


for(source in sources){
    
    marine_diet_files <- 
      map(.x = marine_source_ingredient, .f = \(this_df){
        
        this_allocation_method <- "ge_allocation"
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, " (", this_allocation_method, ") ", source)
        
        these_ingredient_files <- list.files(here("data/spatial/marine_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        these_ingredient_files <- these_ingredient_files[grepl(this_allocation_method, these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    marine_source_file_list[[source]] <- marine_diet_files
    
  }
  


#load all files as rasters and adjust files a  NA -> 0

#this_file <- marine_diet_files[[6]]


marine_source_rast_list <- list()

for(source in sources){
  
  marine_rast_list <- 

  map(.x = marine_source_file_list[[source]], .f = \(this_file){
  
  
  this_rast <- rast(this_file)
  this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
  
  this_rast[is.na(this_rast)] <- 0
  names(this_rast) <- this_source
  return(this_rast)
})
  
  marine_source_rast_list[[source]] <- marine_rast_list
  
}



#stack and sum the pressures for all ingredients and source combinations

marine_diet_raster <- list()

options(warn = 0)
for(source in sources){
  
  #stack rasters for each sourcing combination
  marine_diet_raster[[source]] <- app(rast(marine_source_rast_list[[source]]), sum)
  marine_diet_raster[[source]][marine_diet_raster[[source]]==0] <- NA
  
  #reproject raster 
  marine_diet_raster[[source]] <- project(marine_diet_raster[[source]], moll_crs)
  
}
options(warn = 1)



# MASS ALLOCATION




```


Combine all values for a plant-dominant diet

```{r}
plant_source_ingredient <- 
  plant_diets |> 
  select(ingredient) |> 
  mutate(source_1 = c(1, 1, 2, 4, 1, 1, 3, 1, 3, 4, 2, 2, 1),
         source_2 = c(2, 2, 3, 3, 2, 2, 3, 3, 5, 1, 3, 5, 2),
         source_3 = c(3, 3, 1, 5, 3, 3, 4, 2, 1, 2, 1, 3, 5)) |> 
   mutate(row = row_number()) |> 
  group_split(row) 


# 
#   select(ingredient, source_1 = `3`, source_2 = `4`, source_3 = `18`) |> 
#   mutate(source_1 = case_when(ingredient == "linseed oil"~ as.integer(1),
#                             TRUE ~ source_1),
#          source_2 = case_when(ingredient == "linseed oil"~ as.integer(1),
#                             TRUE ~ source_2),
#          source_3 = case_when(ingredient == "linseed oil"~ as.integer(1),
#                             TRUE ~ source_3)) |> 
#   mutate(row = row_number()) |> 
#   group_split(row) 



#test dataframe
this_df <- plant_source_ingredient[[6]]

sources <- c("source_1", "source_2", "source_3")

plant_source_file_list <- list()


#Now prepare the files corresponding to the first source for each ingredient for each pressure

for(source in sources){
    
    plant_diet_files <- 
      map(.x = plant_source_ingredient, .f = \(this_df){
        
        this_allocation_method <- "ge_allocation"
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient)
        
        
        these_ingredient_files <- list.files(here("data/spatial/plant_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
        these_ingredient_files <- these_ingredient_files[grepl(this_allocation_method, these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    plant_source_file_list[[source]] <- plant_diet_files
    
  }
  


#adjust layers and replace NA with 0

plant_source_rast_list <- list()

for(source in sources){
  
  plant_rast_list <- 

  map(.x = plant_source_file_list[[source]], .f = \(this_file){
  
  
  this_rast <- rast(this_file)
  this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
  
  this_rast[is.na(this_rast)] <- 0
  names(this_rast) <- this_source
  return(this_rast)
})
  
  plant_source_rast_list[[source]] <- plant_rast_list
  
}



#stack and sum the pressures for all ingredients and source combinations

plant_diet_raster <- list()

oldw <- getOption("warn")
options(warn = -1) #just turns off the warnings of southern latitudes issues from converting from Gall-Peters to Mollweide
for(source in sources){
  
  #stack rasters for each sourcing combination
  plant_diet_raster[[source]] <- app(rast(plant_source_rast_list[[source]]), sum)
  plant_diet_raster[[source]][plant_diet_raster[[source]]==0] <- NA
  
  #reproject raster 
  plant_diet_raster[[source]] <- project(plant_diet_raster[[source]], moll_crs)
  
}
options(warn = oldw)

marine_diet_raster
plant_diet_raster




```

Figure - maps comparing diet and sourcing

```{r}

max_marine_value <- marine_diet_raster |> map(no_na_values) |> unlist() |> as.numeric() |> max()
max_plant_value <- plant_diet_raster|> map(no_na_values) |> unlist() |> as.numeric() |> max()

max_cum_value <- max(max_marine_value, max_plant_value)

(sum_marine_value <- marine_diet_raster |> map(no_na_values) |> map(sum))
(sum_plant_value <- plant_diet_raster|> map(no_na_values) |> map(sum))


# MARINE DIETS

#source 1
marine_diet1_df <- terra::as.data.frame(marine_diet_raster[["source_1"]], xy=TRUE)


(marine_sourcing_p1 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries, fill= "grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = marine_diet1_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(
                      colours = final_palette, 
                       limits = c(0, max_cum_value)
                       )+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = "Raw material origin\ncombination 1", angle = 90)
  )
  


#source 2
marine_diet2_df <- terra::as.data.frame(marine_diet_raster[["source_2"]], xy=TRUE)

(marine_sourcing_p2 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries, fill="grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = marine_diet2_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = final_palette, 
                       limits = c(0, max_cum_value))+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = "Raw material origin\ncombination 2", angle = 90)
  )

#source 3

marine_diet3_df <- terra::as.data.frame(marine_diet_raster[["source_3"]], xy=TRUE)

(marine_sourcing_p3 <- 
  ggplot()+
  geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries, fill="grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = marine_diet3_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = final_palette, 
                       limits = c(0, max_cum_value))+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = "Raw material origin\ncombination 3", angle = 90)
  )



#PLANT DIETS

#source 1
plant_diet1_df <- terra::as.data.frame(plant_diet_raster[["source_1"]], xy=TRUE)


(plant_sourcing_p1 <- 
  ggplot()+
  geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries, fill="grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = plant_diet1_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth = 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = final_palette, 
                       limits = c(0, max_cum_value))+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))
  )
  
#source 2
plant_diet2_df <- terra::as.data.frame(plant_diet_raster[["source_2"]], xy=TRUE)

(plant_sourcing_p2 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries, fill="grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = plant_diet2_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth = 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = final_palette, 
                       limits = c(0, max_cum_value))+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                barwidth = 10, barheight = 0.5, units = "cm"))
  )

#source 3

plant_diet3_df <- terra::as.data.frame(plant_diet_raster[["source_3"]], xy=TRUE)

(plant_sourcing_p3 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries, fill="grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = plant_diet3_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth = 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = final_palette, 
                       limits = c(0, max_cum_value))+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))

  )

# bar_legend <- get_legend(plant_sourcing_p3)
# 
# 
# source_diet_map <- 
#   ggarrange(
#     ggarrange(marine_sourcing_p1, plant_sourcing_p1,
#               marine_sourcing_p2, plant_sourcing_p2,
#               marine_sourcing_p3, plant_sourcing_p3,
#               nrow = 3, 
#               ncol = 2,
#               legend = "none"),
#     
#     bar_legend,
#     nrow = 2,
#     ncol=1,
#     heights = c(11,1))


source_diet_map <- 
  (marine_sourcing_p1|plant_sourcing_p3)/
  (marine_sourcing_p2|plant_sourcing_p2)/
  (marine_sourcing_p3|plant_sourcing_p1)+
  plot_layout(guides = "collect")+
  plot_annotation(tag_levels = "a", #surely there is a better way than below?
                  title = "                    Marine-dominant feed                                Plant-dominant feed") & 
  theme(plot.tag = element_text(size=9), 
        plot.tag.position = c(0.1, 1),
        legend.position = "bottom")
          

ggsave(plot = source_diet_map, filename = here("figures/source_vs_diet_visualisation_map.jpg"), device = "jpg", dpi = 600, width = 18, height = 15.5, units = "cm")


```


#Permutation analysis - Energetic allocation

For each of the 2000 combinations for marine-dominant and plant-dominant diets, identify the source of each ingredient and for each pressure. So perform a similar function for each above but summarise the rasters. We also need to repeat this for each allocation method.

Marine diets - Energetic Allocation

```{r}



marine_source_ingredient <- 
  marine_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "guar meal"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)



sources <- bind_rows(marine_source_ingredient) |> select(-c("ingredient", "row")) |>  names()   

marine_source_file_list <- list()


this_df <- marine_source_ingredient[[1]]
source <- sources[[1]]


#Now list the files for all ingredients and sources

for(source in sources){
    
    marine_diet_files <- 
      map(.x = marine_source_ingredient, .f = \(this_df){
        
        this_allocation_method <- "ge_allocation"
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/marine_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        these_ingredient_files <- these_ingredient_files[grepl("ge_allocation", these_ingredient_files)]
        
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    marine_source_file_list[[source]] <- marine_diet_files
    
  }
  

#bring in rasters to list

marine_source_rast_list <- list()

for(source in sources){
  
  message("processing....", source)
  
  marine_rast_list <- 

  map(.x = marine_source_file_list[[source]], .f = \(this_file){
  
  
  this_rast <- rast(this_file)
  this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
  
  # this_rast[is.na(this_rast)] <- 0
  # names(this_rast) <- this_source
  return(this_rast)
})
  
  marine_source_rast_list[[source]] <- marine_rast_list
  
}



#stack and sum the pressures for all ingredients and source combinations

marine_diet_raster <- list()

for(source in sources){
  
  message("processing stack for combination....", source)
  
  #stack rasters for each sourcing combination
    marine_diet_raster[[source]] <- global(app(rast(marine_source_rast_list[[source]]), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(source = source, .before = sum)
  
  
}

marine_diet_raster_df <- bind_rows(marine_diet_raster) |> mutate(diet = "marine diet")

saveRDS(marine_diet_raster_df, file = here("data/tidy_data/pressures/marine_cumulative_pressure_df.rds"))


```

Plant diets - Energetic allocation

```{r}

plant_source_ingredient <- 
  plant_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "linseed oil"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)



sources <- bind_rows(plant_source_ingredient) |> select(-c("ingredient", "row")) |>  names()   

plant_source_file_list <- list()

#test data for loop
# this_df <- plant_source_ingredient[[1]]
# source <- sources[[1]]


#Now list the files for all ingredients and sources

for(source in sources){
    
    plant_diet_files <- 
      map(.x = plant_source_ingredient, .f = \(this_df){
        
        this_allocation_method <- "ge_allocation"
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/plant_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        these_ingredient_files <- these_ingredient_files[grepl("ge_allocation", these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    plant_source_file_list[[source]] <- plant_diet_files
    
  }
  

#bring in rasters to list

plant_source_rast_list <- list()

  
for(source in sources){
  
  message("processing....", source)
  
  plant_rast_list <- 

  map(.x = plant_source_file_list[[source]], .f = \(this_file){
  
  
  this_rast <- rast(this_file)
  this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
  # 
  # this_rast[is.na(this_rast)] <- 0
  # names(this_rast) <- this_source
  return(this_rast)
})
  
  plant_source_rast_list[[source]] <- plant_rast_list
}


#stack and sum the pressures for all ingredients and source combinations

plant_diet_raster <- list()

for(source in sources){
  
  message("processing stack for combination....", source)
  
  #stack rasters for each sourcing combination
  plant_diet_raster[[source]] <- global(app(rast(plant_source_rast_list[[source]]), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(source = source, .before = sum)
  
}

plant_diet_raster_df <- bind_rows(plant_diet_raster) |> mutate(diet = "plant diet")

saveRDS(plant_diet_raster_df, file = here("data/tidy_data/pressures/plant_cumulative_pressure_df.rds"))

```

Energetic allocation - add marine and plant diet data together and save 

```{r}
marine_diet_raster_df <- readRDS(file = here("data/tidy_data/pressures/marine_cumulative_pressure_df.rds"))
plant_diet_raster_df <- readRDS(file = here("data/tidy_data/pressures/plant_cumulative_pressure_df.rds"))

marine_plant_df <- bind_rows(marine_diet_raster_df, plant_diet_raster_df)

saveRDS(marine_plant_df, file = here("data/tidy_data/pressures/marine_plant_pressures_combined_ge.rds"))



```


Energetic allocation - Visualise the differences between the two diets for cumulative pressures
```{r}

marine_plant_df <- readRDS(here("data/tidy_data/pressures/marine_plant_pressures_combined.rds"))


my_pal <- rcartocolor::carto_pal(n = 8, name = "Bold")[c(1, 3, 7, 2)]


(both_diets_agg_p <- ggplot(data = marine_plant_df |> mutate(diet = case_when(diet == "plant diet" ~ "Plant-\ndominant\nfeed",
                                                         diet == "marine diet" ~ "Marine-\ndominant\nfeed"),
                                        diet = factor(diet, levels = c("Plant-\ndominant\nfeed", "Marine-\ndominant\nfeed"))))+
  aes(x = sum, y = diet, colour = diet)+
    geom_jitter(shape =20, size = 1.5, alpha = 0.1, height = 0.4
  )+
  # ggdist::stat_gradientinterval(
  #   width = .3, color = "black", fill_type = "gradient"
  # )+
  scale_colour_manual(values = my_pal[c(4,2)], guide = "none")+
  scale_x_continuous(limits = c(1,4))+
  theme_pubr()+
  geom_boxplot(notch = TRUE, colour = "grey40", width = 0.3, alpha = 0, linewidth=0.5)+
  labs(y = "Diet", x = "Cumulative pressure index")+
  theme(text = element_text(size=8),
        axis.text = element_text(size=8),
        axis.title = element_text(size=8),
        axis.title.y = element_blank()))


ggsave(filename = here("figures/diet_permutation_comparisons_ge.jpg"), device = "jpg", dpi = 600, width = 8.9, height = 5, units="cm")

  
  
```

Energetic allocation - Some statistics for marine-plant comparisons
```{r}

marine_vals <-  marine_plant_df |> filter(diet == "marine diet") |> pull(sum)
plant_vals <- marine_plant_df |> filter(diet == "plant diet") |> pull(sum)


max_marine <- max(marine_vals)
min_marine <- min(marine_vals)

max_plant <- max(plant_vals)
min_plant <- min(plant_vals)

#How much variation in scores within a feed
max_marine/min_marine
max_plant/min_plant

#which marine vals are higher than the plant minimum?
marine_vals[marine_vals>min_plant] |> length()/ length(marine_vals) #98% of marine values

#which plant vals are lower than the marine maximum?
plant_vals[plant_vals<max_marine] |> length()/ length(plant_vals) #98.6%

#median comparisons
median(marine_vals)
median(plant_vals)

((median(marine_vals)/median(plant_vals))-1)*100 #2% higher median CPI for plants

#maximum comparisons
max(marine_vals)
max(plant_vals)

max(plant_vals)/max(marine_vals) #11% higher maximum CPI for plants
min(plant_vals)/min(marine_vals) #min plant diets have CPI 20% > marine diets

#minimum of one, maximum of anotehr
min(plant_vals)/max(marine_vals) #plant diets can have CPI 73% lower than marine diets
min(marine_vals)/max(plant_vals) #marine diets can have CPI 80% lower than plant diets

summary(marine_vals)
summary(plant_vals)

#manual calculation of CI


```



# Disaggregation of permutation by pressure - Energetic allocation

Once we have compared the two diets with different sourcing combinations, it's important to know how they disaggregate into the pressures that are the major drivers.
Pressure Disaggregation - Marine diets

```{r}

marine_source_ingredient <- 
  marine_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "guar meal"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)



sources <- bind_rows(marine_source_ingredient) |> select(-c("ingredient", "row")) |>  names()   

marine_source_file_list <- list()


this_df <- marine_source_ingredient[[1]]
source <- sources[[1]]


#Now list the files for all ingredients and sources

for(source in sources){
    
    marine_diet_files <- 
      map(.x = marine_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/marine_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
        these_ingredient_files <- these_ingredient_files[grepl("ge_allocation", these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    marine_source_file_list[[source]] <- marine_diet_files
    
  }
  


#stack and sum by individual pressures for each ingredient source combination

md_disaggregated_pressures_df_list <- list()


for(source in sources){
  
  message("processing.....", source)
  
  this_combination_all_pressures_list <- marine_source_file_list[[source]]
  
  
  
  #########Separate GHG files, read rasters into a list, stack all GHG layers
  
  message("stacking ghg layers for...", source)
  
  this_combination_ghg_list <- this_combination_all_pressures_list[grep("ghg", this_combination_all_pressures_list)]
  
  these_ghg_rasters <- map(.x = this_combination_ghg_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_ghg_df <- global(app(rast(these_ghg_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "GHG", source = source, .before = sum)
  
  
  
  ##########Separate water files, read rasters into a list, stack all water layers
  
  message("stacking water layers for...", source)
  
  this_combination_water_list <- this_combination_all_pressures_list[grep("water", this_combination_all_pressures_list)]
  
  these_water_rasters <- map(.x = this_combination_water_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_water_df <- global(app(rast(these_water_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Water", source = source, .before = sum)
  
  
  #########Separate N files, read rasters into a list, stack all N layers
  
  message("stacking N layers for...", source)
  
  this_combination_N_list <-  this_combination_all_pressures_list[grep("_N_", this_combination_all_pressures_list)]
  
  these_N_rasters <- map(.x = this_combination_N_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_N_df <- global(app(rast(these_N_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Nitrogen", source = source, .before = sum)
  
  
  #########Separate P files, read rasters into a list, stack all P layers
  
  message("stacking P layers for...", source)
  
  this_combination_P_list <-  this_combination_all_pressures_list[grep("_P_", this_combination_all_pressures_list)]
  
  these_P_rasters <- map(.x = this_combination_P_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_P_df <- global(app(rast(these_P_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Phosphorus", source = source, .before = sum)
  
  
  
  #########Separate disturbance files, read rasters into a list, stack all disturbance layers
  
  message("stacking disturbance layers for...", source)
  
  this_combination_disturbance_list <-  this_combination_all_pressures_list[grep("disturbance", this_combination_all_pressures_list)]
  
  these_disturbance_rasters <- map(.x = this_combination_disturbance_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_disturbance_df <- global(app(rast(these_disturbance_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Disturbance", source = source, .before = sum)
  
  
  md_disaggregated_pressures_df_list[[source]] <- bind_rows(this_ghg_df, this_water_df, this_N_df, this_P_df, this_disturbance_df)
  
}

marine_diet_pressures_df <- bind_rows(md_disaggregated_pressures_df_list) |> mutate(diet = "marine diet")

saveRDS(marine_diet_pressures_df, file = here("data/tidy_data/pressures/marine_disaggregated_pressures_df.rds"))


```

Pressure disaggregation - Plant diets 

```{r}


plant_source_ingredient <- 
  plant_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "linseed oil"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)


sources <- bind_rows(plant_source_ingredient) |> select(-c("ingredient", "row")) |>  names()   

plant_source_file_list <- list()

#test data for loop
# this_df <- plant_source_ingredient[[1]]
# source <- sources[[1]]


#Now list the files for all ingredients and sources

for(source in sources){
    
    plant_diet_files <- 
      map(.x = plant_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/plant_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
        these_ingredient_files <- these_ingredient_files[grepl("ge_allocation", these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    plant_source_file_list[[source]] <- plant_diet_files
    
  }
  
  


#stack and sum by individual pressures for each ingredient source combination

pd_disaggregated_pressures_df_list <- list()


for(source in sources){
  
  message("processing.....", source)
  
  this_combination_all_pressures_list <- plant_source_file_list[[source]]
  
  
  
  #########Separate GHG files, read rasters into a list, stack all GHG layers
  
  message("stacking ghg layers for...", source)
  
  this_combination_ghg_list <- this_combination_all_pressures_list[grep("ghg", this_combination_all_pressures_list)]
  
  these_ghg_rasters <- map(.x = this_combination_ghg_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_ghg_df <- global(app(rast(these_ghg_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "GHG", source = source, .before = sum)
  
  
  
  ##########Separate water files, read rasters into a list, stack all water layers
  
  message("stacking water layers for...", source)
  
  this_combination_water_list <- this_combination_all_pressures_list[grep("water", this_combination_all_pressures_list)]
  
  these_water_rasters <- map(.x = this_combination_water_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_water_df <- global(app(rast(these_water_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Water", source = source, .before = sum)
  
  
  #########Separate N files, read rasters into a list, stack all N layers
  
  message("stacking N layers for...", source)
  
  this_combination_N_list <-  this_combination_all_pressures_list[grep("_N_", this_combination_all_pressures_list)]
  
  these_N_rasters <- map(.x = this_combination_N_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_N_df <- global(app(rast(these_N_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Nitrogen", source = source, .before = sum)
  
  
  #########Separate P files, read rasters into a list, stack all P layers
  
  message("stacking P layers for...", source)
  
  this_combination_P_list <-  this_combination_all_pressures_list[grep("_P_", this_combination_all_pressures_list)]
  
  these_P_rasters <- map(.x = this_combination_P_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_P_df <- global(app(rast(these_P_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Phosphorus", source = source, .before = sum)
  
  
  
  #########Separate disturbance files, read rasters into a list, stack all disturbance layers
  
  message("stacking disturbance layers for...", source)
  
  this_combination_disturbance_list <-  this_combination_all_pressures_list[grep("disturbance", this_combination_all_pressures_list)]
  
  these_disturbance_rasters <- map(.x = this_combination_disturbance_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_disturbance_df <- global(app(rast(these_disturbance_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Disturbance", source = source, .before = sum)
  
  
  pd_disaggregated_pressures_df_list[[source]] <- bind_rows(this_ghg_df, this_water_df, this_N_df, this_P_df, this_disturbance_df)
  
}

plant_diet_pressures_df <- bind_rows(pd_disaggregated_pressures_df_list) |> mutate(diet = "plant diet")

saveRDS(plant_diet_pressures_df, file = here("data/tidy_data/pressures/plant_disaggregated_pressures_df.rds"))

```

Add the marine and plant pressure disaggregation together and save

```{r}

marine_pressure_disagg <- readRDS(here("data/tidy_data/pressures/marine_disaggregated_pressures_df.rds"))
plant_pressure_disagg <- readRDS(here("data/tidy_data/pressures/plant_disaggregated_pressures_df.rds"))

marine_plant_pressure_disagg_df <- bind_rows(marine_pressure_disagg, plant_pressure_disagg)

saveRDS(object = marine_plant_pressure_disagg_df, file = here("data/tidy_data/pressures/marine_plant_disaggreagted_pressures_combined_df.rds"))


marine_pressure_disagg
plant_pressure_disagg


```


Visualise the pressure disaggregation

```{r}

marine_plant_pressure_disagg_df <- 
  readRDS(here("data/tidy_data/pressures/marine_plant_disaggreagted_pressures_combined_df.rds")) |> 
  group_by(diet, source) |> 
  nest() |> 
  mutate(total = map(data, ~(sum(.$sum)))) |> 
  unnest(c(data, total)) |> 
  mutate(prop = sum/total)

my_pal <- rcartocolor::carto_pal(n = 8, name = "Bold")[c(1, 3, 7, 2)]

summarised_pressure_disagg_df <- 
  marine_plant_pressure_disagg_df |> 
  group_by(diet, pressure) |> 
  summarise(mean = mean(prop, na.rm=TRUE),
            sd = sd(prop, na.rm=TRUE),
            median = median(prop, na.rm = TRUE),
            lq = quantile(prop, 0.25),
            uq = quantile(prop, 0.75)) |> 
  mutate(diet = if_else(diet == "marine diet", true = "Marine-dominant feed", false = "Plant-dominant feed"))


both_diets_disagg_p <- 
  ggplot(data= summarised_pressure_disagg_df, aes(x = pressure, y = mean, fill = diet)
       )+
  geom_col( position = "dodge", width = 0.85, alpha = 0.8)+
  geom_point(position = position_dodge(width = 0.7), size=0.6)+
  geom_errorbar(aes(x = pressure, ymin = mean-sd, ymax = mean+sd), position = position_dodge(width=0.7), width = 0, linewidth = 0.5)+
  geom_point(aes(x=pressure, y = mean), colour = "red", shape = 5, position = position_dodge(width = 0.7), size=0.4)+
  coord_polar(clip="off")+
  theme_bw()+
  theme(panel.border = element_blank(),
        axis.title.x = element_blank(),
        #axis.line.y = element_line(colour = "black", size=0.5),
        axis.title.y = element_text(hjust = 0.83),
        legend.position = c(0.84, 0.01),
        legend.title = element_blank(),
        legend.text = element_text(size=6),
        legend.background = element_rect(fill="transparent"),
        legend.key.size = unit(0.4, "cm"),
        plot.margin = unit(c(0, 0.3, 0, 0), "cm"),
        text = element_text(size=8))+
  scale_fill_manual(values = my_pal[c(2, 4)])+
  #scale_y_continuous(breaks = c(0,50,100,150, 200,250,300,350,400,450,500))+
  labs(y = "Prop. of Cumulative Pressure Index")+
  guides(fill = guide_legend(override.aes = list(shape = NA)))

#boxplot version
# ggplot(data= marine_plant_pressure_disagg_df, aes(x = pressure, y = sum, colour = diet, group = interaction(diet, pressure))
#        )+
#   geom_jitter(position = position_dodge2(width=0.8), alpha =0.08)+
#   geom_boxplot(notch = TRUE, fill = "transparent", colour = "grey40")+
#   
#   coord_polar(clip="off")+
#   theme_bw()+
#   theme(panel.border = element_blank(),
#         axis.title.x = element_blank(),
#         legend.title = element_blank(),
#         legend.position = c(0.84, 0.1),
#         legend.background = element_rect(fill="transparent"),
#         legend.key.size = unit(0.4, "cm"),
#         text = element_text(size=8))+
#   scale_colour_manual(values = my_pal[c(2, 4)])+
#   scale_y_continuous(breaks = c(0,50,100,150, 200,250))+
#   labs(y = "Normalised pressure index")+
#   guides(fill = guide_legend(override.aes = list(shape = NA)))

ggsave(filename = here("figures/disaggregation_of_pressures.jpg"), width = 8.9, height=8, units = "cm", device = "jpg", dpi = 600)

```
Combine aggregated and disaggregated figures
```{r}

both_diets_agg_p+both_diets_disagg_p+
  plot_annotation(tag_levels = "a")+
  theme(plot.tag = element_text(size=8))

ggsave(here("figures/both_diet_agg_disagg_combine.jpg"), dpi = 600, width = 18.9, height = 8, units="cm")

```

Some calculation regarding the decomposition of CPI by individual pressures
```{r}

#Disturbance (marine)
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "marine diet") |> pull(prop) |> max()
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "marine diet") |> pull(prop) |> min()
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "marine diet") |> pull(prop) |> mean()
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "marine diet") |> pull(prop) |> sd()

#Disturbance (plant)
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "plant diet") |> pull(prop) |> max()
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "plant diet") |> pull(prop) |> min()
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "plant diet") |> pull(prop) |> mean()
marine_plant_pressure_disagg_df |> filter(pressure == "Disturbance" & diet == "plant diet") |> pull(prop) |> sd()



```


#Disaggregation by pressure and ingredient


Disaggregate by ingredients and pressures -  Marine diets

```{r}

marine_source_ingredient <- 
  marine_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "guar meal"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)



sources <- bind_rows(marine_source_ingredient) |> select(-c("ingredient", "row")) |>  names()   

marine_source_file_list <- list()

# Test data for loop
# this_df <- marine_source_ingredient[[1]]
# source <- sources[[1]]


#Now list the files for all ingredients and sources

for(source in sources){
    
    marine_diet_files <- 
      map(.x = marine_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/marine_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
         these_ingredient_files <- these_ingredient_files[grepl("ge_allocation", these_ingredient_files)]
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    marine_source_file_list[[source]] <- marine_diet_files
    
  }
  


#stack and sum by individual pressures for each ingredient source combination

marine_ingredients <- readRDS(here("data/tidy_data/demand/total_ingredient_demand.rds")) |> 
  filter(diet=="marine_diet" & total_ingredient_demand>0) |> 
  mutate(ingredients = gsub(pattern = "-wild", replacement = "", x = ingredients)) |> pull(ingredients)

this_ingredient_pressure_df_list <- list()
all_ingredients_pressures_df_list <- list()


for(source in sources){
  
  message("processing.....", source)
  
  this_combination_all_pressures_list <- marine_source_file_list[[source]]

  
  for(ingredient in marine_ingredients){
    
    this_ingredient_pressure_list <- this_combination_all_pressures_list[grepl(ingredient, this_combination_all_pressures_list)]
    
    #########Separate GHG files, read rasters into a list, stack all GHG layers
  
  message("stacking ghg layers for...", source, " ", ingredient)
  
  this_combination_ghg_list <- this_ingredient_pressure_list[grep("ghg", this_ingredient_pressure_list)]
  
  these_ghg_rasters <- map(.x = this_combination_ghg_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_ghg_df <- global(app(rast(these_ghg_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "GHG", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  
  ##########Separate water files, read rasters into a list, stack all water layers
  
  message("stacking water layers for...", source, " ", ingredient)
  
  this_combination_water_list <- this_ingredient_pressure_list[grep("water", this_ingredient_pressure_list)]
  
  these_water_rasters <- map(.x = this_combination_water_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_water_df <- global(app(rast(these_water_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Water", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  #########Separate N files, read rasters into a list, stack all N layers
  
  message("stacking N layers for...", source, " ", ingredient)
  
  this_combination_N_list <-  this_ingredient_pressure_list[grep("_N_", this_ingredient_pressure_list)]
  
  these_N_rasters <- map(.x = this_combination_N_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_N_df <- global(app(rast(these_N_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Nitrogen", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  #########Separate P files, read rasters into a list, stack all P layers
  
  message("stacking P layers for...", source, " ", ingredient)
  
  this_combination_P_list <-  this_ingredient_pressure_list[grep("_P_", this_ingredient_pressure_list)]
  
  these_P_rasters <- map(.x = this_combination_P_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_P_df <- global(app(rast(these_P_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Phosphorus", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  #########Separate disturbance files, read rasters into a list, stack all disturbance layers
  
  message("stacking disturbance layers for...", source, " ", ingredient)
  
  this_combination_disturbance_list <-  this_ingredient_pressure_list[grep("disturbance", this_ingredient_pressure_list)]
  
  these_disturbance_rasters <- map(.x = this_combination_disturbance_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_disturbance_df <- global(app(rast(these_disturbance_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Disturbance", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  this_ingredient_pressure_df_list[[ingredient]] <- bind_rows(this_ghg_df, this_water_df, this_N_df, this_P_df, this_disturbance_df)
  
    
    
  }
  
  
  all_ingredients_pressures_df_list[[source]] <- bind_rows(this_ingredient_pressure_df_list)
  
}

marine_diet_pressures_by_ingredient_df <- bind_rows(all_ingredients_pressures_df_list) |> mutate(diet = "marine diet")

saveRDS(marine_diet_pressures_by_ingredient_df, file = here("data/tidy_data/pressures/marine_disaggregated_pressures_by_ingredient_df.rds"))

```

Disaggreagte by ingredients and pressures -  Plant diets

```{r}
plant_source_ingredient <- 
  plant_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "linseed oil"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)


sources <- bind_rows(plant_source_ingredient) |> select(-c("ingredient", "row")) |>  names()   

plant_source_file_list <- list()

#test data for loop
# this_df <- plant_source_ingredient[[1]]
# source <- sources[[1]]


#Now list the files for all ingredients and sources

for(source in sources){
    
    plant_diet_files <- 
      map(.x = plant_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/plant_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
        these_ingredient_files <- these_ingredient_files[grepl("ge_allocation", these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_file <- these_ingredient_files[grepl("ghg", these_ingredient_files)][[this_source_ingredient]]
        
        #nitrogen
        this_N_file <- these_ingredient_files[grepl("_N_", these_ingredient_files)][[this_source_ingredient]]
        
        #phosphorus
        this_P_file <- these_ingredient_files[grepl("_P_", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_file <- these_ingredient_files[grepl("water", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_file, this_ghg_file, this_N_file, this_P_file, this_h2o_file)
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    plant_source_file_list[[source]] <- plant_diet_files
    
  }
  



plant_ingredients <- readRDS(here("data/tidy_data/demand/total_ingredient_demand.rds")) |> 
  filter(diet=="plant_diet" & total_ingredient_demand>0) |> 
  mutate(ingredients = gsub(pattern = "-wild", replacement = "", x = ingredients)) |> pull(ingredients)


this_ingredient_pressure_df_list <- list()
all_ingredients_pressures_df_list <- list()


for(source in sources){
  
  message("processing.....", source)
  
  this_combination_all_pressures_list <- plant_source_file_list[[source]]

  
  for(ingredient in plant_ingredients){
    
    this_ingredient_pressure_list <- this_combination_all_pressures_list[grepl(ingredient, this_combination_all_pressures_list)]
    
    #########Separate GHG files, read rasters into a list, stack all GHG layers
  
  message("stacking ghg layers for...", source, " ", ingredient)
  
  this_combination_ghg_list <- this_ingredient_pressure_list[grep("ghg", this_ingredient_pressure_list)]
  
  these_ghg_rasters <- map(.x = this_combination_ghg_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_ghg_df <- global(app(rast(these_ghg_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "GHG", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  
  ##########Separate water files, read rasters into a list, stack all water layers
  
  message("stacking water layers for...", source, " ", ingredient)
  
  this_combination_water_list <- this_ingredient_pressure_list[grep("water", this_ingredient_pressure_list)]
  
  these_water_rasters <- map(.x = this_combination_water_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_water_df <- global(app(rast(these_water_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Water", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  #########Separate N files, read rasters into a list, stack all N layers
  
  message("stacking N layers for...", source, " ", ingredient)
  
  this_combination_N_list <-  this_ingredient_pressure_list[grep("_N_", this_ingredient_pressure_list)]
  
  these_N_rasters <- map(.x = this_combination_N_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_N_df <- global(app(rast(these_N_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Nitrogen", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  #########Separate P files, read rasters into a list, stack all P layers
  
  message("stacking P layers for...", source, " ", ingredient)
  
  this_combination_P_list <-  this_ingredient_pressure_list[grep("_P_", this_ingredient_pressure_list)]
  
  these_P_rasters <- map(.x = this_combination_P_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_P_df <- global(app(rast(these_P_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Phosphorus", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  #########Separate disturbance files, read rasters into a list, stack all disturbance layers
  
  message("stacking disturbance layers for...", source, " ", ingredient)
  
  this_combination_disturbance_list <-  this_ingredient_pressure_list[grep("disturbance", this_ingredient_pressure_list)]
  
  these_disturbance_rasters <- map(.x = this_combination_disturbance_list, .f = \(this_file){
    
    this_rast <- rast(this_file)
    this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
    
    return(this_rast)
  })
  
  this_disturbance_df <- global(app(rast(these_disturbance_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Disturbance", source = source, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
  
  
  this_ingredient_pressure_df_list[[ingredient]] <- bind_rows(this_ghg_df, this_water_df, this_N_df, this_P_df, this_disturbance_df)
  
    
    
  }
  
  
  all_ingredients_pressures_df_list[[source]] <- bind_rows(this_ingredient_pressure_df_list)
  
}

plant_diet_pressures_by_ingredient_df <- bind_rows(all_ingredients_pressures_df_list) |> mutate(diet = "plant diet")

saveRDS(plant_diet_pressures_by_ingredient_df, file = here("data/tidy_data/pressures/plant_disaggregated_pressures_by_ingredient_df.rds"))

```

Add marine and plant-dominant diet disaggregation by ingredient and pressure

```{r}

marine_diet_pressures_by_ingredient_df <- readRDS(here("data/tidy_data/pressures/marine_disaggregated_pressures_by_ingredient_df.rds"))

plant_diet_pressures_by_ingredient_df <- readRDS(here("data/tidy_data/pressures/plant_disaggregated_pressures_by_ingredient_df.rds"))


marine_plant_pressure_ingredient_disagg_combined_df <- bind_rows(marine_diet_pressures_by_ingredient_df, plant_diet_pressures_by_ingredient_df)

saveRDS(object = marine_plant_pressure_ingredient_disagg_combined_df, file = here("data/tidy_data/pressures/marine_plant_disaggregated_pressures_by_ingredient_combined_df.rds"))

```


Visualise the disaggregation by pressure and ingredient

```{r}

diets <- bind_rows(readRDS(here("data/tidy_data/diet-scenarios/marine_diet.rds")) |> 
  filter(prop >0) |> 
  add_row(groups = "Microingredients", ingredients = "microingredients", diet = "marine_diet", prop = 0.02) |> 
  mutate(ingredients = case_when(ingredients == "canola/camelina oil" ~ "Canola oil",
                                  TRUE ~ str_to_sentence(ingredients)),
         ingredients = gsub(pattern = "-wild", replacement = "", ingredients)) |> 
  mutate(ingredients = factor(ingredients, levels = c("Fishmeal", "Fish oil", "Guar meal", "Soybean meal", "Corn gluten meal", "Wheat gluten", "Wheat", "Microingredients"))),

readRDS(here("data/tidy_data/diet-scenarios/plant_diet.rds")) |> 
  filter(prop >0) |> 
  add_row(groups = "Microingredients", ingredients = "microingredients", diet = "plant_diet", prop = 0.04) |> 
  mutate(ingredients = case_when(ingredients == "canola/camelina oil" ~ "Canola oil",
                                  TRUE ~ str_to_sentence(ingredients)),
         ingredients = gsub(pattern = "-wild", replacement = "", ingredients)) |> 
  mutate(ingredients = factor(ingredients, levels = c("Fishmeal", "Fish oil", "Soy protein concentrate", "Corn gluten meal","Faba beans", "Pea protein concentrate",  "Sunflower meal", "Wheat gluten", "Canola oil", "Linseed oil", "Wheat", "Pea starch", "Corn starch", "Microingredients")))) |> 
  mutate(diet = gsub("_", " ", diet)) |> 
  rename(diet_prop = prop)




marine_plant_pressure_ingredient_disagg_combined_df <- readRDS(here("data/tidy_data/pressures/marine_plant_disaggregated_pressures_by_ingredient_combined_df.rds"))

marine_plant_pressure_ingredient_disagg_combined_df <- 
  marine_plant_pressure_ingredient_disagg_combined_df |> 
  group_by(diet, source) |> 
  nest() |> 
  mutate(total = map(data, ~(sum(.$sum)))) |> 
  unnest(c(data, total)) |> 
  mutate(prop = sum/total) |> 
  mutate(ingredient = str_to_sentence(ingredient)) |> 
  left_join(y = diets, by = c("diet", "ingredient" = "ingredients")) |> 
  mutate(weighted_prop = prop/diet_prop)

#Weighted contributions

summarised_ingredient_pressure_df <- 
  marine_plant_pressure_ingredient_disagg_combined_df |> 
  group_by(diet, pressure, ingredient) |> 
  summarise(mean = mean(weighted_prop, na.rm=TRUE),
            sd = sd(weighted_prop, na.rm=TRUE),
            se = sd/sqrt(500),
            median = median(weighted_prop, na.rm = TRUE),
            lq = quantile(weighted_prop, 0.25),
            uq = quantile(weighted_prop, 0.75), .groups = "drop") |> 
  mutate(diet = if_else(diet == "marine diet", true = "Marine-dominant feed", false = "Plant-dominant feed")) |> 
  mutate(ingredient = if_else(ingredient == "foragefish", true = "Forage fish", false = ingredient)) |> 
  mutate(pressure = factor(pressure, levels = c("Water", "Phosphorus", "Nitrogen", "GHG", "Disturbance")))



(marine_diet_ingredients_p <- 
  ggplot(data = summarised_ingredient_pressure_df |> 
           filter(diet == "Marine-dominant feed"),
                aes(x = reorder(ingredient, mean), y = mean, fill=pressure))+
  coord_flip(clip="off")+
  geom_col(aes(fill=pressure), position = "dodge", width = 0.75, alpha = 0.8)+
  geom_errorbar(aes(x = ingredient, ymin = if_else(mean-sd<0, true = 0, false = mean-sd), ymax = mean+sd), position = position_dodge(width=0.75), width = 0, size = 0.2)+
  scale_fill_manual(values = rev(rcartocolor::carto_pal(n=5, name = "Safe")))+
  #scale_y_continuous(expand = c(0,0), limits = c(0,2))+
  theme_bw()+
  theme(
        #panel.border = element_blank(),
        panel.grid.minor = element_blank(),
        # axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        # axis.text.y = element_blank(),
        # axis.ticks = element_blank(),
        plot.title = element_text(size=8, hjust=0.5),
        legend.position = c(0.8,0.25),
        # legend.title = element_text(size=7, colour = "grey20"),
        # legend.text = element_text(size=7, colour = "grey20"),
        legend.background = element_blank(),
        # legend.key.size = unit(0.3, "cm"),
        #legend.box.spacing = unit(-0.2, "cm"),
        text = element_text(size=8))+
  labs(y = "Wgtd. contribution to Cumulative Pressure", fill = "Pressure", title = "Marine-dominant feed")+
  guides(fill = "none"))


ggsave(here("figures/marine_disaggreation_by_ingredient_pressure.jpg"), dpi = 600, height=12, width=8.9, units="cm")




(plant_diet_ingredients_p <- 
  ggplot(data = summarised_ingredient_pressure_df |> 
           filter(diet == "Plant-dominant feed"),
         aes(x = reorder(ingredient, mean), y = mean, fill=pressure))+
  coord_flip(clip="off")+
  geom_col(aes(fill=pressure), position = "dodge", width = 0.85, alpha = 0.8)+
  geom_errorbar(aes(x = ingredient, ymin = if_else(mean-sd<0, true = 0, false = mean-sd), ymax = mean+sd), position = position_dodge(width=0.85), width = 0, size = 0.2)+
  scale_fill_manual(values = rev(rcartocolor::carto_pal(n=5, name = "Safe")))+
  #scale_y_continuous(expand = c(0,0), limits = c(0,2))+
  theme_bw()+
  theme(
        #panel.border = element_blank(),
        panel.grid.minor = element_blank(),
        # axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        # axis.text.y = element_blank(),
        # axis.ticks = element_blank(),
        plot.title = element_text(size=8, hjust=0.5),
        legend.position = c(0.8,0.13),
        legend.title = element_text(size=7, colour = "grey20"),
        legend.text = element_text(size=7, colour = "grey20"),
        legend.background = element_rect(fill="transparent"),
        legend.key.size = unit(0.3, "cm"),
        #legend.box.spacing = unit(-0.2, "cm"),
        text = element_text(size=8))+
  labs(y = "Wgtd. contribution to Cumulative Pressure", fill = "Pressure", title = "Plant-dominant feed"))+
  guides(fill = guide_legend(reverse = TRUE))


ggsave(here("figures/plant_disaggreation_by_ingredient_pressure.jpg"), dpi = 600, height=12, width=8.9, units="cm")
 

(marine_diet_ingredients_p|plant_diet_ingredients_p) +
  plot_layout(guides = "keep")+
  guides(fill = guide_legend(reverse = TRUE))+
  plot_annotation(tag_levels = "a")&
  theme(plot.tag = element_text(size=8))
ggsave(here("figures/disaggregation_of_pressure_by_ingredient.jpg"), dpi = 600, width = 18, height =13, units="cm")




#Proportional contributions


summarised_ingredient_pressure_df2 <- 
  marine_plant_pressure_ingredient_disagg_combined_df |> 
  group_by(diet, pressure, ingredient) |> 
  summarise(mean = mean(prop, na.rm=TRUE),
            sd = sd(prop, na.rm=TRUE),
            se = sd/sqrt(500),
            median = median(prop, na.rm = TRUE),
            lq = quantile(prop, 0.25),
            uq = quantile(prop, 0.75), .groups = "drop") |> 
  mutate(diet = if_else(diet == "marine diet", true = "Marine-dominant feed", false = "Plant-dominant feed")) |> 
  mutate(ingredient = if_else(ingredient == "foragefish", true = "Forage fish", false = ingredient)) |> 
   mutate(pressure = factor(pressure, levels = c("Water", "Phosphorus", "Nitrogen", "GHG", "Disturbance")))



(marine_diet_ingredients_p2 <- 
  ggplot(data = summarised_ingredient_pressure_df2 |> 
           filter(diet == "Marine-dominant feed"),
                aes(x = reorder(ingredient, mean), y = mean, fill=pressure))+
  coord_flip(clip="off")+
  geom_col(aes(fill=pressure), position = "dodge", width = 0.75, alpha = 0.8)+
  geom_errorbar(aes(x = ingredient, ymin = if_else(mean-sd<0, true = 0, false = mean-sd), ymax = mean+sd), position = position_dodge(width=0.75), width = 0, size = 0.2)+
  scale_fill_manual(values = rev(rcartocolor::carto_pal(n=5, name = "Safe")))+
    scale_y_continuous(expand = c(0,0), limits = c(0,0.302))+
  theme_bw()+
  theme(
        #panel.border = element_blank(),
        panel.grid.minor = element_blank(),
        # axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        # axis.text.y = element_blank(),
        # axis.ticks = element_blank(),
        plot.title = element_text(size=8, hjust=0.5),
        legend.position = c(0.8,0.25),
        legend.title = element_text(size=7, colour = "grey20"),
        legend.text = element_text(size=7, colour = "grey20"),
        legend.background = element_rect(fill="transparent"),
        legend.key.size = unit(0.3, "cm"),
        #legend.box.spacing = unit(-0.2, "cm"),
        text = element_text(size=8))+
  labs(y = "Prop. of Cumulative Pressure Index", fill = "Pressure", title = "Marine-dominant feed")+
  guides(fill = "none"))


ggsave(here("figures/supplementary/marine_disaggreation_by_ingredient_pressure_prop.jpg"), dpi = 600, height=12, width=8.9, units="cm")




(plant_diet_ingredients_p2 <- 
  ggplot(data = summarised_ingredient_pressure_df2 |> 
           filter(diet == "Plant-dominant feed"),
         aes(x = reorder(ingredient, mean), y = mean, fill=pressure))+
  coord_flip(clip="off")+
  geom_col(aes(fill=pressure), position = "dodge", width = 0.85, alpha = 0.8)+
  geom_errorbar(aes(x = ingredient, ymin = if_else(mean-sd<0, true = 0, false = mean-sd), ymax = mean+sd), position = position_dodge(width=0.85), width = 0, size = 0.2)+
  scale_fill_manual(values = rev(rcartocolor::carto_pal(n=5, name = "Safe")))+
  scale_y_continuous(expand = c(0,0), limits = c(0,0.302))+
  theme_bw()+
  theme(
        #panel.border = element_blank(),
        panel.grid.minor = element_blank(),
        # axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        # axis.text.y = element_blank(),
        # axis.ticks = element_blank(),
        plot.title = element_text(size=8, hjust=0.5),
        legend.position = c(0.8,0.13),
        legend.title = element_text(size=7, colour = "grey20"),
        legend.text = element_text(size=7, colour = "grey20"),
        legend.background = element_rect(fill="transparent"),
        legend.key.size = unit(0.3, "cm"),
        #legend.box.spacing = unit(-0.2, "cm"),
        text = element_text(size=8))+
  labs(y = "Prop. of Cumulative Pressure Index", fill = "Pressure", title = "Plant-dominant feed"))+
  guides(fill = guide_legend(reverse = TRUE))


ggsave(here("figures/supplementary/plant_disaggreation_by_ingredient_pressure_prop.jpg"), dpi = 600, height=12, width=8.9, units="cm")


 

(marine_diet_ingredients_p2|plant_diet_ingredients_p2) +
 plot_layout(guides = "keep")+
  guides(fill = guide_legend(reverse = TRUE))+
  plot_annotation(tag_levels = "a")&
  theme(plot.tag = element_text(size=8))
ggsave(here("figures/supplementary/disaggregation_of_pressure_by_ingredient_prop.jpg"), dpi = 600, width = 18, height =13, units="cm")



```



Now some case studies of ingredients 


Fish oil - disturbance

```{r}

list.files(here("data/spatial/marine_diet/int"), pattern = "fish oil_disturbance_km2_gp", full=TRUE) |> rast() |> global(fun = sum, na.rm=TRUE)

list.files(here("data/spatial/plant_diet/int"), pattern = "fish oil_disturbance_km2_gp", full=TRUE) |> rast() |> global(fun = sum, na.rm=TRUE)



marine_diet_disturbance <- 
  bind_rows(
  
  fo_low_epc_r <- list.files(here("data/spatial/marine_diet/int"), pattern = "fish oil_disturbance_km2_gp", full=TRUE) |> pluck(1) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "Eastern Pacific"),
  
  fo_hi_nat_r <- list.files(here("data/spatial/marine_diet/int"), pattern = "fish oil_disturbance_km2_gp", full=TRUE) |> pluck(2) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "North Atlantic")
) |> mutate(pressure = "Disturbance km2",
            diet = "Marine-dominant feed") |> 
  rename(value = lyr.1)



plant_diet_disturbance <- 
  bind_rows(
  
  fo_low_epc_r <- list.files(here("data/spatial/plant_diet/int"), pattern = "fish oil_disturbance_km2_gp", full=TRUE) |> pluck(1) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "Eastern Pacific"),
  
  fo_hi_nat_r <- list.files(here("data/spatial/plant_diet/int"), pattern = "fish oil_disturbance_km2_gp", full=TRUE) |> pluck(2) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "North Atlantic")
) |> mutate(pressure = "Disturbance km2",
            diet = "Plant-dominant feed") |> 
  rename(value = lyr.1)





fish_oil_disturbance_epc_nat <- bind_rows(marine_diet_disturbance, plant_diet_disturbance)

get_bounds <- fish_oil_disturbance_epc_nat |> select(x,y, value) |> rast(type="xyz")
crs(get_bounds) <- gp_crs
fish_oil_bounds <- st_bbox(get_bounds)

fo_bounds_df <- tibble(xmin = fish_oil_bounds["xmin"] , xmax = fish_oil_bounds["xmax"] , ymin = fish_oil_bounds["ymin"] , ymax = fish_oil_bounds["ymax"])


fo_dist_p <- ggplot()+
  geom_rect(data = fo_bounds_df, aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), fill = "aliceblue", alpha = 0.7)+
  geom_tile(data = fish_oil_disturbance_epc_nat, aes(x = x, y = y, fill = value))+
  geom_sf(data = ne_countries(scale = 10, returnclass = "sf") |> st_transform(crs = gp_crs) |> st_crop(y = fish_oil_bounds), size = NA, fill="grey75")+
  facet_grid(rows = vars(source), cols = vars(diet))+
  scale_fill_gradientn(colours = rcartocolor::carto_pal(n = 7, name = "SunsetDark"))+
  theme_bw()+
  theme(axis.text = element_blank(),
        axis.title.y = element_text(vjust = 3, size = 12),
        axis.title.x = element_blank(),
        axis.ticks = element_blank(),
        legend.box.spacing = unit(-0.8, "cm"),
        legend.title = element_text(size = 8))+
  labs(fill = expression(paste("Disturbance (km"^{2}*")")), y = "Fish oil - Disturbance")
  

ggsave(filename = here("explore/fish_oil_comparisons.jpg"), device = "jpg", dpi = 600, width = 18, height = 12, units = "cm")


```


Wheat - GHG

```{r}
list.files(here("data/spatial/marine_diet/int"), pattern = "wheat_N_gp", full=TRUE) |> rast() |> global(fun = sum, na.rm=TRUE)

list.files(here("data/spatial/plant_diet/int"), pattern = "wheat_N_gp", full=TRUE) |> rast() |> global(fun = sum, na.rm=TRUE)



marine_diet_wheat_N <- 
  bind_rows(
  
  wheat_low_ukr_r <- list.files(here("data/spatial/marine_diet/int"), pattern = "wheat_N_gp", full=TRUE) |> pluck(3) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "Ukraine"),
  
  wheat_low_ukr_r <- list.files(here("data/spatial/marine_diet/int"), pattern = "wheat_N_gp", full=TRUE) |> pluck(2) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "Pakistan")
  ) |>  
  mutate(pressure = "kg N eq.",
            diet = "Marine-dominant feed") |> 
  rename(value = N_eq)



plant_diet_wheat_N <- 
  bind_rows(
  
  wheat_low_ukr_r <- list.files(here("data/spatial/plant_diet/int"), pattern = "wheat_N_gp", full=TRUE) |> pluck(3) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "Ukraine"),
  
  wheat_low_ukr_r <- list.files(here("data/spatial/plant_diet/int"), pattern = "wheat_N_gp", full=TRUE) |> pluck(2) |> rast() |> terra::as.data.frame(xy = TRUE) |> mutate(source = "Pakistan")
  ) |>  
  mutate(pressure = "kg N eq.",
            diet = "Plant-dominant feed") |> 
  rename(value = N_eq)



wheat_N_ukr_pak <- bind_rows(marine_diet_wheat_N, plant_diet_wheat_N)

get_bounds_2 <- wheat_N_ukr_pak |> select(x,y, value) |> rast(type="xyz")
crs(get_bounds_2) <- gp_crs
wheat_bounds <- st_bbox(get_bounds_2)

wheat_bounds_df <- tibble(xmin = wheat_bounds["xmin"] , xmax = wheat_bounds["xmax"] , ymin = wheat_bounds["ymin"] , ymax = wheat_bounds["ymax"])

wheat_N_p <- 
  ggplot()+
  geom_rect(data = wheat_bounds_df, aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax), fill = "aliceblue", alpha = 0.7)+
  geom_sf(data = ne_countries(scale = 10, returnclass = "sf") |> st_transform(crs = gp_crs) |> st_crop(y = wheat_bounds), size = 0.3, colour="grey50", fill="grey75")+
  geom_tile(data = wheat_N_ukr_pak, aes(x = x, y = y, fill = value))+
  
  facet_grid(rows = vars(source), cols = vars(diet))+
  scale_fill_gradientn(colours = rcartocolor::carto_pal(n = 7, name = "TealGrn"))+
  theme_bw()+
  theme(axis.text = element_blank(),
        axis.title.y = element_text(vjust = 3, size = 12),
        axis.title.x = element_blank(),
        axis.ticks = element_blank(),
        legend.title = element_text(size = 8),
        plot.margin = unit(c(t = 0, b = 0, l=0.3, r=0.3), "cm"))+
  labs(fill = "kg N eq.", y = "Wheat - Nitrogen pollution")
  

ggsave(filename = here("explore/wheat_comparisons.jpg"), device = "jpg", dpi = 600, width = 18, height = 12, units = "cm")




```

Combine plots

```{r}

(fo_dist_p/wheat_N_p)+
  plot_annotation(tag_levels = "a")+
  plot_layout(heights = c(1.2,1))

ggsave(filename = here("figures/ingredient_casestudies.jpg"), device = "jpg", dpi = 600, width = 18, height = 26, units = "cm")




```
