==> scripts/18_mapping_CPI.Rmd <==
---
title: "Mapping cumulative_pressures"
author: "Rich Cottrell"
date: "14/02/2023"
output: html_document
editor_options: 
  chunk_output_type: console
---
Set up
```{r}
#LIBRARIES
library(tidyverse)
#library(ggpubr)
library(patchwork)
library(here)
library(sf)
library(terra)
library(rnaturalearth)
library(rnaturalearthdata)
library(RColorBrewer)
library(parallel)
library(ggdist)
library(data.table)
library(dtplyr)


# COLOURS
red <- "#B90000"
d_brown <- "#515256"
m_brown <- "#B27B54"
l_brown <- "#BC995F"
green <- "#A8B072"
yellow<- "#EFCE71"
l_yellow <- "#F7F6C1"
light_gray <- "#F8F9FA"
discrete_pal <- c(l_yellow, yellow, green, l_brown, m_brown, d_brown, "firebrick")
continuous_pal <-  c(light_gray, colorRampPalette(discrete_pal, space="Lab", bias = 6)(10000))
final_palette <- c(light_gray, continuous_pal)



#SOURCE FILES
source(here("src/functions.R"))



#SPATIAL INFO
# Base raster wgs84
base_rast <- rast(res=0.5)
ext(base_rast) <- c(-180, 180, -90, 90)

#Gall peters projection
gp_crs <- "+proj=cea +lon_0=0 +x_0=0 +y_0=0 +lat_ts=45 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
#raster template 
eq_area_rast <- project(base_rast, gp_crs) 
values(eq_area_rast) <- 1:ncell(eq_area_rast)
cellSize(eq_area_rast, unit="km", transform = FALSE)

#Mollweide projection
moll_rast <- project(base_rast, "ESRI:54009")
values(moll_rast) <- 1:ncell(moll_rast)
cellSize(moll_rast, unit="km", transform = FALSE)



#SHAPEFILES
countries <-    ne_countries(scale = "medium", returnclass = "sf") |> st_transform(crs = crs(moll_rast))
bbox <- ne_download(scale = 50, type = "wgs84_bounding_box", category = "physical", returnclass = "sf") |> st_transform(crs = crs(moll_rast))
coastline <- ne_coastline(scale = 50, returnclass = "sf") |> st_transform(crs = crs(moll_rast))
select <- dplyr::select
values <- terra::values


#ALLOCATION METHOD (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
# this_allocation_method <- "mass_allocation"

```

Bring in the origin-ingredient combinations within each diet.  

```{r}
marine_diets <- readRDS(here("data/tidy_data/diets/ingredient_source_combinations_marine_diet.rds")) 
plant_diets <- readRDS(here("data/tidy_data/diets/ingredient_source_combinations_plant_diet.rds")) 
```


# Compare marine-dominant and plant-dominant feeds from a selectiom of different sources to illustrate the concept.


Combine all values for pressures for a marine diet

```{r}
marine_source_ingredient <- 
  marine_diets |> 
  select(ingredient) |> 
  mutate(source_1 = c(1, 1, 3, 1, 2, 5, 1),
         source_2 = c(2, 3, 1, 2, 4, 2, 3),
         source_3 = c(3, 5, 5, 4, 3, 2, 4))|> 
   mutate(row = row_number()) |> 
   group_split(row) 
  # select(ingredient, source_1 = `5`, source_2 = `9`, source_3 = `21`) |> 
  # 
  # mutate(source_1 = case_when(ingredient == "guar meal"~ as.integer(1),
  #                           TRUE ~ source_1),
  #        source_2 = case_when(ingredient == "guar meal"~ as.integer(1),
  #                           TRUE ~ source_2),
  #        source_3 = case_when(ingredient == "guar meal"~ as.integer(1),
  #                           TRUE ~ source_3)) |> 
  # 
#test dataframe
this_df <- marine_source_ingredient[[4]]

#Now prepare the files corresponding to the first source for each ingredient for each pressure
sources <- c("source_1", "source_2", "source_3")
source = "source_1"

#mass allocation 
marine_source_file_list <- list()

for(source in sources){
    
    marine_diet_files <- 
      map(.x = marine_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient, " (", this_allocation_method, ") ", source)
        
        these_ingredient_files <- list.files(here("data/spatial/marine_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        these_ingredient_files <- these_ingredient_files[grepl(this_allocation_method, these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_production_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) &!grepl("processing", these_ingredient_files)][[this_source_ingredient]]
         this_disturbance_processing_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_production_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & !grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        this_ghg_processing_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        
        #nutrients
        this_nutrient_production_file <- these_ingredient_files[grepl("nutrient", these_ingredient_files) & !grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        this_nutrient_processing_file <- these_ingredient_files[grepl("nutrient", these_ingredient_files) &grepl("processing", these_ingredient_files)][[this_source_ingredient]]
      
        #water
        this_h2o_production_file <- these_ingredient_files[grepl("water", these_ingredient_files) & !grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        this_h2o_processing_file <- these_ingredient_files[grepl("water", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_production_file,this_disturbance_processing_file, this_ghg_production_file, this_ghg_processing_file, this_nutrient_production_file, this_nutrient_processing_file, this_h2o_production_file, this_h2o_processing_file)
        
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    marine_source_file_list[[source]] <- marine_diet_files
    
  }
  


#just need to add scores as pressures and ingredients are additive at this point

marine_diet_rasters <- 
  
  map(.x = sources, .f = \(this_source){
  
  these_files <- marine_source_file_list[[this_source]]
  
  this_cum_rast <- app(rast(these_files), sum, na.rm=TRUE)
  
  
})


```


Combine all values for a plant-dominant diet

```{r}
plant_source_ingredient <- 
  plant_diets |> 
  select(ingredient) |> 
 mutate(source_1 = c(1, 1, 2, 4, 5, 2, 3, 1, 3, 4, 2, 2, 1),
         source_2 = c(2, 2, 3, 3, 4, 2, 3, 3, 5, 1, 3, 5, 2),
         source_3 = c(3, 3, 1, 5, 3, 1, 4, 2, 1, 2, 1, 3, 5)) |> 
   mutate(row = row_number()) |> 
  group_split(row) 
# 
#   select(ingredient, source_1 = `3`, source_2 = `4`, source_3 = `18`) |> 
#   mutate(source_1 = case_when(ingredient == "linseed oil"~ as.integer(1),
#                             TRUE ~ source_1),
#          source_2 = case_when(ingredient == "linseed oil"~ as.integer(1),
#                             TRUE ~ source_2),
#          source_3 = case_when(ingredient == "linseed oil"~ as.integer(1),
#                             TRUE ~ source_3)) |> 
#   mutate(row = row_number()) |> 
#   group_split(row) 

#test dataframe
this_df <- plant_source_ingredient[[6]]
sources <- c("source_1", "source_2", "source_3")
plant_source_file_list <- list()

#Now prepare the files corresponding to the first source for each ingredient for each pressure
for(source in sources){
    
    plant_diet_files <- 
      map(.x = plant_source_ingredient, .f = \(this_df){
        
  
        
        this_ingredient <- this_df$ingredient
        
        message("processing...", this_ingredient)
        
        
        these_ingredient_files <- list.files(here("data/spatial/plant_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
        these_ingredient_files <- these_ingredient_files[grepl(this_allocation_method, these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
       #disturbance
        this_disturbance_production_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) &!grepl("processing", these_ingredient_files)][[this_source_ingredient]]
         this_disturbance_processing_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_production_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & !grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        this_ghg_processing_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        
        #nutrients
        this_nutrient_production_file <- these_ingredient_files[grepl("nutrient", these_ingredient_files) & !grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        this_nutrient_processing_file <- these_ingredient_files[grepl("nutrient", these_ingredient_files) &grepl("processing", these_ingredient_files)][[this_source_ingredient]]
      
        #water
        this_h2o_production_file <- these_ingredient_files[grepl("water", these_ingredient_files) & !grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        this_h2o_processing_file <- these_ingredient_files[grepl("water", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        
        these_diet_pressures <- c(this_disturbance_production_file,this_disturbance_processing_file, this_ghg_production_file, this_ghg_processing_file, this_nutrient_production_file, this_nutrient_processing_file, this_h2o_production_file, this_h2o_processing_file)
        
        
  
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    plant_source_file_list[[source]] <- plant_diet_files
    
  }
  

#stack rasters  
plant_diet_rasters <- 
  
  map(.x = sources, .f = \(this_source){
  
  these_files <- plant_source_file_list[[this_source]]
  
  this_cum_rast <- app(rast(these_files), sum, na.rm=TRUE)
  
  
})



```

Figure - maps comparing diet and sourcing

```{r}
max_marine_value <- marine_diet_rasters |> map(no_na_values) |> unlist() |> as.numeric() |> max()
max_plant_value <- plant_diet_rasters|> map(no_na_values) |> unlist() |> as.numeric() |> max()
max_cum_value <- max(max_marine_value, max_plant_value)
(sum_marine_value <- marine_diet_rasters |> map(no_na_values) |> map(sum))
(sum_plant_value <- plant_diet_rasters|> map(no_na_values) |> map(sum))

source_marine_1 <- round(sum_marine_value[[1]], digits = 2) 
source_marine_2 <- round(sum_marine_value[[2]], digits = 2)
source_marine_3 <- round(sum_marine_value[[3]], digits = 2)
source_plant_1 <- round(sum_plant_value[[1]], digits = 3)
source_plant_2 <- round(sum_plant_value[[2]], digits = 2)
source_plant_3 <- round(sum_plant_value[[3]], digits = 2)



# MARINE DIETS



#source 1
marine_diet1_df <- terra::as.data.frame(marine_diet_rasters[[1]], xy=TRUE)
(marine_sourcing_p1 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, linewidth=0.08)+
  geom_sf(data = countries, fill= "grey95", colour= "white", linewidth=0.08)+
  geom_tile(data = marine_diet1_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_blank(),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(
                      colours = continuous_pal, 
                       limits = c(0, max_cum_value)
                       #breaks = c(0,max_cum_value*0.25, max_cum_value/2, max_cum_value*.75, max_cum_value)
                      #  labels = c(">0","0.005", "0.010")
                       )+
    coord_sf(clip="off")+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = " Sourcing\ncombination 1", angle = 90)+
    annotate("text", x = 17985650*0.6, y = -7298155*1.35, label = as.expression(bquote(Sigma~CPI==.(format(round(source_marine_1, digits = 2), nsmall=2)))), size = 3)
  )



#source 2
marine_diet2_df <- terra::as.data.frame(marine_diet_rasters[[2]], xy=TRUE)
(marine_sourcing_p2 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, linewidth=0.08)+
  geom_sf(data = countries, fill="grey95", colour= "white", linewidth=0.15)+
  geom_tile(data = marine_diet2_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_blank(),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
    coord_sf(clip="off")+
  scale_fill_gradientn(colours = continuous_pal,
                       limits = c(0, max_cum_value)#,
                      # breaks = c(0,0.005, 0.010),
                      #  labels = c(">0","0.005", "0.010")
                      )+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = "Sourcing\ncombination 2", angle = 90)+
    annotate("text", x = 17985650*0.6, y = -7298155*1.35, label = as.expression(bquote(Sigma~CPI==.(format(round(source_marine_2, digits = 2), nsmall=2)))), size = 3)
  )


#source 3
marine_diet3_df <- terra::as.data.frame(marine_diet_rasters[[3]], xy=TRUE)
(marine_sourcing_p3 <- 
  ggplot()+
  geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, linewidth=0.08)+
  geom_sf(data = countries, fill="grey95", colour= "white", linewidth=0.15)+
  geom_tile(data = marine_diet3_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_blank(),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
    coord_sf(clip="off")+
  scale_fill_gradientn(colours = continuous_pal, 
                       limits = c(0, max_cum_value)#,
                      # breaks = c(0,0.005, 0.010),
                      #  labels = c(">0","0.005", "0.010")
                      )+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
     annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = "Sourcing\ncombination 3", angle = 90)+
    annotate("text", x = 17985650*0.6, y = -7298155*1.35, label = as.expression(bquote(Sigma~CPI==.(format(round(source_marine_3, digits = 2), nsmall=2)))), size = 3)
  )



#PLANT DIETS
#source 1
plant_diet1_df <- terra::as.data.frame(plant_diet_rasters[[1]], xy=TRUE)
(plant_sourcing_p1 <- 
  ggplot()+
  geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, linewidth=0.08)+
  geom_sf(data = countries, fill="grey95", colour= "white", linewidth=0.15)+
  geom_tile(data = plant_diet1_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth = 0.15)+
  theme(panel.grid = element_blank(),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = continuous_pal, 
                       limits = c(0, max_cum_value)#,
                      # breaks = c(0,0.005, 0.010),
                      #  labels = c(">0","0.005", "0.010")
                      )+
    coord_sf(clip="off")+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = 17985650*0.6, y = -7298155*1.35, label = as.expression(bquote(Sigma~CPI==.(format(round(source_plant_1, digits = 2), nsmall=2)))), size = 3)
  )



#source 2
plant_diet2_df <- terra::as.data.frame(plant_diet_rasters[[2]], xy=TRUE)
(plant_sourcing_p2 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, linewidth=0.08)+
  geom_sf(data = countries, fill="grey95", colour= "white", linewidth=0.15)+
  geom_tile(data = plant_diet2_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth = 0.15)+
  theme(panel.grid = element_blank(),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
    coord_sf(clip="off")+
  scale_fill_gradientn(colours = continuous_pal, 
                       limits = c(0, max_cum_value)#,
                      # breaks = c(0,0.005, 0.010),
                      #  labels = c(">0","0.005", "0.010")
                      )+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = 17985650*0.6, y = -7298155*1.35, label = as.expression(bquote(Sigma~CPI==.(format(round(source_plant_2, digits = 2), nsmall=2)))), size = 3)
  )

  
#source 3
plant_diet3_df <- terra::as.data.frame(plant_diet_rasters[[3]], xy=TRUE)
(plant_sourcing_p3 <- 
  ggplot()+
   geom_sf(data = bbox, colour="grey75", fill= "aliceblue", alpha=0.6, linewidth=0.08)+
  geom_sf(data = countries, fill="grey95", colour= "white", linewidth=0.15)+
  geom_tile(data = plant_diet3_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline, colour="grey45", linewidth = 0.15)+
  theme(panel.grid = element_blank(),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=10),
        legend.text = element_text(size=10),
        legend.direction = "horizontal",
        legend.position = "bottom",
      legend.box.spacing = unit(-1, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
    coord_sf(clip="off")+
  scale_fill_gradientn(colours = continuous_pal, 
                       limits = c(0, max_cum_value)#,
                      # breaks = c(0,0.005, 0.010),
                      #  labels = c(">0","0.005", "0.010")
                      )+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 10, barheight = 0.5, units = "cm"))+
    annotate("text", x = 17985650*0.6, y = -7298155*1.35, label = as.expression(bquote(Sigma~CPI==.(format(round(source_plant_3, digits=2), nsmall=2)))), size =3)
  )
# bar_legend <- get_legend(plant_sourcing_p3)
# 
# 
# source_diet_map <- 
#   ggarrange(
#     ggarrange(marine_sourcing_p1, plant_sourcing_p1,
#               marine_sourcing_p2, plant_sourcing_p2,
#               marine_sourcing_p3, plant_sourcing_p3,
#               nrow = 3, 
#               ncol = 2,
#               legend = "none"),
#     
#     bar_legend,
#     nrow = 2,
#     ncol=1,
#     heights = c(11,1))
source_diet_map <- 
  (marine_sourcing_p1|plant_sourcing_p1 )/
  (marine_sourcing_p2|plant_sourcing_p2)/
  (marine_sourcing_p3|plant_sourcing_p3)+
  plot_layout(guides = "collect")+
  plot_annotation(tag_levels = "a", #surely there is a better way than below?
                  title = "                   Fish-dominant feed                       Plant-dominant feed") & 
  theme(plot.tag = element_text(size=10), 
        plot.tag.position = c(0.1, 1),
        legend.position = "bottom")
          

          
ggsave(plot = source_diet_map, filename = sprintf(here("figures/source_vs_diet_visualisation_map_%s.jpg"), this_allocation_method), device = "jpg", dpi = 600, width = 18, height = 16, units = "cm")
ggsave(plot = source_diet_map, filename = sprintf(here("figures/source_vs_diet_visualisation_map_%s.pdf"), this_allocation_method), device = "pdf", dpi = 600, width = 18, height = 16, units = "cm")

```

Save one cumulative pressure map for conceptual plot
```{r}
    
  (ggplot()+
  geom_sf(data = bbox |> st_transform(crs = "ESRI:54009"), colour="grey75", fill= "aliceblue", alpha=0.6, size = 0.3)+
  geom_sf(data = countries |> st_transform(crs = "ESRI:54009"), fill="grey90", colour= "white", linewidth=0.15)+
  geom_tile(data = marine_diet3_df, aes(x = x, y=y, fill=sum))+
  geom_sf(data = coastline |> st_transform(crs = "ESRI:54009"), colour="grey45", linewidth= 0.15)+
  theme(panel.grid = element_line(linewidth=0.25, colour="grey75"),
        panel.background = element_rect(fill = "transparent"),
        axis.title = element_blank(),
        legend.title = element_text(size=14),
        legend.text = element_text(size=14),
        legend.direction = "horizontal",
        legend.position = "bottom",
        legend.background = element_rect(fill = "transparent"),
      legend.box.spacing = unit(-0.3, "cm"),
        plot.margin = margin(r=-0.5,l=-0.5))+
  scale_fill_gradientn(colours = final_palette, 
                       limits = c(0, max_cum_value)
                       # breaks = c(0,0.005, 0.010),
                       # labels = c(">0","0.005", "0.010")
                       )+
    guides(fill = guide_colorbar(direction = "horizontal", title.position = "top", title = "Cumulative pressure index", title.hjust = 0.5,
                                 barwidth = 12, barheight = 0.6, units = "cm"))
    #annotate("text", x = -22000000, y = sum(c(-9008064, 9009955))|> mean(), label = "Raw material origin\ncombination 3", angle = 90)+
    #annotate("text", x = 17985650*0.8, y = -7298155*1.2, label = as.expression(bquote(Sigma~CPI==.(source_marine_3))), size = 3)
  )

ggsave(here("figures/single_CPI_map.jpeg"), device = "jpg", dpi = 600, width = 8.9, height =7, units = "cm")



```



#clear the environment
```{r}
rm(list = ls(all.names = TRUE))
```
