==> scripts/19_CPI_permutation.Rmd <==
---
title: "CPI Permutation"
author: "Rich Cottrell"
date: "14/02/2023"
output: html_document
---

Set up

```{r}
library(tidyverse)
library(ggpubr)
library(patchwork)
library(here)
library(sf)
library(terra)
library(rnaturalearth)
library(rnaturalearthdata)
library(RColorBrewer)
library(ggdist)
library(data.table)
library(dtplyr)

select <- dplyr::select
values <- terra::values


#ALLOCATION METHOD (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"


marine_run <- !file.exists(sprintf(here("data/tidy_data/pressures/marine_disaggregated_pressures_by_ingredient_df_%s.rds"), this_allocation_method))
plant_run <- !file.exists(sprintf(here("data/tidy_data/pressures/plant_disaggregated_pressures_by_ingredient_df_%s.rds"), this_allocation_method))

```

Bring in the origin-ingredient combinations within each diet.

```{r}
marine_diets <- readRDS(here("data/tidy_data/diets/ingredient_source_combinations_marine_diet.rds")) 
plant_diets <- readRDS(here("data/tidy_data/diets/ingredient_source_combinations_plant_diet.rds")) 
```

#Disaggregation by pressure and ingredient

Disaggregate by ingredients and pressures - Marine diets - takes ~ 2.5 hours

```{r, eval= marine_run}

marine_source_ingredient <- 
  marine_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "guar meal"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)

#create vector for the sources
sources <- bind_rows(marine_source_ingredient) |> select(-c("ingredient", "row")) |>  names()  
#create vector for the value chain stage
stages <- c("production", "processing")



marine_source_file_list <- list()

# # Test data for loop
 this_df <- marine_source_ingredient[[2]]
source <- sources[[1]]


#Now list the files for all ingredients and sources
for(source in sources){
    
    marine_diet_files <- 
      map(.x = marine_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("Isolating pressure files for each ingredient-source combination - ", this_ingredient, "...", source)
        
        these_ingredient_files <- list.files(here("data/spatial/marine_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        
         these_ingredient_files <- these_ingredient_files[grepl(this_allocation_method, these_ingredient_files)]
        #sources
         
        this_source_ingredient <- this_df |> pull(source)
        
        
        #ISOLATE THE FILES FOR EACH INGREDIENT AND STAGE IN VALUE CHAIN
        
        #disturbance
        this_disturbance_production_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_disturbance_processing_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) & grepl("processing",these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_production_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_ghg_processing_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        #nutrients
        this_nutrient_production_file <- these_ingredient_files[grepl("_nutrient_", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_nutrient_processing_file <- these_ingredient_files[grepl("_nutrient_", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_production_file <- these_ingredient_files[grepl("water", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_h2o_processing_file <- these_ingredient_files[grepl("water", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        
        #concatenate and return files needed for this ingredient
        these_diet_pressures <- c(this_disturbance_production_file,this_disturbance_processing_file, #disturbance 
                                  this_ghg_production_file, this_ghg_processing_file,  #ghg 
                                  this_nutrient_production_file, this_nutrient_processing_file, #nutrients
                                  this_h2o_production_file, this_h2o_processing_file) #water
        
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    marine_source_file_list[[source]] <- marine_diet_files #add these files to list
    
  }





  #stack and sum by individual pressures for each ingredient source combination and stage in the value chain

marine_ingredients <- 
  readRDS(here("data/tidy_data/demand/total_ingredient_demand_by_diet.rds")) |> 
  filter(diet=="marine_diet" & ingredient_demand_tonnes>0) |> pull(ingredients)





this_ingredient_pressure_df_list <- list()

all_ingredients_pressures_df_list <- list()


for(source in sources){
  
  message("Stacking pressure files - marine diet ", source)
  
  this_combination_all_pressures_list <- marine_source_file_list[[source]]
  
  #isolate the value chain stage
  
  for(stage in stages){
    
    this_stage_pressures_list <- this_combination_all_pressures_list[grepl(stage, this_combination_all_pressures_list)]
    
    for(ingredient in marine_ingredients){
      
      this_ingredient_pressure_list <- this_stage_pressures_list[grepl(paste0(ingredient, "_"), this_stage_pressures_list)]
      
      #########Separate GHG files, read rasters into a list, stack all GHG layers
      
      message("GHG pressure layer for - marine diet...", source, " - ", ingredient, " - ", stage)
      
      this_combination_ghg_list <- this_ingredient_pressure_list[grep("ghg", this_ingredient_pressure_list)]
      
      these_ghg_rasters <- map(.x = this_combination_ghg_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_ghg_df <- global(app(rast(these_ghg_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "GHG", source = source, stage = stage, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      
      ##########Separate water files, read rasters into a list, stack all water layers
      
      message("Water pressure layer for - marine diet...",source, " - ", ingredient, " - ", stage)
      
      this_combination_water_list <- this_ingredient_pressure_list[grep("water", this_ingredient_pressure_list)]
      
      these_water_rasters <- map(.x = this_combination_water_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_water_df <- global(app(rast(these_water_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Water", source = source, stage= stage, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      #########Separate nutrient files, read rasters into a list, stack all nutrient layers
      
      message("Nutrient pressure layer for - marine diet...", source, " - ", ingredient, " - ", stage)
      
      this_combination_nutrient_list <-  this_ingredient_pressure_list[grep("_nutrient_", this_ingredient_pressure_list)]
      
      these_nutrient_rasters <- map(.x = this_combination_nutrient_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_nutrient_df <- global(app(rast(these_nutrient_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Nutrients", source = source,stage= stage,  ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      #########Separate disturbance files, read rasters into a list, stack all disturbance layers
      
      message("Disturbance pressure layer for - marine diet...", source, " - ", ingredient, " - ", stage)
      
      this_combination_disturbance_list <-  this_ingredient_pressure_list[grep("disturbance", this_ingredient_pressure_list)]
      
      these_disturbance_rasters <- map(.x = this_combination_disturbance_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_disturbance_df <- global(app(rast(these_disturbance_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Disturbance", source = source, stage= stage,  ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      this_ingredient_pressure_df_list[[ingredient]] <- bind_rows(this_ghg_df, this_water_df, this_nutrient_df, this_disturbance_df)
      
      
      
    }
    all_ingredients_pressures_df_list[[paste("source", source, "stage", stage)]] <- do.call(bind_rows, this_ingredient_pressure_df_list)
  }
  
  
}

marine_diet_pressures_by_ingredient_df <- bind_rows(all_ingredients_pressures_df_list) |> mutate(diet = "marine diet")
saveRDS(marine_diet_pressures_by_ingredient_df, file = sprintf(here("data/tidy_data/pressures/marine_disaggregated_pressures_by_ingredient_df_%s.rds"), this_allocation_method))

marine_diet_pressures_by_ingredient_df |> pull(source) |> unique()
```

Disaggreagte by ingredients and pressures - Plant diets - takes ~6 hours

```{r, eval= plant_run}

plant_source_ingredient <- 
  plant_diets |>
  select(1:2001) |> 
  rename_at(vars(-ingredient), ~ paste("source", .x)) |>
  mutate(row = row_number()) |> 
  mutate_at(vars(-ingredient, -row), ~ case_when(ingredient == "linseed oil"~ as.integer(1),
                                           TRUE ~ .)) |> 
  group_split(row)

#Create vector for the sources
sources <- bind_rows(plant_source_ingredient) |> select(-c("ingredient", "row")) |>  names() 
#create vector for the value chain stage
stages <- c("production", "processing")

plant_source_file_list <- list()


#test data for loop
this_df <- plant_source_ingredient[[1]]
source <- sources[[1]]



#Now list the files for all ingredients and sources
for(source in sources){
    
    plant_diet_files <- 
      map(.x = plant_source_ingredient, .f = \(this_df){
        
        this_ingredient <- this_df$ingredient
        
        message("Isolating pressure files for each ingredient-source combination - ", this_ingredient, "...", source)
        
        #ISOLATE THE FILES NEEDED
        
        these_ingredient_files <- list.files(here("data/spatial/plant_diet/pressures"), pattern = paste0(this_ingredient, "_"), full.names = TRUE)
        these_ingredient_files <- these_ingredient_files[grepl(this_allocation_method, these_ingredient_files)]
        
        #sources
        this_source_ingredient <- this_df |> pull(source)
        
        #disturbance
        this_disturbance_production_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_disturbance_processing_file <- these_ingredient_files[grepl("disturbance", these_ingredient_files) & grepl("processing",these_ingredient_files)][[this_source_ingredient]]
        
        #ghg
        this_ghg_production_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_ghg_processing_file <- these_ingredient_files[grepl("ghg", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        #nutrients
        this_nutrient_production_file <- these_ingredient_files[grepl("_nutrient_", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_nutrient_processing_file <- these_ingredient_files[grepl("_nutrient_", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        #water
        this_h2o_production_file <- these_ingredient_files[grepl("water", these_ingredient_files) & grepl("production", these_ingredient_files)][[this_source_ingredient]]
        this_h2o_processing_file <- these_ingredient_files[grepl("water", these_ingredient_files) & grepl("processing", these_ingredient_files)][[this_source_ingredient]]
        
        
        #concatenate and return files needed for this ingredient
        these_diet_pressures <- c(this_disturbance_production_file,this_disturbance_processing_file, #disturbance 
                                  this_ghg_production_file, this_ghg_processing_file,  #ghg 
                                  this_nutrient_production_file, this_nutrient_processing_file, #nutrients
                                  this_h2o_production_file, this_h2o_processing_file) #water
        
        return(these_diet_pressures)
        
      }) |> 
      unlist()
    
    plant_source_file_list[[source]] <- plant_diet_files
    
}


  
plant_ingredients <- readRDS(here("data/tidy_data/demand/total_ingredient_demand_by_diet.rds")) |> 
  filter(diet=="plant_diet" & ingredient_demand_tonnes>0) |> 
  pull(ingredients)

#initiate lists 
this_ingredient_pressure_df_list <- list()
all_ingredients_pressures_df_list <- list()


for(source in sources){
  
  message("Stacking pressure files - plant diet ", source)
  
  this_combination_all_pressures_list <- plant_source_file_list[[source]]
  
  #isolate the production stage
  
  for(stage in stages){
    
    this_stage_pressures_list <- this_combination_all_pressures_list[grepl(stage, this_combination_all_pressures_list)]
    
    for(ingredient in plant_ingredients){
      
      this_ingredient_pressure_list <- this_stage_pressures_list[grepl(paste0(ingredient, "_"), this_stage_pressures_list)]
      
      #########Separate GHG files, read rasters into a list, stack all GHG layers
      
      message("GHG pressure layer for - plant diet...", source, " - ", ingredient, " - ", stage)
      
      
      this_combination_ghg_list <- this_ingredient_pressure_list[grep("ghg", this_ingredient_pressure_list)]
      
      these_ghg_rasters <- map(.x = this_combination_ghg_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_ghg_df <- global(app(rast(these_ghg_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "GHG", source = source, stage= stage,  ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      
      ##########Separate water files, read rasters into a list, stack all water layers
      
      message("Water pressure layer for - plant diet...",source, " - ", ingredient, " - ", stage)
      
      this_combination_water_list <- this_ingredient_pressure_list[grep("water", this_ingredient_pressure_list)]
      
      these_water_rasters <- map(.x = this_combination_water_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_water_df <- global(app(rast(these_water_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Water", source = source, stage= stage,  ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      #########Separate nutrient files, read rasters into a list, stack all nutrient layers
      
      message("Nutrient pressure layer for - plant diet...", source, " - ", ingredient, " - ", stage)
      
      this_combination_nutrient_list <-  this_ingredient_pressure_list[grep("_nutrient_", this_ingredient_pressure_list)]
      
      these_nutrient_rasters <- map(.x = this_combination_nutrient_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      this_nutrient_df <- global(app(rast(these_nutrient_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Nutrients", source = source, stage= stage,  ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      #########Separate disturbance files, read rasters into a list, stack all disturbance layers
      
      message("Disturbance pressure layer for - plant diet...", source, " - ", ingredient, " - ", stage)
      
      this_combination_disturbance_list <-  this_ingredient_pressure_list[grep("disturbance", this_ingredient_pressure_list)]
      
      these_disturbance_rasters <- map(.x = this_combination_disturbance_list, .f = \(this_file){
        
        this_rast <- rast(this_file)
        this_source <- tools::file_path_sans_ext(basename(sources(this_rast)))
        
        return(this_rast)
      })
      
      
      this_disturbance_df <- global(app(rast(these_disturbance_rasters), sum, na.rm=TRUE), sum, na.rm=TRUE) |> mutate(pressure = "Disturbance", source = source, stage= stage, ingredient = gsub(pattern = "_", replacement = "", ingredient), .before = sum)
      
      
      this_ingredient_pressure_df_list[[ingredient]] <- bind_rows(this_ghg_df, this_water_df, this_nutrient_df, this_disturbance_df)
      
      
    }
  
  
     all_ingredients_pressures_df_list[[paste("source", source, "stage", stage)]] <- do.call(bind_rows, this_ingredient_pressure_df_list)

  }
  
 
}


plant_diet_pressures_by_ingredient_df <- bind_rows(all_ingredients_pressures_df_list) |> mutate(diet = "plant diet")
saveRDS(plant_diet_pressures_by_ingredient_df, file = sprintf(here("data/tidy_data/pressures/plant_disaggregated_pressures_by_ingredient_df_%s.rds"), this_allocation_method))
```

Add marine and plant-dominant diet disaggregation by ingredient and pressure

```{r}
marine_diet_pressures_by_ingredient_df <- readRDS(sprintf(here("data/tidy_data/pressures/marine_disaggregated_pressures_by_ingredient_df_%s.rds"), this_allocation_method))

plant_diet_pressures_by_ingredient_df <- readRDS(sprintf(here("data/tidy_data/pressures/plant_disaggregated_pressures_by_ingredient_df_%s.rds"), this_allocation_method))

marine_plant_pressure_ingredient_disagg_combined_df <- bind_rows(marine_diet_pressures_by_ingredient_df, plant_diet_pressures_by_ingredient_df)

saveRDS(object = marine_plant_pressure_ingredient_disagg_combined_df, file = sprintf(here("data/tidy_data/pressures/marine_plant_disaggregated_pressures_by_ingredient_combined_df_%s.rds"),this_allocation_method))
```


#clear the environment

```{r}
rm(list = ls(all.names = TRUE))
```
