==> scripts/21_calculations.Rmd <==
---
title: "Calculating raw materials demands across scenarios"
author: "Rich Cottrell"
date: "20/02/2023"
output: html_document
---

Set up
```{r}

library(tidyverse)
library(here)
library(terra)

#ALLOCATION METHOD (un-comment preference)
source(here("1_choose_allocation_method.R")) #selects from set up script
#this_allocation_method <- "econ_allocation"
#this_allocation_method <- "ge_allocation"
#this_allocation_method <- "mass_allocation"
this_allocation_method

select <- dplyr::select

```


Total raw material demand
```{r}

crop_demand <- 
  readRDS(here("data/tidy_data/demand/total_crop_demand.rds")) |> 
  filter(total_crop_demand >0 ) |>
  rename(demand_from = ingredients,
         demand = total_crop_demand,
         type=FAOSTAT_name) |> 
  select(diet, demand_from, type, demand)

fish_demand <- 
  bind_rows(readRDS(here("data/tidy_data/demand/embodied_trimmingsfish_demand_by_source_diet.rds")) |> mutate(type="Trimmings"),
                         readRDS(here("data/tidy_data/demand/embodied_foragefish_demand_by_source_diet.rds")) |> mutate(type= "Wild-caught") |> rename(embodied_fish = embodied_foragefish)) |> 
  rename(demand = embodied_fish) |> select(diet, demand_from, fao_area_code, type, demand)


```



Basic calculations of diet vs.marine/terrestrial
```{r}


raw_material_iteration <- 
  map_df(.x = unique(fish_demand$fao_area_code), .f = \(this_area){
  
  return(bind_rows(fish_demand |> 
  filter(fao_area_code == this_area) |> 
  group_by(demand_from, fao_area_code, diet) |> 
  summarise(demand = sum(demand, na.rm=TRUE)) |> 
  ungroup()  |>
  rename(type = demand_from),
  
crop_demand |> 
  group_by(diet, type) |> 
  summarise(demand = sum(demand, na.rm=TRUE)) |>
  mutate(fao_area_code = this_area))|> 
    arrange(diet, fao_area_code, type)) 
  
})
  

raw_material_totals <- 
  raw_material_iteration |> 
  filter(type!="Total") |> 
  group_by(diet, fao_area_code) |>
  summarise(total_biomass = sum(demand)) |> 
  ungroup() |>
  group_by(diet) |>
  summarise(mean_biomass = mean(total_biomass, na.rm=TRUE),
            sd_biomass = sd(total_biomass, na.rm=TRUE))



saveRDS(raw_material_totals, file = sprintf(here("data/tidy_data/demand/total_raw_material_demand_%s.rds"), this_allocation_method))
  



raw_material_totals_by_origin <- 
  raw_material_iteration |> 
  filter(type!="Total") |> 
  mutate(origin = case_when(type %in% c("Fishmeal", "Fish oil") ~ "Marine",
                            TRUE ~ "Terrestrial")) |> 
  group_by(diet, origin, fao_area_code) |>
  summarise(total_biomass_by_origin = sum(demand)) |> 
  ungroup() |> 
  group_by(diet, origin) |> 
  summarise(mean_biomass_by_origin = mean(total_biomass_by_origin, na.rm=TRUE),
            sd_biomass_by_origin = sd(total_biomass_by_origin, na.rm=TRUE))


saveRDS(raw_material_totals_by_origin, file = sprintf(here("data/tidy_data/demand/total_raw_material_demand_by_origin_%s.rds"), this_allocation_method))

  
```

# Differences in embodied fish among regions and the main drivers

```{r}

#Calculations of differences in FMFO demand

raw_material_iteration |> 
  filter(type %in%  c( "Total") & diet == "marine_diet") |> 
  arrange(-demand) |> 
  mutate(diff = demand-demand[1])



#plot differences for supplementary information


raw_material_iteration |> 
  filter(type %in%  c("Fishmeal", "Fish oil", "Total")) |>
  mutate(fao_area_code = factor(fao_area_code),
         diet = case_when(diet == "marine_diet" ~ "Fish-dominant diet", TRUE ~ "Plant-dominant diet")) |> 
  ggplot(aes(x = type, y = demand/1000, fill = fao_area_code))+
  geom_bar(stat = "identity", position = "dodge")+
  facet_grid(cols = vars(diet))+
  theme_bw()+
  scale_fill_manual(values = RColorBrewer::brewer.pal(n=8, name = "BuPu")[3:8])+
  theme(legend.position =  c(0.66,0.76),
        legend.background = element_rect(fill="transparent"),
        panel.grid = element_blank(),
        axis.title.x= element_blank(),
        text = element_text(size=8),
        legend.text = element_text(size=7),
        legend.title = element_text(size=7),
        legend.key.size = unit(0.3, "cm"))+
  labs(y = "Embodied fish demand (1000s tonnes)", fill = "FAO fishing area")

ggsave(filename = sprintf(here("figures/embodied_fish_demand_%s.jpeg"), this_allocation_method), dpi = 600, device = "jpeg", width = 9, height = 6, units = "cm")


#calculations

raw_material_iteration |> 
  filter(type %in%  c("Total") & diet == "marine_diet") |> pull(demand) |> sort() |> pluck(2)-raw_material_iteration |> 
  filter(type %in%  c("Total") & diet == "marine_diet") |> pull(demand) |> sort() |> pluck(1)
  



# Reasons for differences

fread(here("data/large_data/embodied_fishfromtrimmings_per_cell.csv")) |> 
  lazy_dt() |> 
  select(diet, fao_area_code, TaxonName, CommonName, trimmings_weight, byproduct_allocation_factor_fishmeal, byproduct_allocation_factor_fish_oil, total_embodied_fish) |> 
  mutate(source = "trimmings") |> 
  as_tibble() |> 
  group_by(diet, fao_area_code) |> 
  summarise(embodied_fish = sum(total_embodied_fish))

fread(here("data/large_data/embodied_foragefish_per_cell.csv")) |> mutate(source = "forage fish") |> 
  select(diet, fao_area_code, TaxonName, CommonName, forage_allocation_factor_fishmeal, forage_allocation_factor_fish_oil, total_embodied_fish) |> 
  as_tibble() |> 
  group_by(diet, fao_area_code, TaxonName, CommonName, forage_allocation_factor_fishmeal, forage_allocation_factor_fish_oil) |> 
  summarise(embodied_fish = sum(total_embodied_fish)) |> 
  arrange(diet, fao_area_code, -embodied_fish) |> 
  ungroup() |> 
  filter(diet == "marine_diet") |> 
  group_by(diet, fao_area_code) |> 
  nest() |> 
  mutate(cumulative = map(data, ~(cumsum(.$embodied_fish/sum(.$embodied_fish)))))|> 
  unnest(cols = c(data, cumulative)) |> 
  slice(1:3) 
  

```


#Differences betweem diets from the iterations

```{r}

all_data <- readRDS(file = sprintf(here("data/tidy_data/pressures/marine_plant_disaggregated_pressures_by_ingredient_combined_df_%s.rds"),this_allocation_method))


summary_by_total <- 
  all_data |> 
  group_by(source, diet) |> 
  summarise(cpi = sum(sum, na.rm=TRUE))


#average differences in cpi
summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> median()/summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> median()
summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> max()/summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> max()


#variability within each feed scenario
summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> max()/summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> min()
summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> max()/summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> min()


#overlap in CPI

#how many of the fish-dominant scores are lower than the minimum plant-dominant scores
which(summary_by_total |> filter(diet == "marine diet") |> pull(cpi)<summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> min(), TRUE) |> length()/2000

#how many of the fish-dominant scores are greater than the minimum plant-dominant scores
which(summary_by_total |> filter(diet == "marine diet") |> pull(cpi)>summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> min(), TRUE) |> length()/2000



##how many of the plant-dominant scores are lower than the maximum fish-dominant scores
which(summary_by_total |> filter(diet == "plant diet") |> pull(cpi)<summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> max(), TRUE) |> length()/2000

##how many of the plant-dominant scores are higher than the maximum fish-dominant scores
which(summary_by_total |> filter(diet == "plant diet") |> pull(cpi)>summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> max(), TRUE) |> length()/2000


#how many times lower can the CPI of the plant dominant scenario be despite the higher average
summary_by_total |> filter(diet == "plant diet") |> pull(cpi) |> min()/summary_by_total |> filter(diet == "marine diet") |> pull(cpi) |> max()



summary_by_pressure <- 
  all_data |> 
  group_by(diet, source, pressure) |> 
  summarise(cpi_cont = sum(sum, na.rm=TRUE)) |> 
   ungroup() |> 
   group_by(pressure, diet) |> 
   summarise(median = median(cpi_cont),
             mean = mean(cpi_cont),
             min = min(cpi_cont),
             max = max(cpi_cont))


#how much higher is median disturbance pressure in fish-dominant scenario?

summary_by_pressure |> filter(pressure == "Disturbance") |> pull(median) |> pluck(1)/summary_by_pressure |> filter(pressure == "Disturbance") |> pull(median) |> pluck(2)

#how much higher is median GHG, nutrients, and water in plant-dominant feed?
summary_by_pressure |> filter(pressure == "GHG") |> pull(median) |> pluck(2)/summary_by_pressure |> filter(pressure == "GHG") |> pull(median) |> pluck(1)
summary_by_pressure |> filter(pressure == "Nutrients") |> pull(median) |> pluck(2)/summary_by_pressure |> filter(pressure == "Nutrients") |> pull(median) |> pluck(1)
summary_by_pressure |> filter(pressure == "Water") |> pull(median) |> pluck(2)/summary_by_pressure |> filter(pressure == "Water") |> pull(median) |> pluck(1)


#How much do minimum and maximum scores vary across different diets and pressures
summary_by_pressure |> filter(pressure == "GHG" & diet == "marine diet") |> pull(max)/summary_by_pressure |> filter(pressure == "GHG" & diet == "marine diet") |> pull(min)

summary_by_pressure |> filter(pressure == "GHG" & diet == "marine diet") |> pull(max)/summary_by_pressure |> filter(pressure == "GHG" & diet == "plant diet") |> pull(min)

summary_by_pressure |> filter(pressure == "Disturbance" & diet == "marine diet") |> pull(max)/summary_by_pressure |> filter(pressure == "Disturbance" & diet == "plant diet") |> pull(max)
summary_by_pressure |> filter(pressure == "Disturbance" & diet == "plant diet") |> pull(max)/summary_by_pressure |> filter(pressure == "Disturbance" & diet == "marine diet") |> pull(min)



#nutrients differences
summary_by_pressure |> filter(pressure == "Nutrients" & diet == "plant diet") |> pull(max)/summary_by_pressure |> filter(pressure == "Nutrients" & diet == "marine diet") |> pull(max)


summary_by_pressure |> filter(pressure == "Nutrients" & diet == "plant diet") |> pull(min)/summary_by_pressure |> filter(pressure == "Nutrients" & diet == "marine diet") |> pull(max)


#overalp of nutrients
pressure_scores <- 
  all_data |> 
  group_by(diet, source, pressure) |> 
  summarise(cpi_cont = sum(sum, na.rm=TRUE)) |> 
   ungroup()



which(pressure_scores |> filter(diet == "plant diet" & pressure == "Nutrients") |> pull(cpi_cont) > pressure_scores |> filter(diet == "marine diet" & pressure == "Nutrients") |> pull(cpi_cont) |> max(), TRUE) |> length()/2000

```

Breakdown of why FMFO GHGs vary so much - particularly the Northwest Pacific.

```{r}

embodied_fish_ghgs <- readRDS(sprintf(here("data/large_data/embodied_fish_ghgs_%s.rds"), this_allocation_method))


embodied_fish_ghgs |> group_by(fao_area_code, VBDesc) |> 
  summarise(ghg_fm = sum(ei_kg_CO2_eq_fm),
            ghg_fo = sum(ei_kg_CO2_eq_fo)) |> 
  mutate(total_ghg = ghg_fm+ghg_fo) |> 
  arrange(-total_ghg) |> 
  group_by(fao_area_code) |> 
  nest() |> 
  mutate(prop_ghg = map(data, ~(.$total_ghg/sum(.$total_ghg)))) |>
  mutate(total_area_fm_ghgs = map(data, ~(sum(.$ghg_fm)))) |>
  mutate(total_area_fo_ghgs = map(data, ~(sum(.$ghg_fo)))) |>
  mutate(total_area_ghgs = map(data, ~(sum(.$total_ghg)))) |> 
  unnest(cols = c(data, prop_ghg,total_area_fm_ghgs, total_area_fo_ghgs, total_area_ghgs))

embodied_fish_ghgs |> 
  group_by(TaxonName, VBDesc, fao_area_code) |> 
  summarise(ghg_fm = sum(ei_kg_CO2_eq_fm),
            ghg_fo = sum(ei_kg_CO2_eq_fo)) |> 
  mutate(total_ghg = ghg_fm+ghg_fo) |> 
  arrange(fao_area_code, -total_ghg) |> filter(fao_area_code %in% c(61,87))

  

```


Breakdown of disturbance variation
```{r}

#breakdown of trophic level by area
readRDS(sprintf(here("data/large_data/fish_tls_by_area_%s.rds"), this_allocation_method)) |> 
  group_by(fao_area_code) |> 
  summarise(weighted_mean_tl = sum(weighted_tl)/sum(total_embodied_fish))

#which species contribute the most to fish from the different regions and what is their trophic level?
readRDS(sprintf(here("data/large_data/fish_tls_by_area_%s.rds"), this_allocation_method)) |> 
  group_by(fao_area_code, CommonName, DietTroph) |> 
  summarise(total_embodied_fish = sum(total_embodied_fish)) |> 
  ungroup() |> 
  arrange( -total_embodied_fish) #blue whiting is a large contributor with a tl of 4.13


#how much does distrubance vary for fmfo?
marine_fo_dist <- 
  
  list.files(here("data/spatial/marine_diet/int"), 
                        pattern = "fish oil_production_disturbance_moll|fish oil_processing_disturbance_moll", full.names = TRUE) |>
  map(.f = \(this_file){
    
    this_area <- str_sub(this_file, start = nchar(tools::file_path_sans_ext(this_file))-1, end = nchar(tools::file_path_sans_ext(this_file)))
    
    this_stage <- str_extract(pattern = "production|processing", string = this_file)
    
    return(global(rast(this_file), fun = sum, na.rm=TRUE) |> 
      select(sum) |> 
      mutate(area = this_area,
             stage = this_stage,
             diet = "Fish-dominant"))
  })


plant_fo_dist <- 
  
  list.files(here("data/spatial/plant_diet/int"), 
                        pattern = "fish oil_production_disturbance_moll|fish oil_processing_disturbance_moll", full.names = TRUE) |>
  map(.f = \(this_file){
    
    this_area <- str_sub(this_file, start = nchar(tools::file_path_sans_ext(this_file))-1, end = nchar(tools::file_path_sans_ext(this_file)))
   
     this_stage <- str_extract(pattern = "production|processing", string = this_file)
    
    return(global(rast(this_file), fun = sum, na.rm=TRUE) |> 
      select(sum) |> 
      mutate(area = this_area,
             stage = this_stage,
             diet = "Plant-dominant"))
  })
  

fish_disturbance_df <- 
  bind_rows(marine_fo_dist, plant_fo_dist) |> 
  mutate(area_name = case_when(area == 27 ~ "Northeast Atlantic",
                          area == 31 ~ "Western-central Atlantic",
                          area == 61 ~ "Northwest Pacific",
                          area == 67 ~ "Northeast Pacific",
                          area == 87 ~ "Southeast Pacific")) 

#fishing in area 27 compared to min (31)
fish_disturbance_df |> filter(stage == "production" & diet == "Fish-dominant") |>  pull(sum) |> max()/fish_disturbance_df |> filter(stage == "production" & diet == "Fish-dominant") |>  pull(sum) |> min()

#at least how many times greater is area 27 than anywhere else in disturbance
fish_disturbance_df |> filter(stage == "production" & diet == "Fish-dominant") |> arrange(-sum) |>  pull(sum) |> pluck(1) /fish_disturbance_df |> filter(stage == "production" & diet == "Fish-dominant") |> arrange(-sum) |>  pull(sum) |> pluck(2)



#bring in the FAO regions shape file

fao_areas_shp <- 
  read_sf("https://www.fao.org/fishery/geoserver/fifao/ows?service=WFS&request=GetFeature&version=1.0.0&typeName=fifao:FAO_AREAS_CWP_NOCOASTLINE&outputFormat=json") |> filter(F_LEVEL == "MAJOR") |> 
  filter(F_CODE %in% c(27, 31, 61, 67, 87)) |> 
  mutate(F_AREA = as.numeric(F_AREA)) |> 
  group_by(F_AREA) |> 
  group_split()

this_area <- fao_areas_shp[[5]]


map_df(.x=fao_areas_shp, .f = \(this_area){
  
  npp_rast <- rast(here("data/spatial/00-net-primary-productivity/npp_2013_2022_mean.tif"))
  
  fao_area <- this_area$F_CODE
  
  npp_sum <- crop(npp_rast, ext(this_area)) |> values() |> sum(na.rm=TRUE)
  
  return(tibble(fao_area_code = fao_area, npp_sum = npp_sum))

})


#look at the productivity of the regions

prop_npp <- 
  readRDS(file = sprintf(here("data/large_data/prop_npp_%s.rds"), this_allocation_method)) |> 
  group_by(diet, fao_area_code, LonCentre, LatCentre) |> 
  summarise(npp = sum(mean_npp_t_C_yr, na.rm=TRUE),
            ppr_fo = sum(ppr_prop_fo, na.rm = TRUE)) |> 
ungroup() |> 
  group_by(fao_area_code, diet) |> 
  summarise(prop_npp = sum(ppr_fo)/sum(npp)) |> 
  mutate(fao_area_code = factor(fao_area_code)) |> 
  mutate(diet = case_when(diet == "marine_diet" ~ "Fish-dominant",
                          TRUE ~ "Plant-dominant"))
  

my_pal <- rcartocolor::carto_pal(n = 8, name = "Bold")[c(1, 3, 7, 2)]

ggplot(data = prop_npp, aes(x=fao_area_code, y = prop_npp, fill = diet)) +
  geom_col(position = "dodge", alpha=0.8)+
  labs(x = "FAO fishing area", y = "Proportion NPP appropriated", fill = "Feed scenario")+
  theme_bw()+
  theme(text = element_text(size=8),
        legend.text = element_text(size=8),
        legend.title = element_text(size=8),
        legend.position = c(0.75,0.85),
        legend.key.size = unit(0.3, "cm"),
        panel.grid = element_blank())+
  scale_fill_manual(values = rev(my_pal[c(4,2)]))+
  scale_y_continuous(labels = scales::scientific)
  

ggsave(filename = sprintf(here("figures/npp_appropriation_%s.jpg"), this_allocation_method), dpi=600, width = 9.3, height=7, units = "cm")
  
```



Some statistics on the number of species/crop raw materials/countries/marine regions needed to support demand. 

```{r}
forage_fish <- data.table::fread(here("data/tidy_data/demand/embodied_fish_per_cell.csv"))
forage_fish_md <- forage_fish |> filter(diet == "marine_diet") |> as_tibble()
unique(forage_fish$CommonName)
sum(forage_fish_md$total_catch)


#Number of countries for crops
countries_regions <-
  c(list.files("/mnt/rdsi/github/feed_pressure_mapping/data/spatial/marine_diet/pressures", full.names = TRUE), list.files("/mnt/rdsi/github/feed_pressure_mapping/data/spatial/plant_diet/pressures", full.names = TRUE)) |>
  map(.f = \(this_element){
    sans_ext <- tools::file_path_sans_ext(this_element)
    
    last_3 <-  str_sub(string = sans_ext, start = nchar(sans_ext)-2, end = nchar(sans_ext))
    return(last_3)
  }) |> 
  unlist() |> 
  unique() 


length(countries_regions[!countries_regions %in% c("EPC", "NAT", "OTH")]) # 18 countries used and 3 marine regions


#number of crop types
total_crop_demand <- readRDS("/mnt/rdsi/github/feed_pressure_mapping/data/tidy_data/demand/total_crop_demand.rds") 
total_crop_demand |> pull(FAOSTAT_name) |> unique()
total_crop_demand |> filter(FAOSTAT_name == "Pulses nes")

#forage fish embodied demand by regions
embodied_fish_bysource <- readRDS(here("data/tidy_data/demand/embodied_fish_demand_by_source_diet.rds")) |> 
  mutate(source_code = case_when(source_code == "EPC" ~ "Eastern Pacific",
                                 source_code == "NAT" ~ "North Atlantic",
                                 source_code == "OTH" ~ "All other regions")) |> 
  mutate(source_code = factor(source_code, levels = c("Eastern Pacific", "North Atlantic", "All other regions"))) |> 
  mutate(diet = if_else(diet=="marine_diet", true = "Fish-dominant\nfeed", false = "Plant-dominant\nfeed"))

#how much more embodied fish does a marine diet create relative to plant?
source_codes <- unique(embodied_fish_bysource$source_code) |> as.character()
diff_list_ge <- list()
diff_list_mass <- list()
amount_list <- list()

for(s in source_codes){
 
  marine_plant_ratio_ge <-  
    (embodied_fish_bysource |> filter(source_code == s & allocation == "Energetic allocation" & diet == "Fish-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))/(embodied_fish_bysource |> filter(source_code == s & allocation == "Energetic allocation" & diet == "Plant-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))
  
  marine_plant_ratio_mass <-  
    (embodied_fish_bysource |> filter(source_code == s & allocation == "Mass allocation" & diet == "Fish-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))/(embodied_fish_bysource |> filter(source_code == s & allocation == "Mass allocation" & diet == "Plant-dominant\nfeed" & demand_from == "Total") |> pull(embodied_fish))
  
  
 diff_list_ge[which(source_codes==s)] <- marine_plant_ratio_ge
 
 diff_list_mass[which(source_codes==s)] <- marine_plant_ratio_mass
 
}

unlist(diff_list_ge) #~2.2 across all regions and allocation methods=
unlist(diff_list_mass)

#Differences among regions within marine-dominant feed
embodied_fish_bysource |> 
  filter(diet == "Fish-dominant\nfeed" & demand_from == "Total") |> 
  arrange(-embodied_fish) |> 
  mutate(region_diff = max(embodied_fish)-embodied_fish)

embodied_fish_bysource |>
  filter(demand_from =="Total") |> 
  group_by(diet, allocation) |> 
  summarise(mean = mean(embodied_fish, na.rm=TRUE),
            sd = sd(embodied_fish, na.rm=TRUE))


ggplot(data = embodied_fish_bysource) +
  aes(x = demand_from, y = embodied_fish/1000, fill = source_code)+
  geom_bar(stat= "identity", position = "dodge")+
  facet_grid(rows = vars(diet), cols = vars(allocation), scales = "fixed")+
  
  labs(fill = "Origin", y = "Embodied fish demand (1000s T)")+
  scale_fill_manual(values = c("#edf8b1", "#7fcdbb", "#2c7fb8"))+
  theme_bw()+
  theme(legend.position = c(0.65, 0.35),
        legend.background = element_rect(fill="transparent"), 
        axis.title.x = element_blank(),
        panel.grid = element_blank(), 
        text = element_text(size=12),
        legend.text = element_text(size = 10))+
  guides(fill = guide_legend(title.position = "top"))

ggsave(filename = here("figures/presos/embodied_forage_fish_demand.jpg"), device="jpg", dpi = 600, width = 18, height = 15, units="cm")
ggsave(filename = here("figures/supplementary/embodied_forage_fish_demand.jpg"), device="jpg", dpi = 600, width = 18, height = 15, units="cm")



#find reasons for differences in demand
#blue whiting looks like the main reason with European sprat also. Comparing the embodied fish versus the catch both Sprat and Whiting rank as higher in embodied fish than for catch
(embodied_fish_rank <- 
  fread("/mnt/rdsi/github/feed_pressure_mapping/data/tidy_data/demand/embodied_fish_per_cell.csv") |> 
  lazy_dt() |> 
  group_by(diet, source_code, TaxonName, CommonName) |> 
  summarise(total_demand_ge = sum(total_embodied_fish_ge, na.rm = TRUE),
            total_demand_mass = sum(total_embodied_fish_mass, na.rm = TRUE)) |> 
  filter(diet == "marine_diet") |> 
  arrange(-total_demand_mass) |> 
  as_tibble())


(fish_catch_rank <- 
  fread("/mnt/rdsi/github/feed_pressure_mapping/data/tidy_data/demand/embodied_fish_per_cell.csv") |> 
  lazy_dt() |> 
  group_by(diet, source_code, TaxonName, CommonName) |> 
  summarise(total_catch = sum(total_catch, na.rm = TRUE)) |> 
  filter(diet == "marine_diet") |> 
  arrange(-total_catch) |> 
  as_tibble())


left_join(embodied_fish_rank, fish_catch_rank) |> mutate(disproportion_mass =  total_demand_mass/total_catch) |> arrange(-disproportion_mass)
```
